% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolutebiomassplot_prep_DD.R
\name{absolutebiomassplot_prep_DD}
\alias{absolutebiomassplot_prep_DD}
\title{Prepare DDUST data for absolute biomass plot}
\usage{
absolutebiomassplot_prep_DD(
  dd_mle,
  dd_mcmc = NULL,
  dd_sim = NULL,
  scenarios = NULL
)
}
\arguments{
\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{dd_mcmc}{Placeholder. No functionality for MCMC build yet. A list of model fits from tmbstan::tmbstan() with one list element per scenario. Only needed if MCMC was used.}

\item{dd_sim}{Placeholder. No functionality for MCMC build yet. A list of outputs from SSAND::simulate_DD() with one list element per scenario. Only required if MCMC was used. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}
}
\value{
A data frame with date, fleet, obs, exp, ub, lb, scenario
}
\description{
Prepare DDUST data for absolute biomass plot
}
\examples{
\dontrun{
data <- absolutebiomassplot_prep_DD(dd_mle)
absolutebiomassplot(data)
}
}
