% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{add_scenarios}
\alias{add_scenarios}
\title{Generate a detailed model output appendix
This code produced a text file that contains code that can be pasted directly into the "Detailed model outputs" section of a stock assessment report
Code can be customised for the number of scenarios, model type and output type.}
\usage{
add_scenarios(
  filename = paste0(getwd(), "/scenarios_appendix.txt"),
  model = "SS",
  MCMC = TRUE,
  scenarios = 1:2,
  length = TRUE,
  age = FALSE,
  caal = TRUE,
  discard = FALSE
)
}
\arguments{
\item{filename}{Directory and filename where text file will be saved. Defaults to working directory}

\item{model}{"SS" or "DD" for Stock Synthesis or DDUST respectively. Will populate the relevant content}

\item{MCMC}{TRUE or FALSE if MCMC was used. Will populate the relevant content}

\item{scenarios}{A vector of scenario numbers used. Will populate the appendix.}

\item{length}{TRUE or FALSE if length compositions were used. Will populate the relevant content}

\item{age}{TRUE or FALSE if age compositions were used. Will populate the relevant content}

\item{caal}{TRUE or FALSE if conditional age-at-length compositions were used. Will populate the relevant content}

\item{discard}{TRUE or FALSE if discard inputs were used. Will populate the relevant content}
}
\value{
A text file that contains code that can be pasted directly into the "Detailed model outputs" section of a stock assessment report
}
\description{
Generate a detailed model output appendix
This code produced a text file that contains code that can be pasted directly into the "Detailed model outputs" section of a stock assessment report
Code can be customised for the number of scenarios, model type and output type.
}
\examples{
\dontrun{
add_scenarios()
add_scenarios(model="SS", MCMC=FALSE, scenarios=1:12, discard=TRUE)
}
}
