% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andreplot_prep_SS.R
\name{andreplot_prep_SS}
\alias{andreplot_prep_SS}
\title{Prepare SS data for Andre plot
Andre plots show mean age and standard deviation in conditional age-at-length data.
Left plots are mean age at length by size-class (obs. and exp.) with confidence intervals obtained by adding the appropriate number of standard errors of mean to the data.
Right plots in each pair are SE of mean age-at-length (obs. and exp.) with confidence intervals based on the chi-square distribution.}
\usage{
andreplot_prep_SS(ss_mle, scenarios = NULL, sex_code = NULL, CI = 0.95)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{sex_code}{Specify which sex to show on the graph. Sex codes correspond to Stock Synthesis sex codes (1 is female, 2 is male, 3 and 0 are combined.)}

\item{CI}{Confident interval. Default is 0.95.}
}
\value{
A data frame with columns "year" (num), "fleet" (num), "length" (num), "obs" (num), "pred" (num), "low" (num), "upp" (num), "label" (chr), "scenario" (int), "CI" (num)
}
\description{
Prepare SS data for Andre plot
Andre plots show mean age and standard deviation in conditional age-at-length data.
Left plots are mean age at length by size-class (obs. and exp.) with confidence intervals obtained by adding the appropriate number of standard errors of mean to the data.
Right plots in each pair are SE of mean age-at-length (obs. and exp.) with confidence intervals based on the chi-square distribution.
}
\examples{
data <- andreplot_prep_SS(ss_mle, sex_code=1)
andreplot(data, fleet = 1)
}
