% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_data.R
\docType{data}
\name{biological_data}
\alias{biological_data}
\title{Fake data to be used in investigate_biology functions.}
\format{
\subsection{\code{biological_data}}{

A sample set of length and age data
}
}
\usage{
biological_data
}
\description{
A sample set of length and age data
}
\keyword{datasets}
