% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassplot_prep_SS.R
\name{biomassplot_prep_SS}
\alias{biomassplot_prep_SS}
\title{Prepare Stock Synthesis data for SSAND::biomassplot()
If using MCMC, the "interval", "prob_lower" and "prob_upper" column will return NA when med="MCMC".}
\usage{
biomassplot_prep_SS(
  ss_mle,
  ss_mcmc = NULL,
  scenarios = NULL,
  period = NULL,
  biomass_type = "relative",
  biomass_definition = "spawning",
  selectivity_fleet = 2,
  intervals = c(0.2, 0.5, 0.7, 0.9, 0.95)
)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{A list of outputs from r4ss::SSgetMCMC() with one element per scenario. Only needed if MCMC was used. Will automatically reformat as a list if a single r4ss::SSgetMCMC() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{period}{Optional: time period of stock assessment.}

\item{biomass_type}{The type of biomass used. Options are "relative" or "absolute" (character).}

\item{biomass_definition}{The definition of biomass used. Options are "spawning" or "vulnerable" (character).}

\item{selectivity_fleet}{Fleet number used for the selectivity applied to calculate vulnerable biomass.}

\item{intervals}{A vector of credible interval values to be displayed on banded MCMC plot (numeric). For example, "0.9" denotes the 90\% credible interval.}
}
\value{
If MCMC is being used, returns a data frame with variables called rownum, scenario, year, value, med, interval, prob_lower, prob_upper, biomass_type, biomass_definition. If MLE, returns a data frame with variables called year, value, lower, upper, scenario, biomass_type, biomass_definition.
}
\description{
Prepare Stock Synthesis data for SSAND::biomassplot()
If using MCMC, the "interval", "prob_lower" and "prob_upper" column will return NA when med="MCMC".
}
\examples{
data <- biomassplot_prep_DD(dd_mle)
biomassplot(data)

# Vulnerable biomass
data <- biomassplot_prep_SS(ss_mle,
                            biomass_type = 'absolute',
                            biomass_definition = 'vulnerable',
                            intervals = c(0.2, 0.5, 0.7, 0.9, 0.95),
                            scenarios = 1)

# MCMC
data <- biomassplot_prep_SS(ss_mle, ss_mcmc)
biomassplot(data)

# Ensemble model
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc)
data <- biomassplot_prep_SS(ss_mle, ss_mcmc_ens)
biomassplot(data)
}
