% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomasstable_prep_DD.R
\name{biomasstable_prep_DD}
\alias{biomasstable_prep_DD}
\title{Prepare DDUST data for biomass table}
\usage{
biomasstable_prep_DD(
  dd_mle,
  dd_mcmc,
  dd_sim,
  end_year = NULL,
  credible_interval = 0.95,
  scenarios = NULL,
  period = NULL,
  biomass_type = "relative"
)
}
\arguments{
\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{dd_mcmc}{A list of model fits from tmbstan::tmbstan() with one list element per scenario. Only needed if MCMC was used.}

\item{dd_sim}{A list of outputs from DDUST::simulate_DD() with one list element per scenario. Only required if MCMC was used. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{end_year}{The final year of biomass to be plotted. Calculated by default, but option is included in case model is run into the future.}

\item{credible_interval}{The credible interval to use to create dashed lines in the final plot}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{period}{Optional: time period of stock assessment.}

\item{biomass_type}{The type of biomass used. Options are "relative" or "absolute" (character).}
}
\value{
A data frame that summarised biomass estimate, ready for use in biomasstable()
}
\description{
Prepare DDUST data for biomass table
}
\examples{
\dontrun{
library(DDUST)
dd_sim <- simulate_DDUST(dd_mle,dd_mcmc)

# <<table_biomass, results='asis', echo=FALSE>>=
data <- biomasstable_prep_DD(dd_mle,dd_mcmc,dd_sim)
biomasstable(data, label="tab:biomass")
# @
}
}
