% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomasstable_prep_SS.R
\name{biomasstable_prep_SS}
\alias{biomasstable_prep_SS}
\title{Prepare Stock Synthesis data for biomass table}
\usage{
biomasstable_prep_SS(
  ss_mle,
  ss_mcmc,
  end_year = NULL,
  credible_interval = 0.95,
  scenarios = NULL,
  period = NULL,
  biomass_type = "relative",
  biomass_definition = "spawning",
  selectivity_fleet = 2
)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{A list of outputs from r4ss::SSgetMCMC() with one element per scenario. Only needed if MCMC was used. Will automatically reformat as a list if a single r4ss::SSgetMCMC() output (i.e. one scenario) is entered.}

\item{end_year}{The final year of biomass to be plotted. Calculated by default, but option is included in case model is run into the future.}

\item{credible_interval}{The credible interval to use to create dashed lines in the final plot}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{period}{Optional: time period of stock assessment.}

\item{biomass_type}{The type of biomass used. Options are "relative" or "absolute" (character).}

\item{biomass_definition}{The definition of biomass used. Options are "spawning" or "vulnerable" (character).}

\item{selectivity_fleet}{Fleet number used for the selectivity applied to calculate vulnerable biomass.}
}
\value{
A data frame that summarises biomass estimate, ready for use in biomasstable()
}
\description{
Prepare Stock Synthesis data for biomass table
}
\examples{
# <<table_biomass, results='asis', echo=FALSE>>=
data <- biomasstable_prep_SS(ss_mle,ss_mcmc)
biomasstable(data, label="tab:biomass")
# @
}
