% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caal_agefitplot_prep_SS.R
\name{caal_agefitplot_prep_SS}
\alias{caal_agefitplot_prep_SS}
\title{Prepare Stock Synthesis data for plot to show age fits from conditional-age-at-length data}
\usage{
caal_agefitplot_prep_SS(ss_mle, sex_code = 1, scenarios = NULL)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{sex_code}{Specify which sex to show on the graph. Sex codes correspond to Stock Synthesis sex codes (1 is female, 2 is male, 3 and 0 are combined.)}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}
}
\value{
A data frame with year (int), bin (int), obs (num), exp (num), scenario (int), sex (int), fleet (int).
}
\description{
Prepare Stock Synthesis data for plot to show age fits from conditional-age-at-length data
}
\examples{
data <- caal_agefitplot_prep_SS(ss_mle)
caal_agefitplot(data, scenario=2,show_fits=FALSE)
caal_agefitplot(data, scenario=2)
}
