% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{catalogue}
\alias{catalogue}
\title{Generate a catalogue of all available plots}
\usage{
catalogue(dir = getwd())
}
\arguments{
\item{dir}{Directory where catalogue will be saved. Defaults to working directory.}
}
\value{
A html file containing a catalogue of all available plots
}
\description{
Generate a catalogue of all available plots
}
\examples{
\dontrun{
catalogue()
}
}
