% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchabilityplot_prep_DD.R
\name{catchabilityplot_prep_DD}
\alias{catchabilityplot_prep_DD}
\title{Prepare DDUST data for seasonal catchability pattern}
\usage{
catchabilityplot_prep_DD(
  dd_mle,
  dd_mcmc = NULL,
  dd_sim = NULL,
  scenarios = NULL,
  fleets = NULL
)
}
\arguments{
\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{dd_mcmc}{A list of model fits from tmbstan::tmbstan() with one list element per scenario. Only needed if MCMC was used.}

\item{dd_sim}{A list of outputs from SSAND::simulate_DD() with one list element per scenario. Only required if MCMC was used. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{fleets}{A vector of which fleet/s to plot (numeric).}
}
\value{
A data frame with q (num), month (int), month_point (int), monthnames (factor), fleet (factor), scenario (factor)
}
\description{
Prepare DDUST data for seasonal catchability pattern
}
\examples{
data <- catchabilityplot_prep_DD(dd_mle)
catchabilityplot(data)

\dontrun{
# MCMC model
library(DDUST)
dd_sim <- simulate_DDUST(dd_mle,dd_mcmc)

data <- catchabilityplot_prep_DD(dd_mle,dd_mcmc,dd_sim)
catchabilityplot(data)

# Ensemble model:
library(DDUST)
dd_sim <- simulate_DDUST(dd_mle,dd_mcmc)
dd_mcmc_ens <- mcmc_ensemble_DD(dd_mcmc,dd_sim)$dd_mcmc
dd_sim_ens <- mcmc_ensemble_DD(dd_mcmc,dd_sim)$dd_sim
data <- catchabilityplot_prep_DD(dd_mle,dd_mcmc_ens,dd_sim_ens)
catchabilityplot(data)
}
}
