% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchabilityplot_prep_SS.R
\name{catchabilityplot_prep_SS}
\alias{catchabilityplot_prep_SS}
\title{Prepare Stock Synthesis data for catchability pattern plot}
\usage{
catchabilityplot_prep_SS(ss_mle, scenarios = NULL)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}
}
\value{
A data frame with q (num), month (int), month_point (int), monthnames (factor), fleet (factor), scenario (factor)
}
\description{
Prepare Stock Synthesis data for catchability pattern plot
}
\examples{
data <- catchabilityplot_prep_SS(ss_mle)
catchabilityplot(data)

\dontrun{
# MCMC
data <- catchabilityplot_prep_SS(ss_mle, ss_mcmc)
catchabilityplot(data)

# Ensemble model
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc)
data <- catchabilityplot_prep_SS(ss_mle, ss_mcmc_ens)
catchabilityplot(data)
}
}
