% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchplot_prep_DD.R
\name{catchplot_prep_DD}
\alias{catchplot_prep_DD}
\title{Prepare DDUST data for catchplot()}
\usage{
catchplot_prep_DD(dd_mle, scenarios = NULL)
}
\arguments{
\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}
}
\value{
A dataframe with date, value, sector, scenario
}
\description{
Prepare DDUST data for catchplot()
}
\examples{
data <- catchplot_prep_DD(dd_mle)
catchplot(data)
}
