% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchplot_prep_SS.R
\name{catchplot_prep_SS}
\alias{catchplot_prep_SS}
\title{Prepare Stock Synthesis data for catchplot()}
\usage{
catchplot_prep_SS(ss_mle, scenarios = NULL)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}
}
\value{
A dataframe with date, value, sector, scenario
}
\description{
Prepare Stock Synthesis data for catchplot()
}
\examples{
data <- catchplot_prep_SS(ss_mle)
catchplot(data, scenarios=1)
}
