% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_scenarios.R
\name{check_scenarios}
\alias{check_scenarios}
\title{A function used within SSAND to check if SS or DD input into prep functions are a single scenario or list of scenarios}
\usage{
check_scenarios(x, model, type)
}
\arguments{
\item{x}{dd_mle or ss_mle}

\item{model}{"DD" or "SS"}

\item{type}{"MLE" or "MCMC"}
}
\value{
response determining if input is a list of scenarios or a single scenario
}
\description{
A function used within SSAND to check if SS or DD input into prep functions are a single scenario or list of scenarios
}
\examples{
check_scenarios(ss_mle,"SS","MLE")
}
