% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comms_biomassposterior.R
\name{comms_biomassposterior}
\alias{comms_biomassposterior}
\title{Animations and graphics for communication packages}
\usage{
comms_biomassposterior(
  data1,
  dir = paste0(getwd(), "/"),
  end_year,
  xlab1 = NULL,
  ylab1 = NULL,
  font = "Meta"
)
}
\arguments{
\item{data1}{Output of mcmc_finalbiomassposterior_prep_SS(ss_mle,ss_mcmc_ens,cut=FALSE)}

\item{dir}{Folder to save outputs. Default is getwd()}

\item{end_year}{Year of biomass to report. E.g. If SS model inputs end in
2024, enter 2025.}

\item{xlab1}{x axis label for comms_biomasposterior and comms_histogram_animation.
If blank, defaults to paste0("Biomass in ",end_year)}

\item{ylab1}{y axis label for comms_biomasposterior and comms_histogram_animation.
Should be NULL.}

\item{font}{Font to use on all assets}
}
\description{
Plots have been formatted for consistent visual assets
\itemize{
\item \code{comms_video_suite()}: An umbrella function that creates all assets
\item \code{comms_biomassposterior()}:  Generate MCMC final biomass posterior density
plots and save x-axis information for histogram animations
\item \code{comms_histogram_animation()}: Generate histogram animation
\item \code{comms_hairymcmc()}: Generate hairy biomass plot and save x-axis
information for spaghetti plot
\item \code{comms_spahettiplot()}: Generate spaghetti plot
}
}
\examples{
\dontrun{
library(SSAND)
library(extrafont)
# font_import()
# loadfonts(device = "win")

end_year <- 2025
fig_dir <- paste0(getwd(),'/')
load("data.Rdata")
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc)

data1 <- mcmc_finalbiomassposterior_prep_SS(ss_mle,ss_mcmc_ens,cut=FALSE)
comms_biomassposterior(data1,
                       dir = fig_dir,
                       end_year = end_year)
}
}
