% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comms_hairymcmc.R
\name{comms_hairymcmc}
\alias{comms_hairymcmc}
\title{Animations and graphics for communication packages}
\usage{
comms_hairymcmc(
  data2,
  dir = paste0(getwd(), "/"),
  font = "Meta",
  xlab2 = "Year",
  ylab2 = "Spawning biomass (relative)"
)
}
\arguments{
\item{data2}{Output of biomassplot_prep_SS(ss_mle,ss_mcmc_ens, scenarios=1)}

\item{dir}{Folder to save outputs. Default is getwd()}

\item{font}{Font to use on all assets}

\item{xlab2}{x axis label for comms_hairymcmc and comms_spaghettiplot. Default
is "Year"}

\item{ylab2}{y axis label for comms_hairymcmc and comms_spaghettiplot. Default
is "Spawning biomass (relative)"}
}
\description{
Plots have been formatted for consistent visual assets
\itemize{
\item \code{comms_video_suite()}: An umbrella function that creates all assets
\item \code{comms_biomassposterior()}:  Generate MCMC final biomass posterior density
plots and save x-axis information for histogram animations
\item \code{comms_histogram_animation()}: Generate histogram animation
\item \code{comms_hairymcmc()}: Generate hairy biomass plot and save x-axis
information for spaghetti plot
\item \code{comms_spahettiplot()}: Generate spaghetti plot
}
}
\examples{
\dontrun{
library(SSAND)
library(extrafont)
# font_import()
# loadfonts(device = "win")

end_year <- 2025
fig_dir <- paste0(getwd(),'/')
load("data.Rdata")
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc)

data2 <- biomassplot_prep_SS(ss_mle,ss_mcmc_ens, scenarios=1)

axis_info <- comms_hairymcmc(data2,
                             dir = fig_dir)
}
}
