% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comms_histogram_animation.R
\name{comms_histogram_animation}
\alias{comms_histogram_animation}
\title{Animations and graphics for communication packages}
\usage{
comms_histogram_animation(
  data2,
  end_year,
  bins = c(50000, 10000, 5000, 1000, 500, 100),
  dir = paste0(getwd(), "/"),
  xlab1 = NULL,
  ylab1 = NULL,
  xlim1 = c(0, 1),
  xbreaks1 = c(0, 0.2, 0.4, 0.6, 0.8, 1),
  font = "Meta"
)
}
\arguments{
\item{data2}{Output of biomassplot_prep_SS(ss_mle,ss_mcmc_ens, scenarios=1)}

\item{end_year}{Year of biomass to report. E.g. If SS model inputs end in
2024, enter 2025.}

\item{bins}{Vector of bins for histogram animation}

\item{dir}{Folder to save outputs. Default is getwd()}

\item{xlab1}{x axis label for comms_biomasposterior and comms_histogram_animation.
If blank, defaults to paste0("Biomass in ",end_year)}

\item{ylab1}{y axis label for comms_biomasposterior and comms_histogram_animation.
Should be NULL.}

\item{xlim1}{The first object in the list output by comms_biomassposterior.
A vector of the xlims from comms_biomassposterior, used in
comms_histogram_animation}

\item{xbreaks1}{The second object in the list output by comms_biomassposterior.
A vector of the xbreaks from comms_biomassposterior, used in
comms_histogram_animation}

\item{font}{Font to use on all assets}
}
\description{
Plots have been formatted for consistent visual assets
\itemize{
\item \code{comms_video_suite()}: An umbrella function that creates all assets
\item \code{comms_biomassposterior()}:  Generate MCMC final biomass posterior density
plots and save x-axis information for histogram animations
\item \code{comms_histogram_animation()}: Generate histogram animation
\item \code{comms_hairymcmc()}: Generate hairy biomass plot and save x-axis
information for spaghetti plot
\item \code{comms_spahettiplot()}: Generate spaghetti plot
}
}
\examples{
\dontrun{
library(SSAND)
library(extrafont)
# font_import()
# loadfonts(device = "win")

end_year <- 2025
fig_dir <- paste0(getwd(),'/')
load("data.Rdata")
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc)

data1 <- mcmc_finalbiomassposterior_prep_SS(ss_mle,ss_mcmc_ens,cut=FALSE)
data2 <- biomassplot_prep_SS(ss_mle,ss_mcmc_ens, scenarios=1)

x_info <- comms_biomassposterior(data1,
                                 dir = fig_dir,
                                 end_year = end_year)

comms_histogram_animation(data2,
                          end_year = end_year,
                          dir = fig_dir,
                          xlim1 = x_info[[1]],
                          xbreaks1 = x_info[[2]])
}
}
