% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comms_spaghettiplot.R, R/comms_video_suite.R,
%   R/comms_youtube_suite.R
\name{comms_video_suite}
\alias{comms_video_suite}
\alias{comms_spaghettiplot}
\title{Animations and graphics for communication packages}
\usage{
comms_spaghettiplot(
  data3,
  axis_info,
  dir = paste0(getwd(), "/"),
  font = "Meta",
  xlab2 = "Year",
  ylab2 = "Spawning biomass (relative)"
)

comms_video_suite(
  data1,
  data2,
  data3,
  xlab1 = NULL,
  ylab1 = NULL,
  xlim1 = c(0, 1),
  xbreaks1 = c(0, 0.2, 0.4, 0.6, 0.8, 1),
  xlab2 = "Year",
  ylab2 = "Spawning biomass (relative)",
  dir,
  end_year,
  bins = c(50000, 10000, 5000, 1000, 500, 100),
  font = "Meta",
  axis_info = NULL
)

comms_video_suite(
  data1,
  data2,
  data3,
  xlab1 = NULL,
  ylab1 = NULL,
  xlim1 = c(0, 1),
  xbreaks1 = c(0, 0.2, 0.4, 0.6, 0.8, 1),
  xlab2 = "Year",
  ylab2 = "Spawning biomass (relative)",
  dir,
  end_year,
  bins = c(50000, 10000, 5000, 1000, 500, 100),
  font = "Meta",
  axis_info = NULL
)
}
\arguments{
\item{data3}{Output of spaghettiplot_prep_SS(ss_mle)}

\item{axis_info}{A list generated from comms_hairymcmc containing vectors for
x limits, x breaks, y limits and y breaks, used in comms_spaghettiplot}

\item{dir}{Folder to save outputs. Default is getwd()}

\item{font}{Font to use on all assets}

\item{xlab2}{x axis label for comms_hairymcmc and comms_spaghettiplot. Default
is "Year"}

\item{ylab2}{y axis label for comms_hairymcmc and comms_spaghettiplot. Default
is "Spawning biomass (relative)"}

\item{data1}{Output of mcmc_finalbiomassposterior_prep_SS(ss_mle,ss_mcmc_ens,cut=FALSE)}

\item{data2}{Output of biomassplot_prep_SS(ss_mle,ss_mcmc_ens, scenarios=1)}

\item{xlab1}{x axis label for comms_biomasposterior and comms_histogram_animation.
If blank, defaults to paste0("Biomass in ",end_year)}

\item{ylab1}{y axis label for comms_biomasposterior and comms_histogram_animation.
Should be NULL.}

\item{xlim1}{The first object in the list output by comms_biomassposterior.
A vector of the xlims from comms_biomassposterior, used in
comms_histogram_animation}

\item{xbreaks1}{The second object in the list output by comms_biomassposterior.
A vector of the xbreaks from comms_biomassposterior, used in
comms_histogram_animation}

\item{end_year}{Year of biomass to report. E.g. If SS model inputs end in
2024, enter 2025.}

\item{bins}{Vector of bins for histogram animation}
}
\description{
Plots have been formatted for consistent visual assets
\itemize{
\item \code{comms_video_suite()}: An umbrella function that creates all assets
\item \code{comms_biomassposterior()}:  Generate MCMC final biomass posterior density
plots and save x-axis information for histogram animations
\item \code{comms_histogram_animation()}: Generate histogram animation
\item \code{comms_hairymcmc()}: Generate hairy biomass plot and save x-axis
information for spaghetti plot
\item \code{comms_spahettiplot()}: Generate spaghetti plot
}

Plots have been formatted for consistent visual assets
\itemize{
\item \code{comms_video_suite()}: An umbrella function that creates all assets
\item \code{comms_biomassposterior()}:  Generate MCMC final biomass posterior density
plots and save x-axis information for histogram animations
\item \code{comms_histogram_animation()}: Generate histogram animation
\item \code{comms_hairymcmc()}: Generate hairy biomass plot and save x-axis
information for spaghetti plot
\item \code{comms_spahettiplot()}: Generate spaghetti plot
}

Plots have been formatted for consistent visual assets
\itemize{
\item \code{comms_video_suite()}: An umbrella function that creates all assets
\item \code{comms_biomassposterior()}:  Generate MCMC final biomass posterior density
plots and save x-axis information for histogram animations
\item \code{comms_histogram_animation()}: Generate histogram animation
\item \code{comms_hairymcmc()}: Generate hairy biomass plot and save x-axis
information for spaghetti plot
\item \code{comms_spahettiplot()}: Generate spaghetti plot
}
}
\examples{
\dontrun{
library(SSAND)
library(extrafont)
# font_import()
# loadfonts(device = "win")

end_year <- 2025
fig_dir <- paste0(getwd(),'/')
load("data.Rdata")
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc)

data2 <- biomassplot_prep_SS(ss_mle,ss_mcmc_ens, scenarios=1)
data3 <- spaghettiplot_prep_SS(ss_mle)

axis_info <- comms_hairymcmc(data2,
                             dir = fig_dir)

comms_spaghettiplot(data3,
                    axis_info = axis_info,
                    dir = fig_dir)
}
\dontrun{
library(SSAND)
library(extrafont)
# font_import()
# loadfonts(device = "win")

end_year <- 2025
fig_dir <- paste0(getwd(),'/')
load("data.Rdata")
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc)

data1 <- mcmc_finalbiomassposterior_prep_SS(ss_mle,ss_mcmc_ens,cut=FALSE)
data2 <- biomassplot_prep_SS(ss_mle,ss_mcmc_ens, scenarios=1)
data3 <- spaghettiplot_prep_SS(ss_mle)

comms_video_suite <- function(data1,
                                data2,
                                data3,
                                dir = fig_dir,
                                end_year = 2025)
}
\dontrun{
library(SSAND)
library(extrafont)
# font_import()
# loadfonts(device = "win")

end_year <- 2025
fig_dir <- paste0(getwd(),'/')
load("data.Rdata")
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc)

data1 <- mcmc_finalbiomassposterior_prep_SS(ss_mle,ss_mcmc_ens,cut=FALSE)
data2 <- biomassplot_prep_SS(ss_mle,ss_mcmc_ens, scenarios=1)
data3 <- spaghettiplot_prep_SS(ss_mle)

comms_video_suite <- function(data1,
                                data2,
                                data3,
                                dir = fig_dir,
                                end_year = 2025)
}
}
