% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalageatlengthplot_prep_SS.R
\name{conditionalageatlengthplot_prep_SS}
\alias{conditionalageatlengthplot_prep_SS}
\title{Prepare Stock Synthesis data for conditional age-at-length plot}
\usage{
conditionalageatlengthplot_prep_SS(
  ss_mle,
  scenarios = NULL,
  display_threshold = 0.001,
  sex_code = NULL
)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{display_threshold}{The minimum value to be displayed on the graph (numeric). Observed or Pearson values below this will not be represented a bubble.}

\item{sex_code}{Specify which sex to show on the graph. Sex codes correspond to Stock Synthesis sex codes (1 is female, 2 is male, 3 and 0 are combined.)}
}
\value{
A data frame with year, bin, fleet, lbin_low, pearson
}
\description{
Prepare Stock Synthesis data for conditional age-at-length plot
}
\examples{
data <- conditionalageatlengthplot_prep_SS(ss_mle,sex_code=1)
conditionalageatlengthplot(data, show_fits=FALSE)
}
