% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationplot_prep_DD.R
\name{correlationplot_prep_DD}
\alias{correlationplot_prep_DD}
\title{Prepare DDUST data frame for correlationplot}
\usage{
correlationplot_prep_DD(
  dd_sdr,
  dd_mle = NULL,
  dd_mcmc = NULL,
  dd_sim = NULL,
  other_quantities = NULL,
  scenario = 1
)
}
\arguments{
\item{dd_sdr}{The SD report from a model run. Extracted from model output using TMB::sdreport() as per the example.}

\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{dd_mcmc}{A list of model fits from tmbstan::tmbstan() with one list element per scenario. Only needed if MCMC was used.}

\item{dd_sim}{A list of outputs from SSAND::simulate_DD() with one list element per scenario. Only required if MCMC was used. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{other_quantities}{A vector of other derived quantities to display on plot (character).}

\item{scenario}{The scenarios to plot (numeric). Shows scenario 1 if left blank.}
}
\value{
A data frame for correlation plot with X1 (fac), X2 (fac) and value (num)
}
\description{
Prepare DDUST data frame for correlationplot
}
\examples{
\dontrun{
library(DDUST)
dd_sdr <- TMB::sdreport(dd_mle[[1]]$model)
data <- correlationplot_prep_DD(dd_sdr)
correlationplot(data)

data <- correlationplot_prep_DD(dd_sdr=dd_sdr, dd_mle=dd_mle, dd_mcmc=dd_mcmc, dd_sim=dd_sim)
correlationplot(data, sample=10)
}
}
