% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationplot_prep_SS.R
\name{correlationplot_prep_SS}
\alias{correlationplot_prep_SS}
\title{Prepare Stock Synthesis data frame for correlationplot}
\usage{
correlationplot_prep_SS(ss_mle, ss_mcmc = NULL, scenario = 1, parameters)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{A list of outputs from r4ss::SSgetMCMC() with one element per scenario. Only needed if MCMC was used. Will automatically reformat as a list if a single r4ss::SSgetMCMC() output (i.e. one scenario) is entered.}

\item{scenario}{The scenario to plot (numeric). Shows scenario 1 if left blank. Can be overridden in the plotting function.}

\item{parameters}{A vector of parameters to include on plot (character). Use SSAND::extract_SS_parameters() to find a list of parameter names used.}
}
\value{
A data frame for correlation plot with X1 (fac), X2 (fac) and value (num)
}
\description{
Prepare Stock Synthesis data frame for correlationplot
}
\examples{
parameters <- extract_SS_parameters(ss_mle)[c(2,3,4,26),]
data <- correlationplot_prep_SS(ss_mle, scenario = 1, parameters = parameters)
correlationplot(data)

parameters <- extract_SS_parameters(ss_mle)[c(2,3,4,26),]
data <- correlationplot_prep_SS(ss_mle, ss_mcmc, scenario = 1, parameters = parameters)
correlationplot(data)
}
