% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpueplot_prep_DD.R
\name{cpueplot_prep_DD}
\alias{cpueplot_prep_DD}
\title{Prepare DDUST data for catch per unit effort (cpue) fit plot}
\usage{
cpueplot_prep_DD(
  dd_mle,
  dd_mcmc = NA,
  dd_sim = NA,
  scenarios = NULL,
  intervals = c(0.2, 0.5, 0.7, 0.9, 0.95),
  sigma_I = FALSE
)
}
\arguments{
\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{dd_mcmc}{A list of model fits from tmbstan::tmbstan() with one list element per scenario. Only needed if MCMC was used.}

\item{dd_sim}{A list of outputs from SSAND::simulate_DD() with one list element per scenario. Only required if MCMC was used. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{intervals}{A vector of credible interval values to be displayed on banded MCMC plot (numeric). For example, "0.9" denotes the 90\% credible interval.}

\item{sigma_I}{Set to TRUE to define error bars using sigma I (estimated error) rather than observed CPUE SD.}
}
\value{
A data frame with variables date, fleet, obs, exp, ub, lb, scenario
}
\description{
Prepare DDUST data for catch per unit effort (cpue) fit plot
}
\examples{
data <- cpueplot_prep_DD(dd_mle)
cpueplot(data)
}
