% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpueplot_prep_SS.R
\name{cpueplot_prep_SS}
\alias{cpueplot_prep_SS}
\title{Prepare Stock Synthesis data for catch per unit effort (cpue) fit plot}
\usage{
cpueplot_prep_SS(
  ss_mle,
  ss_mcmc = NULL,
  scenarios = NULL,
  month_override = TRUE
)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{A list of outputs from r4ss::SSgetMCMC() with one element per scenario. Only needed if MCMC was used. Will automatically reformat as a list if a single r4ss::SSgetMCMC() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{month_override}{Set to TRUE to change month to 1 for each year. Useful for monthly models.}
}
\value{
A data frame with variables date, fleet, obs, exp, ub, lb, scenario
}
\description{
Prepare Stock Synthesis data for catch per unit effort (cpue) fit plot
}
\examples{
data <- cpueplot_prep_SS(ss_mle)
cpueplot(data)
}
