% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpueqplot.R
\name{cpueqplot}
\alias{cpueqplot}
\title{A plot that illustrates the impact of catchability rescaling how the model perceives catch rates}
\usage{
cpueqplot(
  data,
  xlab = "Year",
  ylab = "Catch rate (kg/fisher day)",
  colours = c("#FFC000", "#9E480E", "#70AD47"),
  xangle = NULL,
  legend_position = NULL,
  financial_year = FALSE,
  text_size = NULL,
  legend_text_size = NULL,
  text_colour = NULL,
  legend_text_colour = NULL,
  legend_box = NULL,
  legend_title_blank = NULL,
  panel_border = NULL,
  panel_border_colour = NULL
)
}
\arguments{
\item{data}{Output from cpueqplot_prep()}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{ylab}{Label for y-axis (character). Default is "Catch rate (kg/fisher day)".}

\item{colours}{A vector of colours used for scenarios (character).}

\item{text_size}{Text size (num). Default is 12.}

\item{text_size, legend_text_size, text_colour, legend_text_colour, legend_position, legend_box, legend_title_blank, panel_border, panel_border_colour, xangle}{Optional plotting theme overrides. Defaults are controlled by \code{theme_ssand()}
and can be set globally via \code{set_ssand_style()}.}
}
\value{
A plot that illustrates the impact of catchability rescaling how the model perceives catch rates
}
\description{
A plot that illustrates the impact of catchability rescaling how the model perceives catch rates
}
