% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpueqplot_prep_DD.R
\name{cpueqplot_prep_DD}
\alias{cpueqplot_prep_DD}
\title{Prepare data for a plot that illustrates the impact of catchability rescaling how the model perceives catch rates}
\usage{
cpueqplot_prep_DD(cont_dd_mle, split_dd_mle, offset_dd_mle = NULL)
}
\arguments{
\item{cont_dd_mle}{An output of DDUST::makedd_mle(report) for a model scenario that uses continuous catch rates for the entire assessment period}

\item{split_dd_mle}{An output of DDUST::makedd_mle(report) for a model scenario that uses fleets to represent changes in fishing behaviour (e.g. fishery closures)}

\item{offset_dd_mle}{An output of DDUST::makedd_mle(report) for a model scenario that uses uses offset continuous catch rates for the entire assessment period}
}
\value{
A data frame with variables Year (numerical), cpueadjust (numerical), scenario (character)
}
\description{
Prepare data for a plot that illustrates the impact of catchability rescaling how the model perceives catch rates
}
