% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataplot.R
\name{dataplot}
\alias{dataplot}
\title{Data plot}
\usage{
dataplot(
  data,
  xlab = "Year",
  ylab = NULL,
  xbreaks = NULL,
  xlabels = NULL,
  xlim = NULL,
  ylim = NULL,
  colours = fq_palette("alisecolours"),
  size_range = c(0.5, 5),
  scenarios = 1,
  scenario_labels = NULL,
  scenario_order = NULL,
  scales = "free",
  ncol = 2,
  fleet_names = NULL,
  hollow = TRUE,
  xangle = NULL,
  legend_position = NULL,
  financial_year = FALSE,
  text_size = NULL,
  legend_text_size = NULL,
  text_colour = NULL,
  legend_text_colour = NULL,
  legend_box = NULL,
  legend_title_blank = NULL,
  panel_border = NULL,
  panel_border_colour = NULL
)
}
\arguments{
\item{data}{Output from dataplot_prep(). Columns are year (num), typename (fac), size (num), fleet (fac), scenario (int)}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{ylab}{Label for y-axis (character). Default is NULL.}

\item{xbreaks}{A vector of breaks between x-axis labels, used in ggplot2::scale_x_continous() (numeric).}

\item{xlabels}{A vector of labels for the x-axis breaks.}

\item{xlim}{A vector of lower and upper x-axis limits (e.g. c(1950, 2020)) (numeric).}

\item{ylim}{A vector of lower and upper y-axis limits (e.g. c(0,1)) (numeric).}

\item{colours}{A vector of colours used (character).}

\item{size_range}{A two-number vector specifying the minimum and maximum sizes for the circles}

\item{scenarios}{Scenario to be shown on plot (numeric). Default is 1. This was already specified in prep file, but this is a manual override to save running the prep function again.}

\item{scenario_labels}{A vector of customised scenario names (character). Default is "Scenario 1", "Scenario 2", etc.}

\item{scenario_order}{A vector to reorder how scenarios are displayed (character). Use the label names defined in "scenario_labels".
If "scenario_labels" is left blank, the labels will be "Scenario 1", "Scenario 2" etc.
Any scenarios not included in "scenario_order" will be tacked on in the order they appear in the input data.}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'free', see ?ggplot2::facet_wrap for options.}

\item{ncol}{Number of columns for facet_wrap(). Default is .}

\item{fleet_names}{A vector of customised fleet names for legend}

\item{hollow}{A logitcal to switch between two colouring styles. Default is TRUE.}

\item{xangle}{Set to 90 to rotate x-axis labels 90 degrees.}

\item{financial_year}{Set to TRUE if the assessment was based on financial year (logical). Adjusts the x-axis to show full financial year notation.}

\item{text_size, legend_text_size, text_colour, legend_text_colour, legend_position, legend_box, legend_title_blank, panel_border, panel_border_colour, xangle}{Optional plotting theme overrides. Defaults are controlled by \code{theme_ssand()}
and can be set globally via \code{set_ssand_style()}.}
}
\value{
A plot of data used in model
}
\description{
Data plot
}
\examples{
data <- dataplot_prep_SS(ss_mle)
dataplot(data)

data <- dataplot_prep_DD(dd_mle)
dataplot(data)
}
