% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataplot_prep_DD.R
\name{dataplot_prep_DD}
\alias{dataplot_prep_DD}
\title{Prepare DDUST data for dataplot()}
\usage{
dataplot_prep_DD(
  dd_mle,
  scenarios = 1,
  cpue_circle_size = 1,
  absolute_biomass_circle_size = 1,
  retained_catch_name = "Retained catch",
  abundance_indices_name = "Abundance indices",
  financial_year = FALSE
)
}
\arguments{
\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{cpue_circle_size}{Size of cpue circles (default 1)}

\item{absolute_biomass_circle_size}{Size of absolute biomass circles (default 1)}

\item{retained_catch_name}{Label for catch input (default "Retained catch")}

\item{abundance_indices_name}{Label for abundance index (default "Abundance indices")}

\item{financial_year}{Set to TRUE if the assessment was based on financial year (logical).}
}
\value{
A data object with variables called yr (int/num), typename (factor), size (numeric), fleetnames (factor)
}
\description{
Prepare DDUST data for dataplot()
}
\examples{
data <- dataplot_prep_DD(dd_mle)
dataplot(data)
}
