% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataplot_prep_SS.R
\name{dataplot_prep_SS}
\alias{dataplot_prep_SS}
\title{Prepare Stock Synthesis data for dataplot()}
\usage{
dataplot_prep_SS(ss_mle, scenarios = NULL, financial_year = FALSE)
}
\arguments{
\item{ss_mle}{A single model output from r4ss::SS_output()}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{financial_year}{Set to TRUE if the assessment was based on financial year (logical).}
}
\value{
A data object with variables called yr (int/num), typename (factor), size (numeric), fleetnames (factor)
}
\description{
Prepare Stock Synthesis data for dataplot()
}
\examples{
data <- dataplot_prep_SS(ss_mle)
dataplot(data)
}
