% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_mle.R
\docType{data}
\name{dd_mle}
\alias{dd_mle}
\title{Fake data to be used in DD_... prep functions.}
\format{
\subsection{\code{dd_mle}}{

A list of 6 lists. Each lists contains 100 variables:
\describe{
\item{annual_catch, annual_catch_sd}{Annual catch}
\item{annual_recruitment, annual_recruitment_sd}{Annual recruitment}
\item{B, B_sd}{Biomass}
...
}
}
}
\usage{
dd_mle
}
\description{
Fabricated data showing 6 scenarios of DDUST models run through
optim (maximum likelihood estimation) and tmbstan (Markov chain Monte Carlo) ...
}
\keyword{datasets}
