% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_format_logbooks.R
\name{format_logbooks}
\alias{format_logbooks}
\title{Data filtering and manipulation for logbook data
Creates columns for year, month, day, financial year and financial month,
Filters fishing method, CAAB species name, assessment time period and longitudinal and latitudinal extents.}
\usage{
format_logbooks(
  data,
  logbook_type = "QLD",
  variable_names = NULL,
  variable_names_append = NULL,
  end_year = NULL,
  end_year_financial = FALSE,
  species_code = NULL,
  species_name = NULL,
  spatial_extent = NULL,
  assessment_area = NULL,
  methods = NULL,
  other_name = "Other",
  regions_dir = NULL,
  regions_obj = NULL,
  nregions_coarse = 2,
  remove_excess_columns = FALSE,
  regions_coarse_labels = NULL
)
}
\arguments{
\item{data}{A data frame containing the variables for commercial data.}

\item{logbook_type}{(Optional) A character string to automatically enter variable_names based on pre-defined logbook conventions. Options are "QLD", "NSW" and "NSW_monthly". Not needed if using custom variable_names input.}

\item{variable_names}{(Optional) Not needed if using logbook_type. A named vector of logbook field names, named as target variable names.
For example, variable_names = c(species = "CAABSpeciesCommonName", species_code = "CaabSpeciesID", grid = "GridDerived", site = "SiteDerived", operator = "AuthorityChainNumber")
Target variable names are used in subsequent data investigation code.
Target variable names include: species, species_code, grid, site, operator, date, weight, latitude, longitude, ncrew, nboats, method, boatmark, maximum_fishing_day_count
Additional target variable names can be included.}

\item{variable_names_append}{(Optional) If using logbook_type and more variables are to be added or overwritten. A named vector of logbook field names, named as target variable names.
For example, variable_names = c(species = "CAABSpeciesCommonName", species_code = "CaabSpeciesID", grid = "GridDerived", site = "SiteDerived", operator = "AuthorityChainNumber")}

\item{end_year}{A year value specifying the end year of catch reconstruction (numeric). Used to filter data set.}

\item{end_year_financial}{A logical value indicating whether the assessment is in financial year (TRUE) or calendar year (FALSE), used to filter data temporally.}

\item{species_code}{A vector of character strings specifying the CAAB code of the species being assessed or investigated (e.g. 37441911).}

\item{species_name}{A vector of character strings specifying the CAAB species name of the species being assessed or investigated.}

\item{spatial_extent}{A numeric vector specifying the extent of the assessment in the form: c(longitude_west, longitude_east, latitude_north, latitude_south).}

\item{assessment_area}{An optional character string to override the "spatial_extent", for commonly used assessment areas. Pre-defined options are:
\itemize{
\item "EC" for Queensland east coast (c(142,155.1,-9.5,-29)),
\item "GOC" for Gulf of Carpentaria (c(138,144.6,-9.76,-18.2)),
\item "TS" for Torres Strait (c(138,155.1,-9,-12))
}}

\item{methods}{A vector of fishing method types to specify. Anything not included here will be categorised as 'Other'}

\item{other_name}{A name to call other methods e.g. 'Net + Other'. Default is 'Other'.}

\item{regions_dir}{Optional. A file path to a csv file containing a look-up table of grids to regions. File must contain columns "grid","region" and can optionally include "site" and "sitegrid"}

\item{regions_obj}{Optional. A dataframe containing a look-up table of grids to regions. Dataframe must contain columns "grid","region" and can optionally include "site" and "sitegrid"}

\item{nregions_coarse}{A numeric value for the number of coarse regions to be determined. Aggregates latitude bands into this number of groups. For example, if nregions_coarse=3, data will be grouped into "North", "Middle" and "South"}

\item{remove_excess_columns}{Set to TRUE to keep only the columns that are specified within this function and arguments. Default is FALSE.}

\item{regions_coarse_labels}{Optional. A vector of labels for coarse regions. Coarse regions are aggregated from north to south.}
}
\value{
Returns a dataframe containing a filtered, formatted catch dataset.
}
\description{
Data filtering and manipulation for logbook data
Creates columns for year, month, day, financial year and financial month,
Filters fishing method, CAAB species name, assessment time period and longitudinal and latitudinal extents.
}
\examples{
\dontrun{
QLD_commercial_filtered <- format_logbooks(data)
}
}
