% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fplot_prep_SS.R
\name{Fplot_prep_SS}
\alias{Fplot_prep_SS}
\title{Prepare Stock Synthesis data for fishing mortality plot}
\usage{
Fplot_prep_SS(
  ss_mle,
  ss_mcmc = NULL,
  scenarios = NULL,
  intervals = c(0.2, 0.5, 0.7, 0.9, 0.95),
  CI_range = 0.95,
  biomass_definition = "spawning",
  selectivity_fleet = NULL
)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{A list of outputs from r4ss::SSgetMCMC() with one element per scenario. Only needed if MCMC was used. Will automatically reformat as a list if a single r4ss::SSgetMCMC() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{intervals}{A vector of credible interval values to be displayed on banded MCMC plot (numeric). For example, "0.9" denotes the 90\% credible interval.}

\item{CI_range}{Specify credible interval range (numeric). Only activated if mcmc_style==CI.}

\item{biomass_definition}{The definition of biomass used. Options are "spawning" or "vulnerable" (character).}

\item{selectivity_fleet}{Fleet number used for the selectivity applied to calculate vulnerable biomass.}
}
\value{
A data frame with variables date, fleet, obs, exp, ub, lb, scenario
}
\description{
Prepare Stock Synthesis data for fishing mortality plot
}
\examples{
data <- Fplot_prep_SS(ss_mle,ss_mcmc)
Fplot(data)
# MCMC
data <- Fplot_prep_SS(ss_mle, ss_mcmc)
Fplot(data)

# Ensemble model
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc)
data <- Fplot_prep_SS(ss_mle, ss_mcmc_ens)
Fplot(data)
}
