% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq_palette.R
\name{fq_palette}
\alias{fq_palette}
\title{A Fisheries Queensland palette generator}
\usage{
fq_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{DAF}, \code{alisecolours},  \code{fishsex},
\code{biomass}, \code{cols}}

\item{n}{Number of colors desired}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours
}
\description{
Code adapted from karthik/wesanderson
}
\examples{
fq_palette("DAF", 4)
fq_palette("alisecolours")
fq_palette("biomass")

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- fq_palette(21, name = "DAF", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
