% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fq_palette.R
\docType{data}
\name{fq_palettes}
\alias{fq_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
fq_palettes
}
\description{
Use \code{\link{fq_palette}} to construct palettes of desired length.
}
\keyword{datasets}
