% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericplot.R
\name{genericplot}
\alias{genericplot}
\title{Generic plot}
\usage{
genericplot(
  data,
  type = "bar",
  xvar = data$year,
  yvar = data$value,
  fill_var = data$fleet,
  colour_var = data$fleet,
  shape_var = NULL,
  show_points = FALSE,
  legend_title = "Legend",
  xlab = "x axis",
  ylab = "y axis",
  xbreaks = NULL,
  ybreaks = NULL,
  xlabels = NULL,
  ylabels = NULL,
  xlim = NULL,
  ylim = NULL,
  xangle = NULL,
  colours = NULL,
  legend_position = "top",
  scales = "free",
  ncol = 2
)
}
\arguments{
\item{data}{Output from catchplot_prep(). A dataframe with date, value, fleet, scenario}

\item{type}{Specify type of plot to generate (character). Options are "bar" or "line".}

\item{xvar}{A vector to be plotted on the x-axis. See default values for example.}

\item{yvar}{A vector to be plotted on the y-axis. See default values for example.}

\item{fill_var}{A vector to be plotted on the x-axis. See default values for example.}

\item{colour_var}{A vector to guide the colour mapping. See default values for example.}

\item{shape_var}{A vector to guide the shape mapping. See default values for example.}

\item{show_points}{If type=="line", set show_points to TRUE to show points on top of the line plot.}

\item{legend_title}{Text to display as legend title (character).}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{ylab}{Label for y-axis (character). Default is "Retained catch (t)".}

\item{xbreaks}{A vector of breaks between x-axis labels, used in ggplot2::scale_x_continous() (numeric).}

\item{ybreaks}{A vector of breaks between y-axis labels, used in ggplot2::scale_y_continous() (numeric).}

\item{xlabels}{A vector of labels for the x-axis breaks.}

\item{ylabels}{A vector of labels for the y-axis breaks.}

\item{xlim}{A vector of lower and upper x-axis limits (e.g. c(1950, 2020)) (numeric).}

\item{ylim}{A vector of lower and upper y-axis limits (e.g. c(0,1)) (numeric).}

\item{xangle}{Set to 90 to rotate x-axis labels 90 degrees.}

\item{colours}{A vector of colours used (character).}

\item{legend_position}{Position of the legend ("none", "left", "right", "bottom", "top", or two-element numeric vector for x and y position). Default is "top".}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'free', see ?ggplot2::facet_wrap for options.}

\item{ncol}{Number of columns for facet_wrap(). Default is 2.}
}
\value{
Generic plot
}
\description{
Generic plot
}
\examples{
data <- data.frame()
for (scenario in 1:6) {
  for (fleet in 1:2) {
    tmp <- data.frame(year = 1980:2020,
                      value = runif(41),
                      fleet = fleet,
                      scenario = scenario)
    data <- rbind(data, tmp)
  }
}
genericplot(data)
genericplot(data, type = "line", show_points = TRUE)
}
