% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthplot_prep_SS.R
\name{growthplot_prep_SS}
\alias{growthplot_prep_SS}
\title{Prepare Stock  Synthesis data for growth plot}
\usage{
growthplot_prep_SS(ss_mle, ss_mcmc = NULL, scenarios = NULL, season = 1)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{A list of outputs from r4ss::SSgetMCMC() with one element per scenario. Only needed if MCMC was used. Will automatically reformat as a list if a single r4ss::SSgetMCMC() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{season}{A numeric value indicating the season to use (default = 1).}
}
\value{
A data frame with variables age (numeric), value (numeric), upper (numeric), and lower (numeric) and Sex (integer)
}
\description{
For MCMC:
\itemize{
\item Assumes if growth parameters are fixed in first scenario entered, they are fixed for all, and takes fixed values from first scenario listed.
\item Assumes age bins are consistent across scenarios
\item Assumes A1 and A2 are consistent across all scenarios
\item Assumes von Bertalanffy growth option selected
}
}
\examples{
data <- growthplot_prep_SS(ss_mle)
growthplot(data)

\dontrun{
# MCMC
data <- growthplot_prep_SS(ss_mle, ss_mcmc, scenarios=1)
growthplot(data)

# Ensemble model
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc)
data <- growthplot_prep_SS(ss_mle, ss_mcmc_ens)
growthplot(data)
}
}
