% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_misreporting.R
\name{heatplot}
\alias{heatplot}
\title{Heat plot
A plot displaying fishing patterns of each operator over time.
The x-axis shows the fishing operator, the y-axis shows the year, and timelines are filled by the magnitude of catch or effort.}
\usage{
heatplot(
  data,
  species_of_interest = NULL,
  fill_var = "days",
  facet_var = NULL,
  ylab = "Year",
  xlab = "Operator",
  interesting_fishers = NULL,
  interesting_years = NULL,
  filter_weight_lower = NULL,
  filter_weight_upper = NULL,
  filter_days_lower = NULL,
  filter_days_upper = NULL,
  show_operators = FALSE,
  show_values = FALSE,
  text_size = 2,
  clean_up_plots = TRUE,
  ncol = 1,
  legend_position = "top",
  legend_width = 5,
  legend_height = 0.5,
  max_fishers = 100,
  scales = "fixed",
  extract_data = FALSE
)
}
\arguments{
\item{data}{Output of format_logbooks()}

\item{species_of_interest}{The name of the species of interest, as listed in the 'species' column of data.}

\item{fill_var}{Variable by which plot is filled. Options are "days" or "catch". Default is "days".}

\item{facet_var}{Variable by which plot is to be faceted (e.g. "region")}

\item{ylab}{Label for y-axis of plot (character). Default is "".}

\item{xlab}{Label for x-axis of plot (character). Default is "".}

\item{interesting_fishers}{Optional. A vector of fishers/operators by which to filter the dataset.}

\item{interesting_years}{Optional. A vector of years by which to filter the dataset.}

\item{filter_weight_lower}{A minimum weight to be included on the plot.}

\item{filter_weight_upper}{A maximum weight to be included on the plot.}

\item{filter_days_lower}{A minimum number of days to be included on the plot.}

\item{filter_days_upper}{A maximum number of days to be included on the plot.}

\item{show_operators}{Set to TRUE to show operator names on the x-axis}

\item{show_values}{Set to TRUE to show values on the plot}

\item{text_size}{Text size}

\item{clean_up_plots}{Set to TRUE to remove axis marks and gridlines.}

\item{ncol}{Number of columns for facet wrap. Default is 2.}

\item{legend_position}{Position of legend. Default is "top"}

\item{legend_width}{Width of legend. Default is 5 (unit is cm)}

\item{legend_height}{Height of legend. Default is 0.5 (unit is cm)}

\item{max_fishers}{Maximum number of fishers to show on a single plot}

\item{scales}{Scaled for facet_wrap. Default is 'fixed'}

\item{extract_data}{Set to TRUE to return data instead of plot. Default is FALSE.}
}
\value{
Heat plot
}
\description{
Heat plot
A plot displaying fishing patterns of each operator over time.
The x-axis shows the fishing operator, the y-axis shows the year, and timelines are filled by the magnitude of catch or effort.
}
\examples{
\dontrun{heatplot(data = format_logbooks(raw_data))}
}
