% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influenceplot.R
\name{influenceplot}
\alias{influenceplot}
\title{Influence plot}
\usage{
influenceplot(
  data,
  step = TRUE,
  model_names = NULL,
  colours = NULL,
  xlab = "Year",
  xlim = NULL,
  xbreaks = NULL,
  ylab = "Catch rate (kg/day)",
  ylim = NULL,
  ybreaks = NULL,
  legend_position = "top",
  show_CI = FALSE
)
}
\arguments{
\item{data}{A data frame prepared outside of SSAND, containing CPUE model outputs for at least one CPUE model.
Columns should be year (int), cpue (num), cpue_lwr (num), cpue_upr (num), model (num)}

\item{step}{Set to TRUE to step out addition of influences. If FALSE, all models are on one panel.}

\item{model_names}{A vector of labels for CPUE models (e.g. c("Raw", "Add season", "Add fishing power"))}

\item{colours}{A vector of colours used (character).}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{xlim}{A vector of lower and upper x-axis limits (e.g. c(1950, 2020)) (numeric).}

\item{xbreaks}{A vector of breaks between x-axis labels, used in ggplot2::scale_x_continous() (numeric).}

\item{ylab}{Label for y-axis (character). Default is "Catch rate (kg/day)".}

\item{ylim}{A vector of lower and upper y-axis limits (e.g. c(0,1)) (numeric).}

\item{ybreaks}{A vector of breaks between y-axis labels, used in ggplot2::scale_y_continous() (numeric).}

\item{legend_position}{Position of the legend ("none", "left", "right", "bottom", "top", or two-element numeric vector for x and y position). Default is "top".}

\item{show_CI}{Set to TRUE to show a confidence intervals (logical).}
}
\value{
A plot showing the influence of each term in a CPUE model
}
\description{
Influence plot
}
\examples{
data <- data.frame(year=rep(1988:2024,6),
                   cpue = c(sin(1988:2024 / 10),
                            cos(1988:2024 / 10),
                            cos(sin(1988:2024 / 10)),
                            sin(cos(1988:2024 / 10)),
                            sin(cos(sin(1988:2024 / 10))),
                            cos( sin(cos(1988:2024 / 10)))),
                   cpue_lwr = c(sin(1988:2024 / 10) /1.1,
                                cos(1988:2024 / 10)  /1.1,
                                cos(sin(1988:2024 / 10))  /1.1,
                                sin(cos(1988:2024 / 10)) /1.1,
                                sin(cos(sin(1988:2024 / 10)) /1.1),
                                cos(sin(cos(1988:2024 / 10)))/1.1),
                   cpue_upr = c(sin(1988:2024 / 10) /1.1,
                                cos(1988:2024 / 10)  /0.9,
                                cos(sin(1988:2024 / 10))  /0.9,
                                sin(cos(1988:2024 / 10)) /0.9,
                                sin(cos(sin(1988:2024 / 10)) /0.9),
                                cos(sin(cos(1988:2024 / 10)))/0.9),
                   model=c(rep(1,37),rep(2,37),rep(3,37),
                              rep(4,37),rep(5,37),rep(6,37)))

influenceplot(data)
influenceplot(data, step=FALSE)
influenceplot(data, step=FALSE, show_CI=TRUE)
influenceplot(data, show_CI=TRUE)
influenceplot(data, model_names = c("Raw","Add A","Add B",
                                     "Add C","Add D","Add E"))
}
