% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate.R
\name{investigate}
\alias{investigate}
\title{Investigate raw catch and effort data}
\usage{
investigate(
  data,
  dir = getwd(),
  species_of_interest,
  financial_year = FALSE,
  render = TRUE,
  anonymous = FALSE,
  maximum_fishing_day_count = 1,
  interesting_years = NULL,
  interesting_fishers = NULL,
  show_raw_catch = TRUE,
  show_raw_effort = TRUE,
  show_raw_cpue = TRUE,
  show_regional_overview = TRUE,
  show_fishing_methods = TRUE,
  show_cocaught_species = TRUE,
  show_seasonality = TRUE,
  show_misreporting = TRUE,
  show_unusual_catch = FALSE,
  show_top_fishers = TRUE,
  show_maps = FALSE,
  show_annual_maps = FALSE,
  show_monthly_maps = FALSE,
  show_method_maps = FALSE,
  metadata = NULL,
  upset_n_trips = 25,
  coast_directory = NULL,
  filter_days_lower = 10
)
}
\arguments{
\item{data}{Output from format_logbooks() applied to raw logbook data.}

\item{dir}{Directory of file outputs (.Rmd, .html, and plots and tables). Default is working directory.}

\item{species_of_interest}{The name of the species of interest, as listed in the 'species' column of data. Can use a vector of species to duplicate plots for each species.}

\item{financial_year}{Set to TRUE to display as financial year}

\item{render}{Set to TRUE to render report. Set to FALSE to only produce Rmd file. If render = FALSE, you will need to manually load your data set into the first chunk of the Rmd file, ensuring it is called 'data' to feed into the subsequent chunks.}

\item{anonymous}{Set to TRUE to remove fisher/operator names.}

\item{maximum_fishing_day_count}{Default is 1. Filters data for trips of this maximum number of days, for all plots except those that explore the range of maximum number of fishing days.}

\item{interesting_years}{Optional. A vector of years to highlight in additional plots.}

\item{interesting_fishers}{Optional. A vector of fishers/operators to highlight in additional plots.}

\item{show_raw_catch}{Set to TRUE to show raw catch plots.}

\item{show_raw_effort}{Set to TRUE to show raw effort plots.}

\item{show_raw_cpue}{Set to TRUE to show raw CPUE plots.}

\item{show_regional_overview}{Set to TRUE to show regional overview. Default is TRUE.}

\item{show_fishing_methods}{Set to TRUE to show fishing method plots.}

\item{show_cocaught_species}{Set to TRUE to show plots. Default is TRUE.}

\item{show_seasonality}{Set to TRUE to show seasonality plots. Default is TRUE.}

\item{show_misreporting}{Set to TRUE to show misreporting plots. Default is TRUE.}

\item{show_unusual_catch}{Set to TRUE to show unusually high catch. Default is FALSE.}

\item{show_top_fishers}{Set to TRUE to show top fisher plots. Default is TRUE.}

\item{show_maps}{Set to TRUE to show maps. Default is TRUE. Not that if you're accessing from outside the QDAF network you will have to customise the directory for coastline shapefiles.}

\item{show_annual_maps}{Set to TRUE to show maps faceted by year. Default is FALSE, time intensive.}

\item{show_monthly_maps}{Set to TRUE to show maps faceted by month. Default is FALSE, time intensive.}

\item{show_method_maps}{Set to TRUE to show maps faceted by fishing method. Default is FALSE, time intensive.}

\item{metadata}{(Optional) A character string that will be entered into the Metadata tab}

\item{upset_n_trips}{Cut off value for tail end of upset plot. Do not plot species combinations who frequency is below this value. Default is 25. Depending on the dataset, a lower number might mean the plot takes a very long time, or renders with unreadable dimensions. A higher number might mean that no combinations are shown and the plot may break.}

\item{coast_directory}{(Optional) A file path for a shapefile containing a coastline to be used in maps.}

\item{filter_days_lower}{A minimum number of days to be included on the heat plot for identifying potential misreporting. Default is 10.}
}
\value{
If render==FALSE, returns investigate.Rmd at the specified directory. If render==TRUE, this Rmd is compiled and investigation.html and a folder called "investigate", containing all produced plots and tables, are generated in the specified directory.
}
\description{
This function produces a dashboard to visualise raw catch and effort data.
It produces a document (Rmd, which complies to a HTML) that shows various data summaries and investigations.
If render = FALSE, you will produce just the Rmd (not HTML). This allows you to then tweak the document before rendering. Note that you will need to manually load your data set into the first chunk of the Rmd file, ensuring it is called 'data' to feed into the subsequent chunks.
This function can take a long time to run if using a large dataset. We suggest you test the function on a smaller dataset to ensure it is configured and operating how you intend it to.
Sections can be toggled on and off in order to save compilation time.
Plots and tables are saved to a folder called "investigate" within the specified directory.
}
\examples{
\dontrun{
investigate(format_logbooks(logbooks),
species_of_interest = "Glitterfin snapper",
filter_days_lower = 1,
upset_n_trips = 1)
}
}
