% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_biology.R
\name{investigate_biology}
\alias{investigate_biology}
\title{Investigate raw and worked-up biological data
Produces a document (Rmd, complies to a HTML) that shows various data summaries and investigations.
Plots and tables are saved to a folder called "investigate_biology" within the specified directory.
Sections can be toggled on and off in order to save compilation time.}
\usage{
investigate_biology(
  data,
  year_var = "year",
  month_var = "month",
  length_var = "fl",
  length_label = "Fork length (cm)",
  age_var = "agegrp",
  age_label = "Age group (years)",
  region_var = "region",
  sex_var = "sexcode",
  method_var = "FishingMethodCode",
  sector_var = "sector",
  MLS = NULL,
  show_length = TRUE,
  show_age = TRUE,
  show_age_at_length = TRUE,
  male = "1 - Male",
  female = "2 - Female",
  unknown_sex = "5 - Unknown",
  dir = getwd(),
  render = FALSE
)
}
\arguments{
\item{data}{A table of raw length and age data. Each row is a different sample. Columns are labelled according to your set up in the "_var" arguments of this function.}

\item{year_var}{The name of the year variable used. Default is 'year'}

\item{month_var}{The name of the month variable used. Default is 'month'}

\item{length_var}{The name of the length variable used. Default is 'fl'}

\item{length_label}{The label for the length variable. Default is "Fork length (cm)"}

\item{age_var}{The name of the age variable used. Default is 'agegrp'}

\item{age_label}{The label for the age variable. Default is "Age group (years)"}

\item{region_var}{The name of the region variable used. Default is 'region'}

\item{sex_var}{The name of the sex variable used. Default is 'sexcode'}

\item{method_var}{The name of the fishing method variable used. Default is 'FishingMethodCode'}

\item{sector_var}{The name of the sector variable used. Default is 'sector'}

\item{MLS}{Optional. The value of the minimum legal size. Numeric}

\item{show_length}{Logical. Set to TRUE to show length plots. Default is TRUE.}

\item{show_age}{Logical. Set to TRUE to show length plots. Default is TRUE.}

\item{show_age_at_length}{Logical. Set to TRUE to show length plots. Default is TRUE.}

\item{male}{Label given to male fish in the sex_var column (default is "1 - Male")}

\item{female}{Label given to female fish in the sex_var column (default is "2 - Female")}

\item{unknown_sex}{Label given to fish of unknown sex in the sex_var column (default is "= "5 - Unknown")}

\item{dir}{Directory of file outputs (.Rmd, .html, and plots and tables). Default is working directory.}

\item{render}{Set to TRUE to render report. Set to FALSE to only produce Rmd file.}
}
\value{
If render==FALSE, returns investigate.Rmd at the specified directory. If render==TRUE, this Rmd is compiled and investigation.html and a folder called "investigate", containing all produced plots and tables, are generated in the specified directory.
}
\description{
Investigate raw and worked-up biological data
Produces a document (Rmd, complies to a HTML) that shows various data summaries and investigations.
Plots and tables are saved to a folder called "investigate_biology" within the specified directory.
Sections can be toggled on and off in order to save compilation time.
}
