% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthplot.R
\name{lengthplot}
\alias{lengthplot}
\title{Length composition plot with an option to add model fits}
\usage{
lengthplot(
  data,
  fleet = 1,
  scenario = 1,
  partition = 0,
  point_size = 1,
  colours = c("#AD3D25", "#0085B8", "#FCC17E"),
  financial_year = FALSE,
  xlab = "Length (cm)",
  ylab = "Proportion",
  xbreaks = NULL,
  ybreaks = NULL,
  xlabels = NULL,
  ylabels = NULL,
  xlim = NULL,
  ylim = NULL,
  scales = "fixed",
  ncol = 4,
  direction = "v",
  MLS = NULL,
  MLS_colour = "black",
  MLS_label = "MLS",
  show_fits = TRUE,
  text_size = NULL,
  legend_text_size = NULL,
  text_colour = NULL,
  legend_text_colour = NULL,
  legend_position = NULL,
  legend_box = NULL,
  legend_title_blank = NULL,
  panel_border = NULL,
  panel_border_colour = NULL,
  xangle = NULL
)
}
\arguments{
\item{data}{Output from lengthplot_prep(). A data frame with year (int), fleet (int), bin (int), sex (chr), obs (num), exp (num), scenario (int)}

\item{fleet}{Specify which fleet to plot (numeric). By default, fleet 1 will be shown.}

\item{scenario}{A scenario number to be shown on plot (numeric). This was already specified in prep file, but this is a manual override to save running the prep function again.}

\item{partition}{Specify which partition to plot (numeric). By default, partition 0 will be shown.}

\item{point_size}{Size of points used in ggplot2::geom_line(). Default is 1.}

\item{colours}{A vector of colours used for sex types (character).}

\item{financial_year}{Set to TRUE if the assessment was based on financial year (logical). Adjusts the x-axis to show full financial year notation.}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{ylab}{Label for y-axis (character). Default is "Spawning biomass (relative)".+}

\item{xbreaks}{A vector of breaks between x-axis labels, used in ggplot2::scale_x_continous() (numeric).}

\item{ybreaks}{A vector of breaks between y-axis labels, used in ggplot2::scale_y_continous() (numeric).}

\item{xlabels}{A vector of labels for the x-axis breaks.}

\item{ylabels}{A vector of labels for the y-axis breaks.}

\item{xlim}{A vector of lower and upper x-axis limits (e.g. c(1950, 2020)) (numeric).}

\item{ylim}{A vector of lower and upper y-axis limits (e.g. c(0,1)) (numeric).}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'free', see ?ggplot2::facet_wrap for options.}

\item{ncol}{Number of columns for facet_wrap(). Default is 2.}

\item{direction}{Direction of facet wrap (character). Enter "v" for vertical or "h" for horizontal.}

\item{MLS}{Value of minimum legal size (or equivalent) (numeric). Leave blank to exclude from plot.}

\item{MLS_colour}{Colour of MLS line (character). Default is "black".}

\item{MLS_label}{Label of MLS line (character). Default is "MLS".}

\item{show_fits}{Set to TRUE to show model fits.}

\item{text_size, legend_text_size, text_colour, legend_text_colour, legend_position, legend_box, legend_title_blank, panel_border, panel_border_colour, xangle}{Optional plotting theme overrides. Defaults are controlled by \code{theme_ssand()}}
}
\value{
A length plot, with an option to add model fits
}
\description{
Length composition plot with an option to add model fits
}
\examples{
data <- lengthplot_prep_SS(ss_mle)
lengthplot(data, show_fits=FALSE)
lengthplot(data)
}
