% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodprofileplot_component_options.R
\name{likelihoodprofileplot_component_options}
\alias{likelihoodprofileplot_component_options}
\title{Options for component argument of likelihoodprofileplot_prep_SS}
\usage{
likelihoodprofileplot_component_options(profile_input)
}
\arguments{
\item{profile_input}{The likelihood profile object after running: r4ss::profile(), r4ss::SSgetoutput(), then r4ss::SSsummarize(). See ?likelihoodprofileplot for details.}
}
\value{
A vector of options that can be used in the component argument of likelihoodprofileplot_prep_SS()
}
\description{
Options for component argument of likelihoodprofileplot_prep_SS
}
\examples{
\dontrun{
likelihoodprofileplot_component_options(profile_input)
data <- likelihoodprofileplot_prep_SS(profile_input)
likelihoodprofileplot(data)
}
}
