% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodprofileplot_prep_SS.R
\name{likelihoodprofileplot_prep_SS}
\alias{likelihoodprofileplot_prep_SS}
\title{Prepare Stock Synthesis data for likelihoodprofileplot()
View ?likelihoodprofileplot to view full examples for running likelihood profiles for Stock Synthesis or DDUST.}
\usage{
likelihoodprofileplot_prep_SS(
  profile_input,
  component = c("Total", "Index data", "Length data", "Age data", "Recruitment"),
  parameter = "SR_LN(R0)"
)
}
\arguments{
\item{profile_input}{The likelihood profile object after running: r4ss::profile(), r4ss::SSgetoutput(), then r4ss::SSsummarize(). See example for details.}

\item{component}{The components for which you want to display the likelihood profile.
Use SSAND::likelihoodprofileplot_component_options(profile_input) to see options.
The default is c("Total","Index data", "Length data", "Age data", "Recruitment").}

\item{parameter}{Parameter used in likelihood profile. Use unique(profile_input$pars$Label) for valid options.}
}
\value{
A dataframe with columns x_vector, component and likelihood
}
\description{
Prepare Stock Synthesis data for likelihoodprofileplot()
View ?likelihoodprofileplot to view full examples for running likelihood profiles for Stock Synthesis or DDUST.
}
\examples{
\dontrun{
r4ss::profile(dir = '.', # directory of 4 SS files
              oldctlfile = "control.ctl",
              newctlfile = "control.ctl",
              string = "steep",
              profilevec = c(0.4,0.5,0.6),
              exe = "C:/stocksynthesis/ss_3.30.22.exe")

profile_input <- r4ss::SSsummarize(
  r4ss::SSgetoutput(dirvec = ".",
                    keyvec = 1:3, # 1:length(profilevec)
                    getcovar = FALSE,
                    getcomp = FALSE))


data <- likelihoodprofileplot_prep_SS(profile_input, parameter="SR_BH_steep")
likelihoodprofileplot(data)
}
}
