% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodtable_prep_SS.R
\name{likelihoodtable_prep_SS}
\alias{likelihoodtable_prep_SS}
\title{Prepare Stock Synthesis data for likelihood table}
\usage{
likelihoodtable_prep_SS(ss_mle, scenarios = NULL, components = NULL)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{components}{A vector of likelihood components to show. Options are "TOTAL", "Catch", "Equil_catch", "Survey", "Discard", "Length_comp", "Age_comp", "Recruitment", "InitEQ_Regime", "Forecast_Recruitment", "Parm_priors", "Parm_softbounds", "Parm_devs", "Crash_Pen"}
}
\value{
A data frame that summarises likelihood components
}
\description{
Prepare Stock Synthesis data for likelihood table
}
\examples{
dat <- likelihoodtable_prep_SS(ss_mle)
dat <- likelihoodtable_prep_SS(ss_mle,
  components = c("TOTAL", "Survey", "Length_comp","Age_comp","Recruitment"))
}
