% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbooks.R
\docType{data}
\name{logbooks}
\alias{logbooks}
\title{Fake data to be used in investigate functions.}
\format{
\subsection{\code{logbooks}}{

A sample set of logbook data
}
}
\usage{
logbooks
}
\description{
A sample set of logbook data
}
\keyword{datasets}
