% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managementplot.R
\name{managementplot}
\alias{managementplot}
\title{Management plot}
\usage{
managementplot(
  data,
  management_text = NULL,
  management_pos = NULL,
  base_case,
  xlab = "Year",
  ylab = "Biomass (relative to unfished)",
  xlim = c(min(data$year) - 1, max(data$year) + 1),
  ylim = NULL,
  xbreaks = NULL,
  title = "Species (YEAR) east coast",
  subtitle = "Stock assessment biomass estimate and key management actions",
  title_position = NULL,
  subtitle_position = NULL,
  management_line_width = 7,
  label_offset = 0.3,
  legend_position = c(0.1, 0.4),
  show_result = TRUE,
  scenarios = base_case,
  show_target = TRUE,
  show_limit = TRUE,
  show_final_biomass = TRUE,
  custom_reference_values = NULL,
  custom_reference_colours = c("#FFC000"),
  custom_reference_labels = c("B40"),
  draft_text = "DRAFT",
  draft_position = c(1970, 0.1),
  draft_colour = "red",
  logo_scale = 10,
  logo_padding = 0.05,
  infographic_path = NA,
  infographic_position = c(0.5, 0.4),
  infographic_scale = NULL,
  plotdir = paste0(getwd(), "/management_plot.png"),
  mcmc_style = "banded",
  band_colour = "black",
  show_median = "none",
  legend_name = "Legend",
  legend_entries = c("Biomass estimate", "Range of biomass estimates",
    "PLEASE CUSTOMISE legend_entries", "x", "x", "x", "x", "x", "x", "x", "x", "x"),
  boxplot_outliers = TRUE
)
}
\arguments{
\item{data}{Output of SSAND::biomassplot_prep(). MCMC or MLE inputs are accepted into either. See ?biomassplot_prep_DD or ?biomassplot_prep_SS for more information.}

\item{management_text}{A vector of management measures over time, each starting with the year and a colon. For example '2004: Rezoning of GBRMP'. Use back-slash followed by n to create a line-break}

\item{management_pos}{A vector of vertical positions, corresponding to the management info. Defaults to 1 for each management measure.}

\item{base_case}{The scenario number that pertains to the base case (numeric). If NA, no base case was selected.}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{ylab}{Label for y-axis (character). Default is "Biomass (relative to unfished)".}

\item{xlim}{A vector of lower and upper x-axis limits (e.g. c(1950, 2020)) (numeric).}

\item{ylim}{A vector of lower and upper y-axis limits (e.g. c(0,1)) (numeric).}

\item{xbreaks}{A vector of breaks between x-axis labels, used in ggplot2::scale_x_continous() (numeric).}

\item{title}{Title of plot (recommend: "Species name (YEAR) stock")}

\item{subtitle}{Subtitle of plot (recommend: "Stock assessment biomass estimate and key management actions")}

\item{title_position}{Vertical position of title}

\item{subtitle_position}{Vertical position of subtitle}

\item{management_line_width}{Width of management line (recommend: 7)}

\item{label_offset}{Amount of offset for management labels (recommend: 0.3)}

\item{legend_position}{Two-element vector with proportional x and y position of top left corner of legend (default c(0.1, 0.4))}

\item{show_result}{Set to TRUE if a result was found for the assessment (logical). Set to FALSE if no result was found (e.g. 2023 red spot king prawn)}

\item{scenarios}{A vector of scenario numbers to be shown on plot (numeric). This was already specified in prep file, but this is a manual override to save running the prep function again.}

\item{show_target}{Flag to include 60\% target reference line (Boolean)}

\item{show_limit}{Flag to include 20\% limit reference line (Boolean)}

\item{show_final_biomass}{Flag to show final biomass at the end of the trajectory. Only works if result==TRUE}

\item{custom_reference_values}{Vector of heights for custom reference lines (e.g. c(0.4))}

\item{custom_reference_colours}{Vector of colours for custom reference lines (e.g. c("#FFC000"))}

\item{custom_reference_labels}{Vector of labels for custom reference lines (e.g. c('B40')}

\item{draft_text}{Watermark text (chr)}

\item{draft_position}{Vector of x and y position of watermark}

\item{draft_colour}{Colour of watermark}

\item{logo_scale}{Scale (as a percentage) of QG logo for overlay onto plot (recommend: 7, adjusting as necessary)}

\item{logo_padding}{Padding around QG logo as percentage of plot width and height (recommend: 0.05)}

\item{infographic_path}{Directory of infographic to be overlaid onto plot (use NA if not infographic exists)}

\item{infographic_position}{Vector of x and y position of top-left corner of infographic, as fraction of plot dimensions (e.g. c(0.5, 0.4))}

\item{infographic_scale}{Scale (multiplier) of infographic for overlay onto plot (defaults to 25\% the plot width)}

\item{plotdir}{Filepath, including file name and extension (.png) of where final management plot should be saved}

\item{mcmc_style}{The type of MCMC plot to be displayed (character). Options are "banded", "hairy", "boxplot", "CI" and "joy", the default is "banded". Only one option can be selected.}

\item{band_colour}{Colour of bands (character). Only used when mcmc_style=="banded". Input one colour, bands will be distinguished using an alpha.}

\item{show_median}{Type of median shown. Default "annual_biomass" shows the median of each year,
"trajectory" shows median trajectory based on biomass in final year,
"parameters" uses median of each parameter (not yet implemented)}

\item{legend_name}{Name of legend}

\item{legend_entries}{A vector of labels for the legend.}

\item{boxplot_outliers}{Set to FALSE to remove outlier points from boxplot. Default is TRUE.}
}
\value{
A management plot, saved in the location specified in plotdir
}
\description{
Management plot
}
\examples{
management_text <- c(
"1980: Management measure",
"1988: Management measure",
"2004: - Many things \n - happened in \n - this year",
"2010: Management measure",
"2021: - Many things \n - happened in \n - this year")

management_pos <- c(0.8, 0.6, 1.0, 0.5, 0.2)

data <- biomassplot_prep_SS(ss_mle)
# OR: data <- biomassplot_prep_SS(ss_mle, ss_mcmc)
# OR: data <- biomassplot_prep_DD(dd_mle)
# OR: data <- biomassplot_prep_DD(dd_mle, dd_mcmc, dd_sim)

# Produce a management plot with a base case (scenario 1), result found,
# and the cloud is determined from the base case uncertainty:
\dontrun{
managementplot(data, management_text, management_pos,
                base_case = 1, show_result = TRUE, scenarios = 1)
}

# Produce a management plot with a base case (scenario 1), result found,
# and the cloud is determined from scenario uncertainty:
\dontrun{
managementplot(data, management_text, management_pos, base_case = 1,
                show_result = TRUE, scenarios = unique(data$scenario))
}

# Produce a management plot where no result was found (and therefore no base case),
# so the cloud is replaced with individual spaghetti strands:
\dontrun{
managementplot(data,management_text, management_pos, base_case = NA,
                show_result = FALSE, scenarios = unique(data$scenario))
}
}
