% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_ensemble_SS.R
\name{mcmc_ensemble_SS}
\alias{mcmc_ensemble_SS}
\title{MCMC ensemble preparation for Stock Synthesis}
\usage{
mcmc_ensemble_SS(ss_mcmc, scenarios = NULL, weighting = NULL)
}
\arguments{
\item{ss_mcmc}{A list of outputs from r4ss::SSgetMCMC() with one element per scenario. Only needed if MCMC was used. Will automatically reformat as a list if a single r4ss::SSgetMCMC() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenarios to be included in the ensemble.}

\item{weighting}{A vector of integers to weight each scenario, one integer per scenario.
The MCMC iterations for each scenario are repeated the number of times indicated by the corresponding element in the weighting vector.
For example, for an ensemble with four scenarios, to double the weighting of the first scenario you would enter weighting=c(2,1,1,1).}
}
\value{
A list containing one dataframe of all specified ss_mcmc objects combined, limited to columns that are common between ss_mcmc objects, duplicated as per the specified weighting.
An extra column "ensemble" is appended, for datachecks in other functions.
}
\description{
MCMC ensemble preparation for Stock Synthesis
}
\examples{
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc,scenarios=c(1,2))
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc,scenarios=c(1,2), weighting=c(3,1))
data <- biomassplot_prep_SS(ss_mle,ss_mcmc_ens)
biomassplot(data)
}
