% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_finalbiomassposteriorplot_prep_DD.R
\name{mcmc_finalbiomassposterior_prep_DD}
\alias{mcmc_finalbiomassposterior_prep_DD}
\title{Prepare DDUST data for mcmc_finalbiomassposteriorplot}
\usage{
mcmc_finalbiomassposterior_prep_DD(
  dd_mle,
  dd_mcmc,
  dd_sim,
  credible_interval = 0.95,
  end_year = NULL,
  cut = TRUE
)
}
\arguments{
\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{dd_mcmc}{A list of model fits from tmbstan::tmbstan() with one list element per scenario. Only needed if MCMC was used.}

\item{dd_sim}{A list of outputs from SSAND::simulate_DD() with one list element per scenario. Only required if MCMC was used. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{credible_interval}{The credible interval to use to create dashed lines in the final plot}

\item{end_year}{The final year of biomass to be plotted. Calculated by default, but option is included in case model is run into the future.}

\item{cut}{Set to TRUE to limit the modeled density to the range of biomass estimates produced by the MCMC. If FALSE, the plot may place density outside of the range of biomass values estimated by the MCMC.}
}
\value{
A list of 5 objects
}
\description{
Prepare DDUST data for mcmc_finalbiomassposteriorplot
}
\examples{
\dontrun{
# MCMC model
library(DDUST)
dd_sim <- simulate_DDUST(dd_mle,dd_mcmc)
data <- mcmc_finalbiomassposterior_prep_DD(dd_mle,dd_mcmc,
                                              dd_sim,scenarios=1)
mcmc_finalbiomassposteriorplot(data)

# Ensemble model:
library(DDUST)
dd_sim <- simulate_DDUST(dd_mle,dd_mcmc)

dd_mcmc_ens <- mcmc_ensemble_DD(dd_mcmc,dd_sim,scenarios=c(1,2))$dd_mcmc
dd_sim_ens <- mcmc_ensemble_DD(dd_mcmc,dd_sim,scenarios=c(1,2))$dd_sim
data <- mcmc_finalbiomassposterior_prep_DD(dd_mle,dd_mcmc_ens,
                                              dd_sim_ens,scenarios=1)
mcmc_finalbiomassposteriorplot(data)
}
}
