% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_finalbiomassposteriorplot_prep_SS.R
\name{mcmc_finalbiomassposterior_prep_SS}
\alias{mcmc_finalbiomassposterior_prep_SS}
\title{Prepare Stock Synthesis data for mcmc_finalbiomassposteriorplot}
\usage{
mcmc_finalbiomassposterior_prep_SS(
  ss_mle,
  ss_mcmc,
  credible_interval = 0.95,
  end_year = NULL,
  cut = TRUE
)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario, or a single representative scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{Output of mcmc_ensemble_SS(). A list containing one dataframe of all specified ss_mcmc objects combined, limited to columns that are common between ss_mcmc objects, duplicated as per the specified weighting.
An extra column "ensemble" is appended, for data checks in other functions.}

\item{credible_interval}{The credible interval to use to create dashed lines in the final plot}

\item{end_year}{The final year of biomass to be plotted. Calculated by default, but option is included in case model is run into the future.}

\item{cut}{Set to TRUE to limit the modeled density to the range of biomass estimates produced by the MCMC. If FALSE, the plot may place density outside of the range of biomass values estimated by the MCMC.}
}
\value{
A list of 5 objects
}
\description{
Prepare Stock Synthesis data for mcmc_finalbiomassposteriorplot
}
\examples{
# MCMC model
data <- mcmc_finalbiomassposterior_prep_SS(ss_mle,ss_mcmc)
mcmc_finalbiomassposteriorplot(data)

# Ensemble model
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc)
data <- mcmc_finalbiomassposterior_prep_SS(ss_mle,ss_mcmc_ens)
mcmc_finalbiomassposteriorplot(data)
}
