% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_median_position_DD.R
\name{mcmc_median_position_DD}
\alias{mcmc_median_position_DD}
\title{Extract DDUST MCMC iteration position that produces median final biomass}
\usage{
mcmc_median_position_DD(dd_mle, dd_mcmc, dd_sim, scenarios = NULL)
}
\arguments{
\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{dd_mcmc}{A list of model fits from tmbstan::tmbstan() with one list element per scenario. Only needed if MCMC was used.}

\item{dd_sim}{A list of outputs from SSAND::simulate_DD() with one list element per scenario. Only required if MCMC was used. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{scenarios}{A vector of scenario numbers to be be included in median calculation (numeric). Function will return one median for each scenario.}
}
\value{
A vector of numbers with one entry per scenario entered, where each number is the corresponding MCMC run that produced the median final biomass ratio.
}
\description{
Extract DDUST MCMC iteration position that produces median final biomass
}
\examples{
\dontrun{
library(DDUST)
dd_sim <- simulate_DDUST(dd_mle,dd_mcmc)
pos <- mcmc_median_position_DD(dd_mle, dd_mcmc, dd_sim)

# Ensemble model
dd_mcmc_ens <- mcmc_ensemble_DD(dd_mcmc,dd_sim,scenarios=c(1,2))$dd_mcmc
dd_sim_ens <- mcmc_ensemble_DD(dd_mcmc,dd_sim,scenarios=c(1,2))$dd_sim
pos <- mcmc_median_position_DD(dd_mle,dd_mcmc_ens,dd_sim_ens)[1]
}
}
