% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_median_position_SS.R
\name{mcmc_median_position_SS}
\alias{mcmc_median_position_SS}
\title{Extract Stock Synthesis MCMC iteration position that produces median final biomass}
\usage{
mcmc_median_position_SS(ss_mle, ss_mcmc, scenarios = NULL)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{A list of outputs from r4ss::SSgetMCMC() with one element per scenario. Only needed if MCMC was used. Will automatically reformat as a list if a single r4ss::SSgetMCMC() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenario numbers to be be included in median calculation (numeric). Function will return one median for each scenario.}
}
\value{
A vector of numbers with one entry per scenario entered, where each number is the corresponding MCMC run that produced the median final biomass ratio.
}
\description{
Extract Stock Synthesis MCMC iteration position that produces median final biomass
}
\examples{
pos <- mcmc_median_position_SS(ss_mle, ss_mcmc)

# Ensemble model
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc)
pos <- mcmc_median_position_SS(ss_mle, ss_mcmc_ens)
}
