% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_posteriordensityplot_prep_DD.R
\name{mcmc_posteriordensityplot_prep_DD}
\alias{mcmc_posteriordensityplot_prep_DD}
\title{Posterior DDUST density plot for MCMC run}
\usage{
mcmc_posteriordensityplot_prep_DD(
  dd_mle,
  dd_mcmc,
  dd_sim,
  scenario = 1,
  other_quantities = c()
)
}
\arguments{
\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{dd_mcmc}{A list of model fits from tmbstan::tmbstan() with one list element per scenario. Only needed if MCMC was used.}

\item{dd_sim}{A list of outputs from SSAND::simulate_DD() with one list element per scenario. Only required if MCMC was used. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{scenario}{A single number representing the scenario to plot (numeric). Shows first scenario if left blank. Can be overridden in the plotting function.}

\item{other_quantities}{Other derived quantities from dd_mle to plot, see names(dd_mle) for options}
}
\value{
Posterior density plot
}
\description{
Posterior DDUST density plot for MCMC run
}
\examples{
\dontrun{
library(DDUST)
dd_sim <- simulate_DDUST(dd_mle,dd_mcmc)
data <- mcmc_posteriordensityplot_prep_DD(dd_mle, dd_mcmc, dd_sim, scenario = 1)
mcmc_posteriordensityplot(data)
}
}
