% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_posteriordensityplot_prep_SS.R
\name{mcmc_posteriordensityplot_prep_SS}
\alias{mcmc_posteriordensityplot_prep_SS}
\title{Posterior Stock Synthesis density plot for MCMC run}
\usage{
mcmc_posteriordensityplot_prep_SS(
  ss_mle,
  ss_mcmc,
  scenario = NULL,
  parameters,
  likelihoods = c("TOTAL"),
  show_objective_function = FALSE
)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{A list of outputs from r4ss::SSgetMCMC() with one element per scenario. Only needed if MCMC was used. Will automatically reformat as a list if a single r4ss::SSgetMCMC() output (i.e. one scenario) is entered.}

\item{scenario}{A single number representing the scenario to plot (numeric). Shows first scenario if left blank. Can be overridden in the plotting function.}

\item{parameters}{A vector of parameter to include on plot. Hint: use SSAND::extract_SS_parameters() to find a list of options.}

\item{likelihoods}{A vector to specify additional likelihoods to include on plot}

\item{show_objective_function}{Set to TRUE to include objective function on posterior density plot.}
}
\value{
data to produce density plot, trace plot
}
\description{
Posterior Stock Synthesis density plot for MCMC run
}
\examples{
parameters <- extract_SS_parameters(ss_mcmc)[c(2:10,449),]
data <- mcmc_posteriordensityplot_prep_SS(ss_mle, ss_mcmc, scenario = 1, parameters)
mcmc_posteriordensityplot(data)
}
