% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_rhatplot.R
\name{mcmc_rhatplot}
\alias{mcmc_rhatplot}
\title{Plot of Rhat values for each parameter}
\usage{
mcmc_rhatplot(
  data,
  ncol = 3,
  parameter_labels = NULL,
  ylab = "Parameter",
  show_point = FALSE,
  scenarios = NULL,
  scenario_labels = NULL,
  scenario_order = NULL,
  xangle = NULL,
  legend_position = NULL,
  financial_year = FALSE,
  text_size = NULL,
  legend_text_size = NULL,
  text_colour = NULL,
  legend_text_colour = NULL,
  legend_box = NULL,
  legend_title_blank = NULL,
  panel_border = NULL,
  panel_border_colour = NULL
)
}
\arguments{
\item{data}{Output from mcmc_rhatplot_prep() with columns parameter (fac), Rhat (num), scenario (fac), group (fac), xmax (num)}

\item{ncol}{Number of columns for facet_wrap(). Default is 3.}

\item{parameter_labels}{y-axis parameter labels, e.g. expression(xi)}

\item{ylab}{Label for y-axis (character). Default is "Fishing mortality".}

\item{show_point}{Set to TRUE to change to point type plot. Default is FALSE, displaying bar type plot}

\item{scenarios}{A vector of scenario numbers to be shown on plot (numeric). This was already specified in prep file, but this is a manual override to save running the prep function again.}

\item{scenario_labels}{A vector of customised scenario names (character). Default is "Scenario 1", "Scenario 2", etc.}

\item{scenario_order}{A vector to reorder how scenarios are displayed (character). Use the label names defined in "scenario_labels".
If "scenario_labels" is left blank, the labels will be "Scenario 1", "Scenario 2" etc.
Any scenarios not included in "scenario_order" will be tacked on in the order they appear in the input data.}

\item{text_size, legend_text_size, text_colour, legend_text_colour, legend_position, legend_box, legend_title_blank, panel_border, panel_border_colour, xangle}{Optional plotting theme overrides. Defaults are controlled by \code{theme_ssand()}
and can be set globally via \code{set_ssand_style()}.}
}
\value{
A plot of R-hat values for parameter estimates
}
\description{
Plot of Rhat values for each parameter
}
\examples{
data <- mcmc_rhatplot_prep_SS(ss_mle, ss_mcmc,
           parameters = extract_SS_parameters(ss_mle)[c(2:10),])
mcmc_rhatplot(data)
}
