% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_rhatplot_prep_DD.R
\name{mcmc_rhatplot_prep_DD}
\alias{mcmc_rhatplot_prep_DD}
\title{Prepare data fro mcmc_rhatplot}
\usage{
mcmc_rhatplot_prep_DD(dd_mle, dd_mcmc, scenarios = NULL)
}
\arguments{
\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{dd_mcmc}{A list of model fits from tmbstan::tmbstan() with one list element per scenario. Only needed if MCMC was used.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}
}
\value{
A dataframe with variables parameter (numeric), group (factor), scenario (factor)
}
\description{
Prepare data fro mcmc_rhatplot
}
\examples{
\dontrun{
library(DDUST)
data <- mcmc_rhatplot_prep_DD(dd_mle,dd_mcmc)
mcmc_rhatplot(data)
}
}
