% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_rhatplot_prep_SS.R
\name{mcmc_rhatplot_prep_SS}
\alias{mcmc_rhatplot_prep_SS}
\title{Prepare Stock Synthesis data for mcmc_rhatplot}
\usage{
mcmc_rhatplot_prep_SS(ss_mle, ss_mcmc, scenarios = NULL, parameters = NULL)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{A list of outputs from r4ss::SSgetMCMC() with one element per scenario. Only needed if MCMC was used. Will automatically reformat as a list if a single r4ss::SSgetMCMC() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{parameters}{A vector of parameters to include on plot (character). Use SSAND::extract_SS_parameters() to find a list of parameter names used.}
}
\value{
A dataframe with variables  parameter (factor), rhat (num), scenario (factor), group (fct), xmax (num)
}
\description{
Prepare Stock Synthesis data for mcmc_rhatplot
}
\examples{
data <- mcmc_rhatplot_prep_SS(ss_mle, ss_mcmc,
           parameters = extract_SS_parameters(ss_mle)[c(2:10),])
mcmc_rhatplot(data)
}
