% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_fishers.R
\name{nfishersplot}
\alias{nfishersplot}
\title{Number of fishers plot}
\usage{
nfishersplot(
  data,
  species_of_interest = NULL,
  interesting_years = NULL,
  interesting_fishers = NULL,
  show_region = FALSE,
  show_latband = FALSE,
  show_latband_regions = FALSE,
  remove_NA_latband = FALSE,
  show_total = FALSE,
  extract_data = FALSE
)
}
\arguments{
\item{data}{Output of format_logbooks()}

\item{species_of_interest}{The name of the species of interest, as listed in the 'species' column of data.}

\item{interesting_years}{Optional. A vector of years to filter the dataset; only fishers operating in this years will be included.}

\item{interesting_fishers}{Optional. A vector of fishers/operators to filter the dataset by.}

\item{show_region}{Set to TRUE to display fishers by region}

\item{show_latband}{Set to TRUE to display fishers by latitude band}

\item{show_latband_regions}{Set to TRUE to display fishers by latitude band grouping}

\item{remove_NA_latband}{Set to TRUE to remove data without valid latitude band.}

\item{show_total}{Set to TRUE to show total number of fishers in fishery on plot.}

\item{extract_data}{Set to TRUE to return data instead of plot. Default is FALSE.}
}
\value{
A plot of the number of fishers in the fishery
}
\description{
Number of fishers plot
}
\examples{
\dontrun{nfishersplot(data = format_logbooks(raw_data))}
}
