% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametertable.R
\name{parametertable}
\alias{parametertable}
\title{Parameter table}
\usage{
parametertable(
  data,
  label,
  parameter_names = NULL,
  column_names = NULL,
  caption = "Summary of parameter estimates",
  align = NULL,
  round = NULL
)
}
\arguments{
\item{data}{Dataframe output from parametertable_prep()}

\item{label}{A label for the table that can be referenced elsewhere in the report using LaTeX syntax}

\item{parameter_names}{An optional vector of customised parameter labels}

\item{column_names}{An optional vector of column names}

\item{caption}{Caption for table}

\item{align}{An optional vector of alignment for the table, input as a character string that is one character longer than the number of columns in the table. Default is centered.}

\item{round}{A vector of digits, one value per row of the table, used to round the parameter estimates}
}
\value{
Code that can be used within a code chunk in a RMarkdown style document to produce a parameter estimate table
}
\description{
Parameter table
}
\examples{
# \rowcolors{2}{white}{light-gray}
# <<table_mcmc1, results='asis', echo=FALSE>>=
parameters <- extract_SS_parameters(ss_mle)[2:4,]
data <- parametertable_prep_SS(ss_mle=ss_mle, parameters=parameters, scenario=1)
parametertable(data, label="tab:param")
# @
}
