% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametertable_prep_DD.R
\name{parametertable_prep_DD}
\alias{parametertable_prep_DD}
\title{Prepare DDUST data for parameter summary table}
\usage{
parametertable_prep_DD(
  dd_mle = NULL,
  dd_mcmc = NULL,
  scenario = 1,
  parameters = NULL
)
}
\arguments{
\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{dd_mcmc}{A list of model fits from tmbstan::tmbstan() with one list element per scenario. Only needed if MCMC was used.}

\item{scenario}{The scenario to display (numeric). Shows scenario 1 if left blank. Can be overridden in the plotting function.}

\item{parameters}{A vector of parameters to include on plot (character). See names(dd_mle) for options.}
}
\value{
A data frame that summarises parameter estimate, ready for use in parametertable()
}
\description{
Prepare DDUST data for parameter summary table
}
\examples{
# \rowcolors{2}{white}{light-gray}
# <<table_mcmc1, results='asis', echo=FALSE>>=
data <- parametertable_prep_DD(dd_mle=dd_mle)
parametertable(data, label="tab:param")
# @

\dontrun{
library(DDUST)
dd_sim <- simulate_DDUST(dd_mle,dd_mcmc)
# \rowcolors{2}{white}{light-gray}
# <<table_mcmc1, results='asis', echo=FALSE>>=
data <- parametertable_prep_DD(dd_mcmc=dd_mcmc)
parametertable(data, label="tab:param")
# @

# \rowcolors{2}{white}{light-gray}
# <<table_mcmc1, results='asis', echo=FALSE>>=
dd_mcmc_ens <- mcmc_ensemble_DD(dd_mcmc,dd_sim,scenarios=c(1,2))$dd_mcmc
data <- parametertable_prep_DD(dd_mcmc=dd_mcmc_ens)
parametertable(data, label="tab:param")
# @
}
}
