% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametertable_prep_SS.R
\name{parametertable_prep_SS}
\alias{parametertable_prep_SS}
\title{Prepare Stock Synthesis data for parameter summary table}
\usage{
parametertable_prep_SS(
  ss_mle = NULL,
  ss_mcmc = NULL,
  scenario = 1,
  parameters,
  fields = c("Label", "Value", "Phase", "Min", "Max", "Init", "Parm_StDev")
)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{A list of outputs from r4ss::SSgetMCMC() with one element per scenario. Only needed if MCMC was used. Will automatically reformat as a list if a single r4ss::SSgetMCMC() output (i.e. one scenario) is entered.}

\item{scenario}{The scenario to display (numeric). Shows scenario 1 if left blank. Can be overridden in the plotting function.}

\item{parameters}{A vector of parameters to include in tablet (character). Use SSAND::extract_SS_parameters() to find a list of parameter names used.}

\item{fields}{For MLE, a vector of column headers to include in the table (e.g. Value, Phase, Min, Max etc)}
}
\value{
A data frame that summarised parameter estimate, ready for use in parametertable()
}
\description{
Prepare Stock Synthesis data for parameter summary table
}
\examples{
# \rowcolors{2}{white}{light-gray}
# <<table_mcmc1, results='asis', echo=FALSE>>=
parameters <- extract_SS_parameters(ss_mle)[2:4,]
data <- parametertable_prep_SS(ss_mle=ss_mle, parameters=parameters, scenario=1)
parametertable(data, label="tab:param")
# @

# \rowcolors{2}{white}{light-gray}
# <<table_mcmc1, results='asis', echo=FALSE>>=
parameters <- extract_SS_parameters(ss_mle)[2:4,]
data <- parametertable_prep_SS(ss_mcmc=ss_mcmc, parameters=parameters, scenario=1)
parametertable(data, label="tab:param")
# @

# \rowcolors{2}{white}{light-gray}
# <<table_mcmc1, results='asis', echo=FALSE>>=
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc,scenarios=c(1,2))
parameters <- extract_SS_parameters(ss_mle)[2:4,]
data <- parametertable_prep_SS(ss_mcmc=ss_mcmc_ens, parameters=parameters, scenario=1)
parametertable(data, label="tab:param")
# @
}
