% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseplot_prep_DD.R
\name{phaseplot_prep_DD}
\alias{phaseplot_prep_DD}
\title{Prepare DDUST data for phase plot}
\usage{
phaseplot_prep_DD(dd_mle, scenarios = NULL)
}
\arguments{
\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}
}
\value{
A data frame that contains year (int), Bratio (num), F_ (num), scenario (factor), B_MSY (num), F_max (num)
}
\description{
Prepare DDUST data for phase plot
}
\examples{
data <- phaseplot_prep_DD(dd_mle)
phaseplot(data)
}
