% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseplot_prep_SS.R
\name{phaseplot_prep_SS}
\alias{phaseplot_prep_SS}
\title{Prepare Stock Synthesis data for phase plot}
\usage{
phaseplot_prep_SS(ss_mle, scenarios = NULL, financial_year = FALSE)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{financial_year}{Set to TRUE if the assessment was based on financial year (logical).}
}
\value{
A data frame that contains year (int), Bratio (num), F_ (num), scenario (factor), B_MSY (num), F_max (num)
}
\description{
Prepare Stock Synthesis data for phase plot
}
\examples{
data <- phaseplot_prep_SS(ss_mle)
phaseplot(data)
}
