% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinerplot.R
\name{pinerplot}
\alias{pinerplot}
\title{Piner plot
A fleet-based likelihood plot}
\usage{
pinerplot(
  data,
  xlab = NULL,
  ylab = "Change in -log-likelihood",
  xlim = NULL,
  ylim = NULL,
  colours = SSAND::fq_palette("alisecolours"),
  shapes = c(16, 17, 18, 15, 1, 2, 5, 0, 19, 20, 3, 4, 7, 8, 9, 10, 11, 12, 13, 14),
  fleet_names = NULL,
  xangle = NULL,
  legend_position = NULL,
  text_size = NULL,
  legend_text_size = NULL,
  text_colour = NULL,
  legend_text_colour = NULL,
  legend_box = NULL,
  legend_title_blank = NULL,
  panel_border = NULL,
  panel_border_colour = NULL
)
}
\arguments{
\item{data}{Output from pinerplot_prep(). A dataframe with columns x_vector, fleet and likelihood}

\item{xlab}{Label for x-axis (character). # Default is \code{expression(log(italic(R)[0]))}.}

\item{ylab}{Label for y-axis (character). Default is "Change in -log-likelihood".}

\item{xlim}{A vector of lower and upper x-axis limits (e.g. c(1950, 2020)) (numeric).}

\item{ylim}{A vector of lower and upper y-axis limits (e.g. c(0,1)) (numeric).}

\item{colours}{A vector of colours used (character).}

\item{shapes}{A vector of shapes used (character).}

\item{fleet_names}{A vector of customised fleet names for legend (character).}

\item{text_size, legend_text_size, text_colour, legend_text_colour, legend_position, legend_box, legend_title_blank, panel_border, panel_border_colour, xangle}{Optional plotting theme overrides. Defaults are controlled by \code{theme_ssand()}
and can be set globally via \code{set_ssand_style()}.}
}
\value{
Piner plot, a fleet based likelihood plot
}
\description{
Piner plot
A fleet-based likelihood plot
}
\examples{
\dontrun{
r4ss::profile(dir = '.', # directory of 4 SS files
              oldctlfile = "control.ctl",
              newctlfile = "control.ctl",
              string = "steep",
              profilevec = c(0.4,0.5,0.6),
              exe = "C:/stocksynthesis/ss_3.30.22.exe")

profile_input <- r4ss::SSsummarize(
  r4ss::SSgetoutput(dirvec = ".",
                    keyvec = 1:3, # 1:length(profilevec)
                    getcovar = FALSE,
                    getcomp = FALSE))

pinerplot_component_options(profile_input)
data <- pinerplot_prep_SS(profile_input, component="Length_like")
pinerplot(data)
}
}
