% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinerplot_component_options.R
\name{pinerplot_component_options}
\alias{pinerplot_component_options}
\title{Options for component argument of pinerplot_prep_SS}
\usage{
pinerplot_component_options(profile_input)
}
\arguments{
\item{profile_input}{The likelihood profile object after running: r4ss::SS_profile(), r4ss::SSgetoutput(), then r4ss::SSsummarize(). See ?pinerplot for example.}
}
\value{
A vector of options that can be used in the component argument of pinerplot_prep_SS()
}
\description{
Options for component argument of pinerplot_prep_SS
}
\examples{
\dontrun{
r4ss::profile(dir = '.', # directory of 4 SS files
              oldctlfile = "control.ctl",
              newctlfile = "control.ctl",
              string = "steep",
              profilevec = c(0.4,0.5,0.6),
              exe = "C:/stocksynthesis/ss_3.30.22.exe")

profile_input <- r4ss::SSsummarize(
  r4ss::SSgetoutput(dirvec = ".",
                    keyvec = 1:3, # 1:length(profilevec)
                    getcovar = FALSE,
                    getcomp = FALSE))

pinerplot_component_options(profile_input)
data <- pinerplot_prep_SS(profile_input, component="Length_like")
pinerplot(data)
}
}
