% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportionspawningplot_prep_SS.R
\name{proportionspawningplot_prep_SS}
\alias{proportionspawningplot_prep_SS}
\title{Prepare Stock Synthesis data for proportion spawning plot}
\usage{
proportionspawningplot_prep_SS(ss_mle)
}
\arguments{
\item{ss_mle}{list of outputs from r4ss::SS_output(), one element per scenario}
}
\value{
A warning that says "No functionality built for Stock Synthesis proportion spawning pattern plot yet"
}
\description{
Prepare Stock Synthesis data for proportion spawning plot
}
