% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_fishers.R
\name{rank_fishers}
\alias{rank_fishers}
\title{Rank fishers}
\usage{
rank_fishers(data, species_of_interest = NULL, show_year = TRUE, round = TRUE)
}
\arguments{
\item{data}{Output of format_logbooks()}

\item{species_of_interest}{The name of the species of interest, as listed in the 'species' column of data.}

\item{show_year}{Set to TRUE to show year in table. Set to FALSE to show total number of years operator has been in fishery.}

\item{round}{Set to TRUE to round total catch.}
}
\value{
A table of fishers ranked by their total catch
}
\description{
Rank fishers
}
\examples{
\dontrun{rank_fishers(data = format_logbooks(raw_data))}
}
