% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_biologyplots.R
\name{rawbiological_boxplot}
\alias{rawbiological_boxplot}
\title{Box plot for raw biological data}
\usage{
rawbiological_boxplot(
  data,
  xlab = NULL,
  ylab = NULL,
  x_var = "year",
  y_var = "fl",
  facet_var = NULL,
  MLS_x = NULL,
  MLS_y = NULL,
  scales = "fixed",
  ncol = 3,
  extract_data = FALSE
)
}
\arguments{
\item{data}{A table of raw length and age data. Each row is a different sample. Columns are labelled according to your set up in the "_var" arguments of this function.}

\item{xlab}{Label for x-axis (character)}

\item{ylab}{Label for y-axis (character)}

\item{x_var}{Variable to be plotted on x-axis (e.g. "year")}

\item{y_var}{Variable to be plotted on x-axis (e.g. "fl")}

\item{facet_var}{Variable by which plot is to be faceted (e.g. "region")}

\item{MLS_x}{Optional. The value of the minimum legal size, if a horizontal line is to be added.}

\item{MLS_y}{Optional. The value of the minimum legal size, if a vertical line is to be added.}

\item{scales}{Scaled for facet wrap. Default is 'fixed'.}

\item{ncol}{Number of columns for facet wrap. Default is 2.}

\item{extract_data}{Set to TRUE to return data instead of plot. Default is FALSE.}
}
\value{
Box plot for raw biological data
}
\description{
Box plot for raw biological data
}
\examples{
\dontrun{rawbiological_boxplot(data, x_var = 'year', xlab='Year',
                                y_var = 'fl', ylab = 'Fork length (cm)')}
\dontrun{rawbiological_boxplot(data, x_var = 'year', xlab='Year',
                    y_var = 'fl', ylab = 'Fork length (cm)',
                    facet_var='region')}
}
