% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_biologyplots.R
\name{rawbiological_scatterplot}
\alias{rawbiological_scatterplot}
\title{Scatter plot for raw biological data}
\usage{
rawbiological_scatterplot(
  data,
  xlab = NULL,
  ylab = NULL,
  x_var = "agegrp",
  y_var = "fl",
  facet_var = NULL,
  MLS_x = NULL,
  MLS_y = NULL,
  scales = "fixed",
  ncol = 1,
  show_vonBert = FALSE,
  show_vonBert_only = FALSE,
  legend_title = "none",
  legend_position = "none",
  extract_data = FALSE
)
}
\arguments{
\item{data}{A table of raw length and age data. Each row is a different sample. Columns are labelled according to your set up in the "_var" arguments of this function.}

\item{xlab}{Label for x-axis (character)}

\item{ylab}{Label for y-axis (character)}

\item{x_var}{Variable to be plotted on x-axis (e.g. "year")}

\item{y_var}{Variable to be plotted on x-axis (e.g. "fl")}

\item{facet_var}{Variable by which plot is to be faceted (e.g. "region")}

\item{MLS_x}{Optional. The value of the minimum legal size, if a horizontal line is to be added.}

\item{MLS_y}{Optional. The value of the minimum legal size, if a vertical line is to be added.}

\item{scales}{Scaled for facet wrap. Default is 'fixed'.}

\item{ncol}{Number of columns for facet wrap. Default is 2.}

\item{show_vonBert}{Set to TRUE to show von Bertalanffy fitted curves}

\item{show_vonBert_only}{Set to TRUE to show only von Bertalanffy fitted curves. This overrides the facetting function.}

\item{legend_title}{Legend title}

\item{legend_position}{Legend position}

\item{extract_data}{Set to TRUE to return data instead of plot. Default is FALSE.}
}
\value{
Scatter plot for raw biological data
}
\description{
Scatter plot for raw biological data
}
\examples{
\dontrun{rawbiological_scatterplot(mac3, x_var = 'agegrp',
xlab = 'Age group (years)', y_var = 'fl', ylab = 'Fork length (cm)',
facet_var='region', show_vonBert=TRUE, show_vonBert_only=TRUE, legend_position='top')}
}
