% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_rawcatcheffortplot.R
\name{rawcatcheffortplot}
\alias{rawcatcheffortplot}
\title{Raw catch or effort plot}
\usage{
rawcatcheffortplot(
  data,
  y_var = "catch",
  species_of_interest = NULL,
  interesting_years = NULL,
  interesting_fishers = NULL,
  show_proportion = FALSE,
  proportion_only = FALSE,
  threshold = NULL,
  annual = TRUE,
  boxplot = FALSE,
  fill_var = "region",
  facet_var = NULL,
  max_days = 1,
  xlab = "Year",
  ylab = NULL,
  facet_labels = NULL,
  scales = "fixed",
  ncol = 2,
  dodge = FALSE,
  colours = rep(fq_palette("alisecolours"), 10),
  legend_position = "top",
  legend_title = "",
  include_method = NULL,
  n_method = 15,
  include_species = NULL,
  n_species = 15,
  extract_data = FALSE
)
}
\arguments{
\item{data}{Output of format_logbooks()}

\item{y_var}{Variable on the y axis. Either "catch" or "effort"}

\item{species_of_interest}{The name of the species of interest, as listed in the 'species' column of data.}

\item{interesting_years}{Optional. A vector of years to filter the dataset; only fishers operating in this years will be included.}

\item{interesting_fishers}{Optional. A vector of fishers/operators to filter the dataset by.}

\item{show_proportion}{Set to TRUE to show relative proportions}

\item{proportion_only}{Set to TRUE to show only relative proportions}

\item{threshold}{The percentile threshold used to 'zoom' in on the right hand side plot. Only for boxplots.}

\item{annual}{Set to TRUE to aggregate by year. Set to FALSE to show overall catch}

\item{boxplot}{Set to TRUE to display as a boxplot. Set to FALSE to display as a barplot.}

\item{fill_var}{Variable by which plot is to be filled (e.g. "region")}

\item{facet_var}{Variable by which plot is to be faceted (e.g. "region")}

\item{max_days}{Numeric. Filter to include only records that were less than or equal to this maximum number of fishing days (i.e. exclude multi day trips)}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{ylab}{Label for y-axis (character). Default is "Daily catch (kg)".}

\item{facet_labels}{A vector of two strings, labels of each facet panel.}

\item{scales}{Scaled for facet wrap. Default is 'fixed'.}

\item{ncol}{Number of columns for facet wrap. Default is 2.}

\item{dodge}{Set to TRUE to offset bars, set to FALSE to stack them}

\item{colours}{Colours for plot}

\item{legend_position}{Legend position}

\item{legend_title}{Legend text}

\item{include_method}{(Optional) A vector of fishing methods to include (if filling by method)}

\item{n_method}{Maximum number of fishing methods to include (if filling by method)}

\item{include_species}{(Optional) A vector of species to include (if filling by species)}

\item{n_species}{Maximum number of species to include (if filling by species)}

\item{extract_data}{Set to TRUE to return data instead of plot. Default is FALSE.}
}
\value{
Raw catch or effort plot
}
\description{
Raw catch or effort plot
}
