% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitmentdistributiondeviationplot_prep_SS.R
\name{recruitmentdistributiondeviationplot_prep_SS}
\alias{recruitmentdistributiondeviationplot_prep_SS}
\title{Prepare Stock Synthesis data for Recruitment deviation distribution plot}
\usage{
recruitmentdistributiondeviationplot_prep_SS(
  ss_mle,
  ss_mcmc = NULL,
  scenarios = NULL,
  intervals = c(0.2, 0.5, 0.7, 0.9, 0.95),
  CI_range = 0.95
)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{A list of outputs from r4ss::SSgetMCMC() with one element per scenario. Only needed if MCMC was used. Will automatically reformat as a list if a single r4ss::SSgetMCMC() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{intervals}{A vector of credible interval values to be displayed on banded MCMC plot (numeric). For example, "0.9" denotes the 90\% credible interval.}

\item{CI_range}{Specify credible interval range (numeric). Only activated if mcmc_style==CI.}
}
\value{
A data frame with variables called year (int), value (int), ub (num), lb (num), median (num, if MCMC), scenario (factor), method (chr).
}
\description{
Prepare Stock Synthesis data for Recruitment deviation distribution plot
}
