% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitmentpatternplot_prep_SS.R
\name{recruitmentpatternplot_prep_SS}
\alias{recruitmentpatternplot_prep_SS}
\title{Prepare Stock Synthesis data for recruitment pattern plot}
\usage{
recruitmentpatternplot_prep_SS(ss_mle, scenarios = NULL)
}
\arguments{
\item{ss_mle}{list of outputs from r4ss::SS_output(), one element per scenario}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}
}
\value{
A dataframe with recruitment (num), months (num), monthnames (factor), scenario (factor), format (character)
}
\description{
Prepare Stock Synthesis data for recruitment pattern plot
}
\examples{
data <- recruitmentpatternplot_prep_SS(ss_mle)
recruitmentpatternplot(data)
}
