% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_values_DD.R
\name{report_values_DD}
\alias{report_values_DD}
\title{Calculate report values from DDUST output}
\usage{
report_values_DD(
  dd_mle,
  dd_mcmc = NULL,
  dd_sim = NULL,
  base_case = 1,
  catch = NULL,
  number_catch_years = 5,
  prefix = NULL,
  suffix = NULL,
  end_year = NULL,
  credible_interval = 0.95,
  number_F_years = 5,
  biomass_minimum = "any",
  round_F_summary = 2,
  round_FMSY = 3,
  biomass_type = "relative",
  F_relative_to_MSY = TRUE
)
}
\arguments{
\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{dd_mcmc}{Output from mcmc_ensemble_DD()$dd_mcmc. A list that combines all outputs from tmbstan::tmbstan(). Only needed if MCMC was used.}

\item{dd_sim}{Output from mcmc_ensemble_DD()$dd_sim. A list that combines all outputs from DDUST::simulate_DD(). Only required if MCMC was used.}

\item{base_case}{The representative base case or reference scenario from ss_mle. Defaults to 1.}

\item{catch}{A data frame of catch, with columns year (num), value (num) and sector (chr)}

\item{number_catch_years}{Number of recent years to average catch. Default is 5.}

\item{prefix}{Optional string to add to the beginning of variable names within data output. Useful if extracting values for multiple models.}

\item{suffix}{Optional string to add to the end of variable names within data output. Useful if extracting values for multiple models.}

\item{end_year}{Optional adjustment for biomass end year. For example, if catch is entered until 2022, end_year is 2023.}

\item{credible_interval}{Optional adjustment for credible interval. Default is 0.95.}

\item{number_F_years}{Number of recent years to average fishing pressure. Default is 5.}

\item{biomass_minimum}{Method for calculating minimum biomass for MCMC results.
"any" takes the minimum value from any MCMC iteration for any year
"median" takes the minimum value from the annual median biomass for any year}

\item{round_F_summary}{Number of decimal places to round F data for summary table}

\item{round_FMSY}{Number of decimal places to round F_MSY data}

\item{biomass_type}{The type of biomass used. Options are "relative" or "absolute" (character).}

\item{F_relative_to_MSY}{Set to TRUE to report F relative to FMSY}
}
\value{
A list containing named values that could be useful in reporting
}
\description{
Calculate report values from DDUST output
}
\examples{
catch <- rbind(data.frame(year = 1953:2022, sector = "Commercial",
                 value = c(602,674,561,472,436,480,370,404,390,277)),
               data.frame(year = 1953:2022, sector = "Recreational",
                 value = c(750,705,570,535,495,390,506,448,383,303)))
data <- report_values_DD(dd_mle,suffix="_MLE")
\dontrun{
library(DDUST)
dd_sim <- simulate_DDUST(dd_mle,dd_mcmc)
dd_mcmc_ens <- mcmc_ensemble_DD(dd_mcmc,dd_sim,scenarios=c(1,2))$dd_mcmc
dd_sim_ens <- mcmc_ensemble_DD(dd_mcmc,dd_sim,scenarios=c(1,2))$dd_sim

data <- report_values_DD(dd_mle,dd_mcmc_ens,dd_sim_ens,catch=catch)

data <- report_values_DD(dd_mle,dd_mcmc_ens,dd_sim_ens,
                             prefix="DD_",suffix="_1")
}
}
