% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_values_SS.R
\name{report_values_SS}
\alias{report_values_SS}
\title{Calculate report values from Stock Synthesis output}
\usage{
report_values_SS(
  ss_mle,
  ss_mcmc = NULL,
  base_case = 1,
  catch = NULL,
  number_catch_years = 5,
  prefix = NULL,
  suffix = NULL,
  end_year = NULL,
  credible_interval = 0.95,
  number_F_years = 5,
  biomass_minimum = "any",
  round_F_summary = 2,
  round_FMSY = 3,
  biomass_type = "relative",
  biomass_definition = "spawning",
  selectivity_fleet = 2
)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{Output from mcmc_ensemble_SS(). A list that combines all outputs from r4ss::SSgetMCMC(). Only needed if MCMC was used.}

\item{base_case}{The representative base case or reference scenario from ss_mle. Defaults to 1.}

\item{catch}{A data frame of catch, with columns year (num), value (num) and sector (chr)}

\item{number_catch_years}{Number of recent years to average catch. Default is 5.}

\item{prefix}{Optional string to add to the beginning of variable names within data output. Useful if extracting values for multiple models.}

\item{suffix}{Optional string to add to the end of variable names within data output. Useful if extracting values for multiple models.}

\item{end_year}{Optional adjustment for biomass end year. For example, if catch is entered until 2022, end_year is 2023.}

\item{credible_interval}{Optional adjustment for credible interval. Default is 0.95.}

\item{number_F_years}{Number of recent years to average fishing pressure. Default is 5.}

\item{biomass_minimum}{Method for calculating minimum biomass for MCMC results.
"any" takes the minimum value from any MCMC iteration for any year
"median" takes the minimum value from the annual median biomass for any year}

\item{round_F_summary}{Number of decimal places to round F data for summary table}

\item{round_FMSY}{Number of decimal places to round F_MSY data}

\item{biomass_type}{The type of biomass used. Options are "relative" or "absolute" (character).}

\item{biomass_definition}{The definition of biomass used. Options are "spawning" or "vulnerable" (character).}

\item{selectivity_fleet}{Fleet number used for the selectivity applied to calculate vulnerable biomass.}
}
\value{
A list containing named values that could be useful in reporting
}
\description{
Calculate report values from Stock Synthesis output
}
\examples{
catch <- rbind(data.frame(year = 1953:2022, sector = "Commercial",
                 value = c(602,674,561,472,436,480,370,404,390,277)),
               data.frame(year = 1953:2022, sector = "Recreational",
                 value = c(750,705,570,535,495,390,506,448,383,303)))

ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc)

data <- report_values_SS(ss_mle,ss_mcmc_ens,catch=catch)
data <- report_values_SS(ss_mle,ss_mcmc_ens,prefix="SS_",suffix="_1")
data <- report_values_SS(ss_mle,suffix="_MLE")
}
