% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_boxplot.R
\name{residual_boxplot}
\alias{residual_boxplot}
\title{Fitted values vs residuals boxplot.}
\usage{
residual_boxplot(
  model,
  log = FALSE,
  nbreaks = 10,
  zero_line = TRUE,
  type = "deviance"
)
}
\arguments{
\item{model}{A fitted \code{gam} or \code{glm} object as produced by \code{gam()} or \code{glm()}.}

\item{log}{TRUE to log-transform fitted values.}

\item{nbreaks}{The number of breaks to group fitted values. Each group contains an equal number of observations.}

\item{zero_line}{TRUE to add dotted line at \code{y = 0}.}

\item{type}{the type of residuals which should be returned. The alternatives are: "deviance" (default), "pearson", "working", "response", and "partial".}
}
\value{
A boxplot made with ggplot2.
}
\description{
Identify heteroscedasticity arising from a fitted \code{glm()} or \code{gam()}.
}
\examples{
raw_catch_rates <- data.frame(
   year = seq(5,100,5),
   effort = c(118,58,42,35,27,25,21,19,18,10,12,15,13,22,8,9,7,16,17,12),
   kilograms = c(75,81,72,69,35,26,21,18,16,13,12,12,11,9,10,11,15,9,8,5))
model <- glm(kilograms ~ year + effort, data = raw_catch_rates, family = Gamma)
residual_boxplot(model, log = TRUE, nbreaks = 5)

}
