% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_seasonality.R
\name{seasonalityplot}
\alias{seasonalityplot}
\title{Seasonality plot}
\usage{
seasonalityplot(
  data,
  show_cpue = FALSE,
  show_catch = FALSE,
  show_effort = FALSE,
  species_of_interest = NULL,
  facet_var = NULL,
  fill_var = NULL,
  max_days = 1,
  daily = FALSE,
  xlab = "Year",
  ylab = NULL,
  colours = rep(fq_palette("alisecolours"), 10),
  ncol = 2,
  extract_data = FALSE
)
}
\arguments{
\item{data}{Output of format_logbooks()}

\item{show_cpue}{Set to TRUE to show CPUE on the y-axis}

\item{show_catch}{Set to TRUE to show catch on the y-axis}

\item{show_effort}{Set to TRUE to show effort on the y-axis}

\item{species_of_interest}{Optional. The name of the species of interest, as listed in the 'species' column of data. Filters data to only include that species.}

\item{facet_var}{Variable by which plot is to be faceted (e.g. "region")}

\item{fill_var}{Variable by which plot is to be filled (e.g. "region")}

\item{max_days}{Numeric. Filter to include only records that were less than or equal to this maximum number of fishing days (i.e. exclude multi-day trips)}

\item{daily}{Set to TRUE to show daily seasonality. Set to FALSE for monthly.}

\item{xlab}{Label for x-axis of plot (character). Default is "".}

\item{ylab}{Label for y-axis of plot (character). Default is "".}

\item{colours}{A vector of colours used (character).}

\item{ncol}{Number of columns for facet wrap. Default is 1.}

\item{extract_data}{Set to TRUE to return data instead of plot. Default is FALSE.}
}
\value{
Seasonality plot
}
\description{
Seasonality plot
}
\examples{
\dontrun{seasonalityplot(data = format_logbooks(raw_data))}
}
