% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivityplot_prep_SS.R
\name{selectivityplot_prep_SS}
\alias{selectivityplot_prep_SS}
\title{Prepare Stock  Synthesis data for selectivity plot}
\usage{
selectivityplot_prep_SS(ss_mle, ss_mcmc_ens = NULL, scenarios, fleet = NULL)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc_ens}{Output of mcmc_ensemble_SS(). A list containing one dataframe of all specified ss_mcmc objects combined, limited to columns that are common between ss_mcmc objects, duplicated as per the specified weighting.
An extra column "ensemble" is appended, for data checks in other functions.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{fleet}{A numeric representing the fleet to be included in MCMC data prep.}
}
\value{
A data frame with variables fleet (factor), year (int), sex (int), value (num), type (factor), selectivity (num)
}
\description{
Prepare Stock  Synthesis data for selectivity plot
}
\examples{
data <- selectivityplot_prep_SS(ss_mle)
selectivityplot(data)

# Ensemble model
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc)
data <- selectivityplot_prep_SS(ss_mle,ss_mcmc_ens,fleet = 1)
selectivityplot(data)
}
