% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivityplot_prep_SS.R
\name{sensitivityplot_prep_SS}
\alias{sensitivityplot_prep_SS}
\title{Prepare Stock Synthesis data for sensitivityplot()}
\usage{
sensitivityplot_prep_SS(
  ss_mle,
  ss_mcmc = NULL,
  scenarios = NULL,
  parameters = NULL,
  show_B_ratio = TRUE,
  show_B_abs = FALSE,
  show_F_final = FALSE,
  show_MSY = FALSE,
  show_B_MSY = FALSE,
  show_F_MSY = FALSE,
  show_LL = TRUE
)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{A list of outputs from r4ss::SSgetMCMC() with one element per scenario. Only needed if MCMC was used. Will automatically reformat as a list if a single r4ss::SSgetMCMC() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{parameters}{A vector of parameters to include on plot (character). Use SSAND::extract_SS_parameters() to find a list of parameter names used.}

\item{show_B_ratio}{Set to TRUE to show a panel for final biomass ratio (logical).}

\item{show_B_abs}{Set to TRUE to show a panel for final absolute biomass (logical).}

\item{show_F_final}{Set to TRUE to show final annual fishing mortality (logical).}

\item{show_MSY}{Set to TRUE, to show MSY (logical).}

\item{show_B_MSY}{Set to TRUE to show B_MSY (logical).}

\item{show_F_MSY}{Set to TRUE to show F_MSY (logical).}

\item{show_LL}{Set to TRUE to show negative log likelhood or objective function (logical).}
}
\value{
A data frame with scenario (int), name (fac), value (num), ub (num), lb (num) and fixed (boolean).
}
\description{
It is recommended you use SSAND::extract_SS_parameters() to find the parameter names you'd like to include on the plot.
}
\examples{
parm <- extract_SS_parameters(ss_mle)[c(3:6),]
data <- sensitivityplot_prep_SS(ss_mle,
                                ss_mcmc,
                                show_MSY = TRUE,
                                show_LL = TRUE,
                                show_B_ratio = TRUE,
                                parameters = parm)
sensitivityplot(data)
}
