% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghettiplot_prep_DD.R
\name{spaghettiplot_prep_DD}
\alias{spaghettiplot_prep_DD}
\title{Prepare DDUST data for biomassplot()}
\usage{
spaghettiplot_prep_DD(
  dd_mle,
  dd_mcmc = NULL,
  dd_sim = NULL,
  base_case = 1,
  biomass_type = "relative",
  scenarios = NULL,
  show_median = "trajectory"
)
}
\arguments{
\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{dd_mcmc}{A list of model fits from tmbstan::tmbstan() with one list element per scenario. Only needed if MCMC was used.}

\item{dd_sim}{A list of outputs from SSAND::simulate_DD() with one list element per scenario. Only required if MCMC was used. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{base_case}{The scenario number that pertains to the base case (numeric). If NA, no base case was selected.}

\item{biomass_type}{The type of biomass used. Options are "relative" or "absolute" (character).}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{show_median}{Type of median shown. Default "trajectory" shows median trajectory based on biomass in final year.
"annual_biomass" shows the median of each year}
}
\value{
A data frame with variables called year, scenario, base and value
}
\description{
Prepare DDUST data for biomassplot()
}
\examples{
data <- spaghettiplot_prep_DD(dd_mle)
spaghettiplot(data)

\dontrun{
library(DDUST)
dd_sim <- simulate_DDUST(dd_mle,dd_mcmc)
data <- spaghettiplot_prep_DD(dd_mle, dd_mcmc, dd_sim)
spaghettiplot(data)
}
}
