% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghettiplot_prep_SS.R
\name{spaghettiplot_prep_SS}
\alias{spaghettiplot_prep_SS}
\title{Prepare Stock Synthesis data for biomassplot()}
\usage{
spaghettiplot_prep_SS(
  ss_mle,
  ss_mcmc = NULL,
  base_case = 1,
  biomass_type = "relative",
  scenarios = NULL,
  period = NULL,
  show_median = "trajectory"
)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{A list of outputs from r4ss::SSgetMCMC() with one element per scenario. Only needed if MCMC was used. Will automatically reformat as a list if a single r4ss::SSgetMCMC() output (i.e. one scenario) is entered.}

\item{base_case}{The scenario number that pertains to the base case (numeric). If NA, no base case was selected.}

\item{biomass_type}{The type of biomass used. Options are "relative" or "absolute" (character).}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{period}{Optional: time period of stock assessment.}

\item{show_median}{Type of median shown. Default "trajectory" shows median trajectory based on biomass in final year.
"annual_biomass" shows the median of each year}
}
\value{
A data frame with variables called year, scenario, base and value
}
\description{
Prepare Stock Synthesis data for biomassplot()
}
\examples{
data <- spaghettiplot_prep_SS(ss_mle)
spaghettiplot(data)

data <- spaghettiplot_prep_SS(ss_mle, ss_mcmc)
spaghettiplot(data)
}
