% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawningoutputplot_prep_SS.R
\name{spawningoutputplot_prep_SS}
\alias{spawningoutputplot_prep_SS}
\title{Prepare Stock Synthesis data for spawningoutputplot()}
\usage{
spawningoutputplot_prep_SS(ss_mle, ss_mcmc = NULL, scenarios = NULL)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{A list of outputs from r4ss::SSgetMCMC() with one element per scenario. Only needed if MCMC was used. Will automatically reformat as a list if a single r4ss::SSgetMCMC() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}
}
\value{
A dataframe with xvar (int), maturityxfecundity (num), x (chr), scenario (int)
}
\description{
Prepare Stock Synthesis data for spawningoutputplot()
}
\examples{
data <- spawningoutputplot_prep_SS(ss_mle)
spawningoutputplot(data,xaxis="length")
}
