% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srplot.R
\name{srplot}
\alias{srplot}
\title{Stock-recruit curve plot}
\usage{
srplot(
  data,
  xlab = expression(paste("Spawning output (relative to ", italic(B)[0], ")")),
  ylab = expression(paste("Recruitment (relative to ", italic(R)[0], ")")),
  xbreaks = NULL,
  show_bias_adjust = FALSE,
  scenarios = NULL,
  scenario_labels = NULL,
  scenario_order = NULL,
  scales = "free",
  ncol = 2,
  xangle = NULL,
  legend_position = NULL,
  text_size = NULL,
  legend_text_size = NULL,
  text_colour = NULL,
  legend_text_colour = NULL,
  legend_box = NULL,
  legend_title_blank = NULL,
  panel_border = NULL,
  panel_border_colour = NULL
)
}
\arguments{
\item{data}{Output from srplot_prep() with columns year (int), spawn_bio (num), pred_recr (num), exp_recr (num), dev (num), bias_adjusted (num), scenario (int)}

\item{xlab}{Label for x-axis (character). Default is "Spawning output (relative to B0)".}

\item{ylab}{Label for y-axis (character). Default is "Recruitment (relative to R0)".}

\item{xbreaks}{A vector of breaks between x-axis labels, used in ggplot2::scale_x_continous() (numeric).}

\item{show_bias_adjust}{Set to TRUE to include bias adjusted recruitment (set to FALSE for DDUST or REDDUST) (logical).}

\item{scenarios}{A vector of scenario numbers to be shown on plot (numeric). This was already specified in prep file, but this is a manual override to save running the prep function again.}

\item{scenario_labels}{A vector of customised scenario names (character). Default is "Scenario 1", "Scenario 2", etc.}

\item{scenario_order}{A vector to reorder how scenarios are displayed (character). Use the label names defined in "scenario_labels".
If "scenario_labels" is left blank, the labels will be "Scenario 1", "Scenario 2" etc.
Any scenarios not included in "scenario_order" will be tacked on in the order they appear in the input data.}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'free', see ?ggplot2::facet_wrap for options.}

\item{ncol}{Number of columns for facet_wrap(). Default is 2.}

\item{text_size, legend_text_size, text_colour, legend_text_colour, legend_position, legend_box, legend_title_blank, panel_border, panel_border_colour, xangle}{Optional plotting theme overrides. Defaults are controlled by \code{theme_ssand()}
and can be set globally via \code{set_ssand_style()}.}
}
\value{
Plot of S-R curve
}
\description{
Stock-recruit curve plot
}
\examples{
data <- srplot_prep_SS(ss_mle)
srplot(data)

data <- srplot_prep_DD(dd_mle)
srplot(data)
}
