% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srplot_prep_DD.R
\name{srplot_prep_DD}
\alias{srplot_prep_DD}
\title{Prepare DDUST data for stock-recruit curve plot}
\usage{
srplot_prep_DD(dd_mle, dd_mcmc = NULL, scenarios = NULL)
}
\arguments{
\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{dd_mcmc}{A list of model fits from tmbstan::tmbstan() with one list element per scenario. Only needed if MCMC was used.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}
}
\value{
A data frame with variables year (int), spawn_bio (num), pred_rec (num), dev (num), exp_recr (num), bias_adjusted (num), scenario (factor)
}
\description{
Prepare DDUST data for stock-recruit curve plot
}
\examples{
data <- srplot_prep_DD(dd_mle)
srplot(data)
}
