% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srplot_prep_SS.R
\name{srplot_prep_SS}
\alias{srplot_prep_SS}
\title{Prepare Stock Synthesis data for stock-recruit curve plot}
\usage{
srplot_prep_SS(ss_mle, ss_mcmc = NULL, scenarios = NULL)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{A list of outputs from r4ss::SSgetMCMC() with one element per scenario. Only needed if MCMC was used. Will automatically reformat as a list if a single r4ss::SSgetMCMC() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}
}
\value{
A data frame with variables year (int), spawn_bio (num), pred_rec (num), dev (num), exp_recr (num), bias_adjusted (num), scenario (factor)
}
\description{
Prepare Stock Synthesis data for stock-recruit curve plot
}
\examples{
data <- srplot_prep_SS(ss_mle)
srplot(data)
}
