% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_mcmc.R
\docType{data}
\name{ss_mcmc}
\alias{ss_mcmc}
\title{Fake data to be used in ..._SS prep functions.}
\format{
\subsection{\code{ss_mcmc}}{

A list of 6 Stock Synthesis MCMC output objects.
Each element of the list is a data frame where rows are saved iterations of the MCMC and
columns are parameter estimates and derived quantities.
}
}
\usage{
ss_mcmc
}
\description{
Fabricated data showing 6 scenarios of Stock Synthesis models run through
Stock Synthesis (SS_3.30.20) (maximum likelihood estimation) and r4ss::SS_output().
}
\keyword{datasets}
