% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_mle.R
\docType{data}
\name{ss_mle}
\alias{ss_mle}
\title{Fake data to be used in ..._SS prep functions.}
\format{
\subsection{\code{ss_mle}}{

A list of 6 Stock Synthesis output objects.
}
}
\usage{
ss_mle
}
\description{
Fabricated data showing 6 scenarios of Stock Synthesis models run through
Stock Synthesis (SS_3.30.20) (maximum likelihood estimation) and r4ss::SS_output().
}
\keyword{datasets}
