% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssandify.R
\name{ssandify}
\alias{ssandify}
\title{SSANDify plots
A quick wrapper to give other ggplots the 'look and feel' of SSAND. Mostly just a short cut for theme_bw() and default SSAND colours}
\usage{
ssandify(p, legend_title = "", x = NULL, y = NULL)
}
\arguments{
\item{p}{A ggplot object}

\item{legend_title}{Optional legend title. Default is blank}

\item{x}{Optional x-axis label}

\item{y}{Optional y-axis label}
}
\value{
A SSAND-looking plot
}
\description{
SSANDify plots
A quick wrapper to give other ggplots the 'look and feel' of SSAND. Mostly just a short cut for theme_bw() and default SSAND colours
}
\examples{
p <- ggplot2::ggplot(ggplot2::mpg) + ggplot2::geom_point(ggplot2::aes(displ, hwy, colour = class))
ssandify(p)
ssandify(p, x = "This", y = "That")
}
