% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarytable_prep_DD.R
\name{summarytable_prep_DD}
\alias{summarytable_prep_DD}
\title{Prepare DDUST data for summary table}
\usage{
summarytable_prep_DD(
  dd_mle,
  dd_mcmc,
  dd_sim,
  end_year = NULL,
  credible_interval = 0.95
)
}
\arguments{
\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{dd_mcmc}{A list of model fits from tmbstan::tmbstan() with one list element per scenario. Only needed if MCMC was used.}

\item{dd_sim}{A list of outputs from SSAND::simulate_DD() with one list element per scenario. Only required if MCMC was used. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{end_year}{The final year of biomass to be plotted. Calculated by default, but option is included in case model is run into the future.}

\item{credible_interval}{The credible interval to use to create dashed lines in the final plot}
}
\value{
A data frame that summarised biomass and fishing mortality estimates, ready for use in summarytable()
}
\description{
Prepare DDUST data for summary table
}
\examples{
\dontrun{
library(DDUST)
dd_sim <- simulate_DDUST(dd_mle,dd_mcmc)

# <<table_summary, results='asis', echo=FALSE>>=
dd_mcmc_ens <- mcmc_ensemble_DD(dd_mcmc,dd_sim,scenarios=c(1,2))$dd_mcmc
dd_sim_ens <- mcmc_ensemble_DD(dd_mcmc,dd_sim,scenarios=c(1,2))$dd_sim
data <- summarytable_prep_DD(dd_mle,dd_mcmc_ens,dd_sim_ens)
summarytable(data, label="tab:summary")
# @
}
}
