% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarytable_prep_SS.R
\name{summarytable_prep_SS}
\alias{summarytable_prep_SS}
\title{Prepare Stock Synthesis data for summary table}
\usage{
summarytable_prep_SS(
  ss_mle,
  ss_mcmc,
  end_year = NULL,
  credible_interval = 0.95,
  number_F_years = 5
)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{ss_mcmc}{A list of outputs from r4ss::SSgetMCMC() with one element per scenario. Only needed if MCMC was used. Will automatically reformat as a list if a single r4ss::SSgetMCMC() output (i.e. one scenario) is entered.}

\item{end_year}{The final year of biomass to be plotted. Calculated by default, but option is included in case model is run into the future.}

\item{credible_interval}{The credible interval to use to create dashed lines in the final plot}

\item{number_F_years}{Number of recent years to be included in the fishing pressure summary (e.g. number_F_years=5 will take the average of the five most recent years of the model)}
}
\value{
A data frame that summarised parameter estimate, ready for use in parametertable()
}
\description{
Prepare Stock Synthesis data for summary table
}
\examples{
# <<table_summary, results='asis', echo=FALSE>>=
data <- summarytable_prep_SS(ss_mle[[1]],ss_mcmc[[1]])
summarytable(data, label="tab:summary")
# @

# <<table_summary, results='asis', echo=FALSE>>=
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc,scenarios=c(1,2))
data <- summarytable_prep_SS(ss_mle,ss_mcmc_ens)
summarytable(data, label="tab:summary")
# @
}
