% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{Generate a stock assessment report template, configured to your model requirements.}
\usage{
template(
  dir = NULL,
  species = "ssand fish",
  filename = "sandfish",
  stock = "east coast",
  model = "SS",
  MCMC = TRUE,
  scenarios = 1:2,
  length = TRUE,
  age = FALSE,
  caal = TRUE,
  discard = FALSE,
  class_dir = NULL
)
}
\arguments{
\item{dir}{Directory where template will be saved. Defaults to Desktop.}

\item{species}{Name of species (plain language), to populate template}

\item{filename}{Name of species, as one word with underscores, to populate file name and Makefile}

\item{stock}{Name of stock, to populate template}

\item{model}{"SS" or "DD" for Stock Synthesis or DDUST respectively. Will populate the relevant content}

\item{MCMC}{TRUE or FALSE if MCMC was used. Will populate the relevant content}

\item{scenarios}{A vector of scenario numbers used. Will populate the appendix.}

\item{length}{TRUE or FALSE if length compositions were used. Will populate the relevant content}

\item{age}{TRUE or FALSE if age compositions were used. Will populate the relevant content}

\item{caal}{TRUE or FALSE if conditional age-at-length compositions were used. Will populate the relevant content}

\item{discard}{TRUE or FALSE if discard inputs were used. Will populate the relevant content}

\item{class_dir}{Directory of folder containing class file, macros and images.}
}
\value{
A folder containing an Rtex file containing a report template, and the associated files, copied to your directory of choice
}
\description{
A report file (.Rtex) will be save into your working directory (or other directory of choosing), as well as any associated files required for it to run.
}
\details{
Template can be customised to the model, presence of MCMC and data types, and jurisdiction style guides
}
\examples{
\dontrun{
template(dir=paste0(getwd(),"/ssandfish/"),
         species = "ssand fish",
         filename = "sandfish",
         stock = "east coast",
         model = "SS",
         MCMC = TRUE,
         scenarios = 1:2,
         length = TRUE,
         age = FALSE,
         caal = FALSE,
         discard = FALSE)
}
}
