% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_fishers.R
\name{topfishersplot}
\alias{topfishersplot}
\title{Top fishers plot}
\usage{
topfishersplot(
  data,
  show_daily = TRUE,
  species_of_interest = NULL,
  n_top_fishers = 10,
  interesting_years = NULL,
  interesting_fishers = NULL,
  show_region = FALSE,
  show_latband = FALSE,
  show_latband_regions = FALSE,
  remove_NA_latband = FALSE,
  extract_data = FALSE
)
}
\arguments{
\item{data}{Output of format_logbooks()}

\item{show_daily}{Set to TRUE to show daily catch of top fishers. Set to FALSE plot the total catch of fishers.}

\item{species_of_interest}{The name of the species of interest, as listed in the 'species' column of data.}

\item{n_top_fishers}{Numeric. Maximum number of top fishers to display on plot. Default is 20.}

\item{interesting_years}{Optional. A vector of years to filter the dataset; only fishers operating in this years will be included.}

\item{interesting_fishers}{Optional. A vector of fishers/operators to filter the dataset by.}

\item{show_region}{Set to TRUE to display top fishers by region}

\item{show_latband}{Set to TRUE to display top fishers by latitude band}

\item{show_latband_regions}{Set to TRUE to display top fishers by latitude band grouping}

\item{remove_NA_latband}{Set to TRUE to remove data without valid latitude band.}

\item{extract_data}{Set to TRUE to return data instead of plot. Default is FALSE.}
}
\value{
A plot of top fishers in the fishery
}
\description{
Top fishers plot
}
\examples{
\dontrun{topfishersplot(data = format_logbooks(raw_data))}
}
