% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_misreporting.R
\name{unusualcatch}
\alias{unusualcatch}
\title{Unusual catch table}
\usage{
unusualcatch(
  data,
  species_of_interest = NULL,
  interesting_fishers = NULL,
  interesting_years = NULL,
  max_days = 1,
  quantile = 0.9,
  multiplier = 10
)
}
\arguments{
\item{data}{Output of format_logbooks()}

\item{species_of_interest}{The name of the species of interest, as listed in the 'species' column of data.}

\item{interesting_fishers}{A vector of interesting fishers by which the data will be filtered}

\item{interesting_years}{A vector of interesting years by which the data will be filtered}

\item{max_days}{Numeric. Filter to include only records that were less than or equal to this maximum number of fishing days (i.e. exclude multi-day trips)}

\item{quantile}{Operators are identified with maximum daily catch is \code{multiplier} times higher than \code{quantile} percentile}

\item{multiplier}{Operators are identified with maximum daily catch is \code{multiplier} times higher than \code{quantile} percentile}
}
\value{
A dataframe of fishers whose activity has been flagged as unusual. This is defined as any fisher whose maximum catch record is greater than 10 times the 90th quantile of all their catch records. The multiplier and quantile can be specified.
}
\description{
Unusual catch table
}
\examples{
\dontrun{unusualcatch(data = format_logbooks(raw_data))}
}
