% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_misreporting.R
\name{unusualcatchplot}
\alias{unusualcatchplot}
\title{Unusual catch plot}
\usage{
unusualcatchplot(
  data,
  max_days = 1,
  interesting_fishers = NULL,
  extract_data = FALSE
)
}
\arguments{
\item{data}{Output of format_logbooks()}

\item{max_days}{Numeric. Filter to include only records that were less than or equal to this maximum number of fishing days (i.e. exclude multi-day trips)}

\item{interesting_fishers}{A vector of interesting fishers by which the data will be filtered}

\item{extract_data}{Set to TRUE to return data instead of plot. Default is FALSE.}
}
\value{
A plot of fishers whose activity has been flagged as unusual. This is defined as any fisher whose maximum catch record is greater than 10 times the 90th quantile of all their catch records. The multiplier and quantile can be specified.
}
\description{
Unusual catch plot
}
\examples{
\dontrun{unusualcatchplot(data = format_logbooks(raw_data))}
}
