% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\docType{data}
\name{upset}
\alias{upset}
\title{Fake data to be used in upsetplot() function.}
\format{
\subsection{\code{upset}}{

A data frame of fabricated logbook data.
}
}
\usage{
upset
}
\description{
Fabricated data showing sample logbook data with columns species, SiteID and method
}
\keyword{datasets}
