% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webcatchplot.R
\name{webcatchplot}
\alias{webcatchplot}
\title{Web catch plot}
\usage{
webcatchplot(
  data,
  weight_unit = "kg",
  xlab = "Year",
  ylab = "Queensland retained catch (t)",
  xbreaks = NULL,
  ybreaks = NULL,
  xlabels = NULL,
  ylabels = NULL,
  xlim = NULL,
  ylim = NULL,
  xangle = NULL,
  financial_year = FALSE,
  legend_position = c(0.2, 0.6),
  colours = c("#005572", "#31B7C3", "#005572", "#31B7C3"),
  text_size = 12
)
}
\arguments{
\item{data}{A data frame frame from webcatchplot_prep() or created from scratch, with variables called year (num), sector (chr), value (num) and method (chr).}

\item{weight_unit}{Set to "kg" if data are input in kilograms, otherwise tonnes are assumed.}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{ylab}{Label for y-axis (character). Default is "Queensland retained catch (t)".}

\item{xbreaks}{A vector of breaks between x-axis labels, used in ggplot2::scale_x_continous() (numeric).}

\item{ybreaks}{A vector of breaks between y-axis labels, used in ggplot2::scale_y_continous() (numeric).}

\item{xlabels}{A vector of labels for the x-axis breaks.}

\item{ylabels}{A vector of labels for the y-axis breaks.}

\item{xlim}{A vector of lower and upper x-axis limits (e.g. c(1950, 2020)) (numeric).}

\item{ylim}{A vector of lower and upper y-axis limits (e.g. c(0,1)) (numeric).}

\item{xangle}{Set to 90 to rotate x-axis labels 90 degrees.}

\item{financial_year}{Set to TRUE if the assessment was based on financial year (logical). Adjusts the x-axis to show full financial year notation.}

\item{legend_position}{Position of the legend ("none", "left", "right", "bottom", "top",
or two-element numeric vector for x and y position). Default is "top".}

\item{colours}{A vector of colours used (character). Avoid changing. These are from the DAF colour palette.}

\item{text_size}{Text size (num). Default is 12.}
}
\value{
A summary plot of catch for the DAF website
}
\description{
Web catch plot
}
\examples{
data <- webcatchplot_prep_SS(ss_mle)
webcatchplot(data)
}
