% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webcatchplot_prep_SS.R
\name{webcatchplot_prep_SS}
\alias{webcatchplot_prep_SS}
\title{Prepare Stock Synthesis data for webcatchplot()}
\usage{
webcatchplot_prep_SS(
  ss_mle,
  scenario = 1,
  commercial_fleets = c(1, 2, 3),
  commercial_data = c(1970, 2003, 2008, 2023),
  recreational_fleets = c(4, 5, 6),
  recreational_data = c(1995, 2001, 2005, 2011)
)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{scenario}{Which scenario to plot (numeric). Default is 1.}

\item{commercial_fleets}{A vector of fleet numbers in the ss_mle object that contribute to total commercial catch (numeric).}

\item{commercial_data}{A vector of years to define when commercial data exist. Data are entered in pairs.
For example commercial_data = c(1942,1972,1988,2023) implies that data exists from 1942 to 1972, then 1988 to 2023.}

\item{recreational_fleets}{A vector of fleet numbers in the ss_mle object that contribute to total commercial catch (numeric).}

\item{recreational_data}{A vector of years for which recreational data exist.}
}
\value{
A data frame with variables called year (num), sector (chr), value (num) and method (chr)
}
\description{
This prep function works on the assumption that all harvest to be plotted on the web summary is contained in a model fleet.
}
\examples{
data <- webcatchplot_prep_SS(ss_mle)
webcatchplot(data)
}
