% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightplot_prep_DD.R
\name{weightplot_prep_DD}
\alias{weightplot_prep_DD}
\title{Prepare DDUST data for SSAND::weightplot()}
\usage{
weightplot_prep_DD(
  dd_mle,
  scenarios = NULL,
  age_max = NULL,
  age_rec = NULL,
  sex = "female"
)
}
\arguments{
\item{dd_mle}{A list of outputs from DDUST::makefullreport() with one list element per scenario. Will automatically reformat as a list if a single DDUST::makefullreport() output (i.e. one scenario) is entered}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{age_max}{Maximum age to plot, in the unit of measurement used in the model.}

\item{age_rec}{Age of recruitment, in the unit of measurement used in the model.}

\item{sex}{Indicate which sex to plot. Enter 1 for female or 2 for male.}
}
\value{
A dataframe with xvar (int), weight (num), scenario (int), sex (character)
}
\description{
Prepare DDUST data for SSAND::weightplot()
}
\examples{
data <- weightplot_prep_DD(dd_mle, age_max=10, age_rec=2)
weightplot(data)
}
