% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightplot_prep_SS.R
\name{weightplot_prep_SS}
\alias{weightplot_prep_SS}
\title{Prepare Stock Synthesis data for SSAND::weightplot()}
\usage{
weightplot_prep_SS(ss_mle, scenarios = NULL)
}
\arguments{
\item{ss_mle}{A list of outputs from r4ss::SS_output() with one element per scenario. Will automatically reformat as a list if a single r4ss::SS_output() output (i.e. one scenario) is entered.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}
}
\value{
A dataframe with xvar (int), maturityxfecundity (num), x (chr), scenario (int)
}
\description{
Prepare Stock Synthesis data for SSAND::weightplot()
}
\examples{
data <- weightplot_prep_SS(ss_mle)
weightplot(data)
}
