% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolutebiomassplot.R
\name{absolutebiomassplot}
\alias{absolutebiomassplot}
\title{Absolute biomass plot}
\usage{
absolutebiomassplot(
  data,
  show_inputs = TRUE,
  show_fits = TRUE,
  show_observed_error = FALSE,
  fleets = NULL,
  fleet_names = NULL,
  scenarios = NULL,
  scenario_labels = NULL,
  scenario_order = NULL,
  xlab = "Year",
  ylab = "Catch rate (kg/operation day)",
  xbreaks = NULL,
  ybreaks = NULL,
  xlim = NULL,
  ylim = NULL,
  xlabels = NULL,
  ylabels = NULL,
  show_colour = TRUE,
  show_CI_ribbon = TRUE,
  show_error_bar = FALSE,
  show_negative = TRUE,
  point_size = 2,
  colours = c("black", fq_palette("cols")),
  ncol = 2,
  scales = "free",
  seasonal = NULL,
  show_point = TRUE,
  show_line = NULL,
  show_dates_on_axis = FALSE,
  sample = NULL,
  band_colour = "black",
  band_labels = NULL,
  show_median = "median_cpue",
  mcmc_style = "banded",
  line_width = 1,
  hair_width = 0.5,
  input_colour = "black",
  input_range_colour = "black",
  aggregate_scenarios = FALSE,
  CI_range = 0.95,
  alpha = NULL,
  boxplot_outliers = TRUE,
  xangle = NULL,
  legend_position = NULL,
  financial_year = FALSE,
  text_size = NULL,
  legend_text_size = NULL,
  text_colour = NULL,
  legend_text_colour = NULL,
  legend_box = NULL,
  legend_title_blank = NULL,
  panel_border = NULL,
  panel_border_colour = NULL
)
}
\arguments{
\item{data}{Output from absolutebiomassplot_prep(). A data frame with variables date (date), fleet (int), obs (num), exp (num), ub (num), lb (num), scenario (fac)}

\item{show_inputs}{Set to TRUE to show model inputs.}

\item{show_fits}{Set to TRUE to show model fits.}

\item{show_observed_error}{Set to TRUE to display error bars around input data (MCMC only)}

\item{fleets}{A numeric or vector of fleet numbers to plot}

\item{fleet_names}{A vector of customised fleet names for legend}

\item{scenarios}{A vector of scenario numbers to be shown on plot (numeric). This was already specified in prep file, but this is a manual override to save running the prep function again.}

\item{scenario_labels}{Customise facet labels for individual scenarios}

\item{scenario_order}{A vector to reorder how scenarios are displayed (character). Use the label names defined in "scenario_labels".
If "scenario_labels" is left blank, the labels will be "Scenario 1", "Scenario 2" etc.
Any scenarios not included in "scenario_order" will be tacked on in the order they appear in the input data.}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{ylab}{Label for y-axis (character). Default is "Catch rate (kg/operation day)".}

\item{xbreaks}{A vector of breaks between x-axis labels, used in ggplot2::scale_x_continous() (numeric).}

\item{ybreaks}{A vector of breaks between y-axis labels, used in ggplot2::scale_y_continous() (numeric).}

\item{xlim}{A vector of lower and upper x-axis limits (e.g. c(1950, 2020)) (numeric).}

\item{ylim}{A vector of lower and upper y-axis limits (e.g. c(0,1)) (numeric).}

\item{xlabels}{A vector of labels for the x-axis breaks.}

\item{ylabels}{A vector of labels for the y-axis breaks.}

\item{show_colour}{Set to TRUE to activate coloured mode, or FALSE for greyscale (logical).}

\item{show_CI_ribbon}{Set to TRUE to show a confidence interval ribbon (logical).}

\item{show_error_bar}{Set to TRUE to include error bar (logical). If wanting to use confidence interval lines with error bars use make error_bar = FALSE}

\item{show_negative}{Set to TRUE to allow confidence interval to be less than zero (logical).}

\item{point_size}{Size of points used in ggplot2::geom_line(). Default is 1.5.}

\item{colours}{A vector of colours used (character).}

\item{ncol}{Number of columns for facet_wrap(). Default is 2.}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'free', see ?ggplot2::facet_wrap for options.}

\item{seasonal}{Set to TRUE if the model is seasonal (i.e. has more than one timestep per year), to display a more detailed x-axis. If missing, or set to FALSE, the x-axis will just display year, not month.}

\item{show_point}{Set to TRUE to display fit/s as points rather than a line (e.g. if there is only one data point)}

\item{show_line}{Set to FALSE to show only points, not line (if there is only one observation in a scenario for example)}

\item{show_dates_on_axis}{Set to TRUE to show full dates on x-axis as opposed to years.}

\item{sample}{Number of samples to plot from each MCMC chain to ease burden of rendering dense plots (numeric).}

\item{band_colour}{Colour of bands (character). Only used when mcmc_style=="banded". Input one colour, bands will be distinguished using an alpha.}

\item{band_labels}{Labels for bands. Default is NULL and interval is used.}

\item{show_median}{Type of median shown. Default "annual_biomass" shows the median of each year,
"trajectory" shows median trajectory based on biomass in final year,}

\item{mcmc_style}{The type of MCMC plot to be displayed (character). Options are "banded", "hairy", "boxplot", "CI" and "joy", the default is "banded". Only one option can be selected.}

\item{line_width}{Width of median lines (numeric). Default is 1.}

\item{hair_width}{Width of fine MCMC hairs (numeric). Default is 0.5.}

\item{input_colour}{Specify colour of input data (MCMC)}

\item{input_range_colour}{Specify colour of error around input data (MCMC)}

\item{aggregate_scenarios}{Set to TRUE to calculate credible intervals across all scenarios (logical). Only activated if mcmc_style==CI.}

\item{CI_range}{Specify credible interval range (numeric). Only activated if mcmc_style==CI.}

\item{alpha}{Transparency for range (numeric) used in ggplot2::geom_density_ridges(). Default is 0.7.}

\item{boxplot_outliers}{Set to FALSE to remove outlier points from boxplot. Default is TRUE.}

\item{financial_year}{Set to TRUE if the assessment was based on financial year (logical). Adds 1 to each year in the dataset.}

\item{text_size, legend_text_size, text_colour, legend_text_colour, legend_position, legend_box, legend_title_blank, panel_border, panel_border_colour, xangle}{Optional plotting theme overrides. Defaults are controlled by \code{theme_ssand()}
and can be set globally via \code{set_ssand_style()}.}
}
\value{
A plot that shows input data and model fits to CPUE data
}
\description{
Absolute biomass plot
}
\examples{
\dontrun{
data <- absolutebiomassplot_prep_DD(dd_mle)
absolutebiomassplot(data)
absolutebiomassplot(data, xlim=c(1954,2023))
}
}
