% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageplot.R
\name{ageplot}
\alias{ageplot}
\title{Age composition plot with an option to add model fits}
\usage{
ageplot(
  data,
  show_fits = TRUE,
  fleet = 1,
  point_size = 1,
  colours = c("#9E480E", "#7CC8FC"),
  legend_position = "top",
  xlab = "Age (years)",
  ylab = "Proportion",
  xbreaks = NULL,
  ybreaks = NULL,
  xlabels = NULL,
  ylabels = NULL,
  xlim = NULL,
  ylim = NULL,
  financial_year = FALSE,
  scales = "fixed",
  ncol = 4,
  direction = "v",
  MLS = NULL,
  MLS_colour = "black",
  MLS_label = "MLS"
)
}
\arguments{
\item{data}{Output from ageplot_prep(). A data frame with year (int), fleet (int), bin (int), sex (chr), obs (num), exp (num), scenario (num)}

\item{show_fits}{Set to TRUE to show model fits.}

\item{fleet}{Specify which fleet to plot (numeric). By default, fleet 1 will be shown.}

\item{point_size}{Size of points used in ggplot2::geom_line(). Default is 1.}

\item{colours}{A vector of colours used for sex types (character).}

\item{legend_position}{Position of the legend ("none", "left", "right", "bottom", "top", or two-element numeric vector for x and y position). Default is "top".}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{ylab}{Label for y-axis (character). Default is "Spawning biomass (relative)".}

\item{xbreaks}{A vector of breaks between x-axis labels, used in ggplot2::scale_x_continous() (numeric).}

\item{ybreaks}{A vector of breaks between y-axis labels, used in ggplot2::scale_y_continous() (numeric).}

\item{xlabels}{A vector of labels for the x-axis breaks.}

\item{ylabels}{A vector of labels for the y-axis breaks.}

\item{xlim}{A vector of lower and upper x-axis limits (e.g. c(1950, 2020)) (numeric).}

\item{ylim}{A vector of lower and upper y-axis limits (e.g. c(0,1)) (numeric).}

\item{financial_year}{Set to TRUE if the assessment was based on financial year (logical). Adjusts the x-axis to show full financial year notation.}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'free', see ?ggplot2::facet_wrap for options.}

\item{ncol}{Number of columns for facet_wrap(). Default is 2.}

\item{direction}{Direction of facet wrap (character). Enter "v" for vertical or "h" for horizontal.}

\item{MLS}{Value of minimum legal size (or equivalent) (numeric). Leave blank to exclude from plot.}

\item{MLS_colour}{Colour of MLS line (character). Default is "black".}

\item{MLS_label}{Label of MLS line (character). Default is "MLS".}
}
\value{
An age plot, with an option to add model fits
}
\description{
Age composition plot with an option to add model fits
}
\examples{
\dontrun{
data <- ageplot_prep_SS(ss_mle)
ageplot(data, show_fits=FALSE)
ageplot(data)
}
}
