% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andreplot.R
\name{andreplot}
\alias{andreplot}
\title{Andre plot}
\usage{
andreplot(
  data,
  scenario = 1,
  fleet = 1,
  years = NULL,
  xlab = "Length (cm)",
  ylab = c("Age", "Standard deviation (age)"),
  colours = c("black", "grey80", fq_palette("DAF")),
  legend_ratio = c(1, 10),
  xangle = NULL,
  legend_position = "top",
  financial_year = FALSE,
  text_size = NULL,
  legend_text_size = NULL,
  text_colour = NULL,
  legend_text_colour = NULL,
  legend_box = NULL,
  legend_title_blank = NULL,
  panel_border = NULL,
  panel_border_colour = NULL
)
}
\arguments{
\item{data}{Output of andreplot_prep_SS. A data frame with columns "year" (num), "length" (num), "obs" (num), "pred" (num), "low" (num), "upp" (num), "label" (chr), "scenario" (int), "CI" (num)}

\item{scenario}{The scenario to plot. Default is 1.}

\item{fleet}{The fleet to plot. Default is 1.}

\item{years}{A vector of years to plot}

\item{xlab}{Label for x-axis (character). Default is "Length (cm)".}

\item{ylab}{A vector of labels for y-axis (character). Two entries, for each column of the plot. Default is c("Age","Standard deviation (age)").}

\item{colours}{A vector of colours to plot. First is the colour of points, then the colour of the ribbon, then the colour of the expected line.}

\item{legend_ratio}{A vector specifying the relative heights of the legend and the rest of the plot. Default is c(1,10)}

\item{financial_year}{Set to TRUE if the assessment was based on financial year (logical). Adjusts the x-axis to show full financial year notation.}

\item{text_size, legend_text_size, text_colour, legend_text_colour, legend_position, legend_box, legend_title_blank, panel_border, panel_border_colour, xangle}{Optional plotting theme overrides. Defaults are controlled by \code{theme_ssand()}
and can be set globally via \code{set_ssand_style()}.}
}
\value{
Andre plot with an option to filter by fleet
}
\description{
Andre plot
}
\examples{
data <- andreplot_prep_SS(ss_mle, sex_code=1)
andreplot(data, fleet = 1)
}
