% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andreplot.R
\name{andreplot}
\alias{andreplot}
\title{Andre plot}
\usage{
andreplot(
  data,
  scenario = 1,
  years = NULL,
  xlab = "Length (cm)",
  ylab = c("Age", "Standard deviation (age)"),
  colours = c("black", "grey80", fq_palette("DAF")),
  legend_position = "top",
  legend_ratio = c(1, 10)
)
}
\arguments{
\item{data}{Output of andreplot_prep_SS. A data frame with columns "year" (num), "length" (num), "obs" (num), "pred" (num), "low" (num), "upp" (num), "label" (chr), "scenario" (int), "CI" (num)}

\item{scenario}{The scenario to plot. Default is 1.}

\item{years}{A vector of years to plot}

\item{xlab}{Label for x-axis (character). Default is "Length (cm)".}

\item{ylab}{A vector of labels for y-axis (character). Two entries, for each column of the plot. Default is c("Age","Standard deviation (age)").}

\item{colours}{A vector of colours to plot. First is the colour of points, then the colour of the ribbon, then the colour of the expected line.}

\item{legend_position}{Position of the legend ("top" or "bottom"). Default is "top".}

\item{legend_ratio}{A vector specifying the relative heights of the legend and the rest of the plot. Default is c(1,10)}
}
\value{
Andre plot
}
\description{
Andre plot
}
\examples{
data <- andreplot_prep_SS(ss_mle, sex_code=1)
andreplot(data)
}
