% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassplot.R
\name{biomassplot}
\alias{biomassplot}
\title{Biomass plot}
\usage{
biomassplot(
  data,
  xlab = "Year",
  ylab = NULL,
  xbreaks = NULL,
  ybreaks = NULL,
  xlabels = NULL,
  ylabels = NULL,
  xlim = NULL,
  ylim = NULL,
  colours = NULL,
  alpha = NULL,
  line_width = 1,
  hair_width = 0.5,
  annotation_position = min(data$year) + 1,
  show_target_line = TRUE,
  target_value = 0.6,
  show_limit_line = TRUE,
  limit_value = 0.2,
  scenarios = NULL,
  scenario_labels = NULL,
  scenario_order = NULL,
  scales = "fixed",
  ncol = 2,
  sample = NULL,
  show_median = c("trajectory", "annual_biomass"),
  mcmc_style = "banded",
  aggregate_scenarios = FALSE,
  CI_range = 0.95,
  show_CI = TRUE,
  band_colour = "black",
  band_labels = NULL,
  rel_min_height = 0.01,
  ridge_scale = 4.5,
  ridge_colour = c("grey30", "black"),
  shapes = c(16, 18, 17),
  show_final_biomass = FALSE,
  boxplot_outliers = TRUE,
  xangle = NULL,
  legend_position = NULL,
  financial_year = FALSE,
  text_size = NULL,
  legend_text_size = NULL,
  text_colour = NULL,
  legend_text_colour = NULL,
  legend_box = NULL,
  legend_title_blank = NULL,
  panel_border = NULL,
  panel_border_colour = NULL
)
}
\arguments{
\item{data}{If MCMC is being used, a data frame with variables called rownum, scenario, year, value, med, interval, prob_lower, prob_upper, biomass_type, biomass_definition. If MLE, a data frame with variables called year, value, lower, upper, scenario, biomass_type, biomass_definition.}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{ylab}{Label for y-axis (character). Default is "Spawning biomass (relative)".}

\item{xbreaks}{A vector of breaks between x-axis labels, used in ggplot2::scale_x_continous() (numeric).}

\item{ybreaks}{A vector of breaks between y-axis labels, used in ggplot2::scale_y_continous() (numeric).}

\item{xlabels}{A vector of labels for the x-axis breaks.}

\item{ylabels}{A vector of labels for the y-axis breaks.}

\item{xlim}{A vector of lower and upper x-axis limits (e.g. c(1950, 2020)) (numeric).}

\item{ylim}{A vector of lower and upper y-axis limits (e.g. c(0,1)) (numeric).}

\item{colours}{A vector of colours used for median types (character).}

\item{alpha}{Transparency for range (numeric) used in ggplot2::geom_density_ridges(). Default is 0.7.}

\item{line_width}{Width of median lines (numeric). Default is 1.}

\item{hair_width}{Width of fine MCMC hairs (numeric). Default is 0.5.}

\item{annotation_position}{Horizontal position of annotation}

\item{show_target_line}{Set to TRUE to show target reference point line (logical).}

\item{target_value}{target reference point, default is 0.6. Colour for line is second element of colours}

\item{show_limit_line}{show limit reference point}

\item{limit_value}{limit reference point, default is 0.2. Colour for line is third element of colours}

\item{scenarios}{A vector of scenario numbers to be shown on plot (numeric). This was already specified in prep file, but this is a manual override to save running the prep function again.}

\item{scenario_labels}{A vector of customised scenario names (character). Default is "Scenario 1", "Scenario 2", etc.}

\item{scenario_order}{A vector to reorder how scenarios are displayed (character). Use the label names defined in "scenario_labels".
If "scenario_labels" is left blank, the labels will be "Scenario 1", "Scenario 2" etc.
Any scenarios not included in "scenario_order" will be tacked on in the order they appear in the input data.}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'fixed', see ?ggplot2::facet_wrap for options.}

\item{ncol}{Number of columns for facet_wrap(). Default is 2.}

\item{sample}{Number of samples to plot from each MCMC chain to ease burden of rendering dense plots (numeric).}

\item{show_median}{Type of median shown. Default "annual_biomass" shows the median of each year,
"trajectory" shows median trajectory based on biomass in final year}

\item{mcmc_style}{The type of MCMC plot to be displayed (character). Options are "banded", "hairy", "boxplot", "CI" and "joy", the default is "banded". Only one option can be selected.}

\item{aggregate_scenarios}{Set to TRUE to calculate credible intervals across all scenarios (logical). Only activated if mcmc_style==CI.}

\item{CI_range}{Specify credible interval range (numeric). Only activated if mcmc_style==CI.}

\item{show_CI}{Set to TRUE to show CI range on joy plot (logical).}

\item{band_colour}{Colour of bands (character). Only used when mcmc_style=="banded". Input one colour, bands will be distinguished using an alpha.}

\item{band_labels}{Labels for bands. Default is NULL and interval is used.}

\item{rel_min_height}{Only used when mcmc_style=="joy". Lines with heights below this cutoff will be removed (passed to geom_density_ridges). Default is 0.01.}

\item{ridge_scale}{Only used when mcmc_style=="joy". Scale the height of the ridgelines relative to the spacing between them (passed to geom_density_ridges). Default is 4.5.}

\item{ridge_colour}{Two-element vector for the fill and outline of ridges (character). Only used when mcmc_style=="joy".}

\item{shapes}{Vector of shapes to denote different median types. Only used when mcmc_style=="joy".}

\item{show_final_biomass}{Set to TRUE to show final biomass value at the end of the time series.}

\item{boxplot_outliers}{Set to FALSE to remove outlier points from boxplot. Default is TRUE.}

\item{financial_year}{Set to TRUE if the assessment was based on financial year (logical). Adjusts the x-axis to show full financial year notation.}

\item{text_size, legend_text_size, text_colour, legend_text_colour, legend_position, legend_box, legend_title_blank, panel_border, panel_border_colour, xangle}{Optional plotting theme overrides. Defaults are controlled by \code{theme_ssand()}
and can be set globally via \code{set_ssand_style()}.}
}
\value{
Biomass plot
}
\description{
MLE biomass plot displays 95\% confidence interval
Multiple options are available for MCMC plots.
The variable "mcmc_type" can be set to "hairy", "boxplot", "banded", "joy" or "CI" to display different plot types.
On top of this, one or more median types can be overlaid using "show_median".
The types of median to be shown are "annual_biomass" or "trajectory".
}
\examples{
data <- biomassplot_prep_DD(dd_mle)
biomassplot(data)

data <- biomassplot_prep_SS(ss_mle, ss_mcmc)
biomassplot(data, mcmc_style = "banded", show_median = c("annual_biomass","trajectory"))
biomassplot(data, mcmc_style = "boxplot", show_median = c("annual_biomass","trajectory"))
biomassplot(data, mcmc_style = "hairy", show_median = c("annual_biomass","trajectory"))
biomassplot(data, mcmc_style = "CI", show_median = c("annual_biomass","trajectory"), CI_range = 0.9)
#biomassplot(data, mcmc_style = "joy", show_median = c("none"))

# Theme setting:
#biomassplot(data)
#biomassplot(data, text_size = 16)
#set_ssand_style(text_colour = "red")
#biomassplot(data)
}
