% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomasstable.R
\name{biomasstable}
\alias{biomasstable}
\title{Biomass table}
\usage{
biomasstable(
  data,
  label,
  row_names = NULL,
  column_names = NULL,
  caption = NULL,
  align = NULL
)
}
\arguments{
\item{data}{Dataframe output from biomasstable_prep()}

\item{label}{A label for the table that can be referenced elsewhere in the report using LaTeX syntax}

\item{row_names}{An option to customise the text in the Indicator columns}

\item{column_names}{An optional vector of column names}

\item{caption}{Caption for table}

\item{align}{An optional vector of alignment for the table, input as a character string that is one character longer than the number of columns in the table. Default is centered.}
}
\value{
Code that can be used within a code chunk in a RMarkdown style document to produce a biomass table
}
\description{
Biomass table
}
\examples{
# <<table_biomass, results='asis', echo=FALSE>>=
data <- biomasstable_prep_SS(ss_mle,ss_mcmc)
biomasstable(data, label="tab:biomass")
# @
}
