% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caal_agefitplot.R
\name{caal_agefitplot}
\alias{caal_agefitplot}
\title{Plot to show age fits from conditional-age-at-length data}
\usage{
caal_agefitplot(
  data,
  scenario = 1,
  fleet = 1,
  colours = c("grey70", "black", "black"),
  line_width = 1,
  ncol = 3,
  scales = "free",
  point_size = 1.5,
  show_fits = TRUE,
  xlab = "Age (years)",
  ylab = "Proportion",
  xbreaks = NULL,
  ybreaks = NULL,
  xlabels = NULL,
  ylabels = NULL,
  xlim = NULL,
  ylim = NULL,
  xangle = NULL,
  legend_position = NULL,
  financial_year = FALSE,
  text_size = NULL,
  legend_text_size = NULL,
  text_colour = NULL,
  legend_text_colour = NULL,
  legend_box = NULL,
  legend_title_blank = NULL,
  panel_border = NULL,
  panel_border_colour = NULL
)
}
\arguments{
\item{data}{A data frame with year (int), bin (int), obs (num), exp (num), scenario (int), sex (int), fleet (int)}

\item{scenario}{A single scenario number to plot (numeric). Default is 1.}

\item{fleet}{A numeric of fleet number to plot}

\item{colours}{A vector of colours used (character). First element is bar fill, then line colour, then point colour.}

\item{line_width}{Width of lines.}

\item{ncol}{Number of columns for facet_wrap(). Default is 3.}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'free', see ?ggplot2::facet_wrap for options.}

\item{point_size}{Size of points used in ggplot2::geom_line(). Default is 1.5.}

\item{show_fits}{Set to TRUE to show model fits. Set to FALSE to show model 'inputs'.
When TRUE, the input data are transformed into proportions rather than absolute values.}

\item{xlab}{Label for x-axis (character). Default is "Age (years)".}

\item{ylab}{Label for y-axis (character). Default is "Sample size".}

\item{xbreaks}{A vector of breaks between x-axis labels, used in ggplot2::scale_x_continous() (numeric).}

\item{ybreaks}{A vector of breaks between y-axis labels, used in ggplot2::scale_y_continous() (numeric).}

\item{xlabels}{A vector of labels for the x-axis breaks.}

\item{ylabels}{A vector of labels for the y-axis breaks.}

\item{xlim}{A vector of lower and upper x-axis limits (e.g. c(1950, 2020)) (numeric).}

\item{ylim}{A vector of lower and upper y-axis limits (e.g. c(0,1)) (numeric).}

\item{financial_year}{Set to TRUE if the assessment was based on financial year (logical). Adjusts the x-axis to show full financial year notation.}

\item{text_size, legend_text_size, text_colour, legend_text_colour, legend_position, legend_box, legend_title_blank, panel_border, panel_border_colour, xangle}{Optional plotting theme overrides. Defaults are controlled by \code{theme_ssand()}
and can be set globally via \code{set_ssand_style()}.}
}
\value{
A plot of age fits from conditional-age-at-length data
}
\description{
Plot to show age fits from conditional-age-at-length data
}
\examples{
data <- caal_agefitplot_prep_SS(ss_mle)
caal_agefitplot(data, scenario=2,show_fits=FALSE)
caal_agefitplot(data, scenario=2)
}
