% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caal_agefitplot.R
\name{caal_agefitplot}
\alias{caal_agefitplot}
\title{Plot to show age fits from conditional-age-at-length data}
\usage{
caal_agefitplot(
  data,
  scenario = 1,
  colours = c("grey70", "black", "black"),
  line_width = 1,
  ncol = 3,
  scales = "free",
  point_size = 1.5,
  xlab = "Age (years)",
  ylab = "Sample size",
  show_fits = TRUE
)
}
\arguments{
\item{data}{A data frame with year (int), bin (int), obs (num), exp (num), scenario (int), sex (int)}

\item{scenario}{A single scenario number to plot (numeric). Default is 1.}

\item{colours}{A vector of colours used (character). First element is bar fill, then line colour, then point colour.}

\item{line_width}{Width of lines.}

\item{ncol}{Number of columns for facet_wrap(). Default is 3.}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'free', see ?ggplot2::facet_wrap for options.}

\item{point_size}{Size of points used in ggplot2::geom_line(). Default is 1.5.}

\item{xlab}{Label for x-axis (character). Default is "Age (years)".}

\item{ylab}{Label for y-axis (character). Default is "Sample size".}

\item{show_fits}{Set to TRUE to show model fits. Set to FALSE to show model 'inputs'.
When TRUE, the input data are transformed into proportions rather than absolute values.}
}
\value{
A plot of age fits from conditional-age-at-length data
}
\description{
Plot to show age fits from conditional-age-at-length data
}
\examples{
data <- caal_agefitplot_prep_SS(ss_mle)
caal_agefitplot(data, scenario=1,show_fits=FALSE)
caal_agefitplot(data, scenario=1)
}
