% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchabilityplot.R
\name{catchabilityplot}
\alias{catchabilityplot}
\title{Plot catchability pattern}
\usage{
catchabilityplot(
  data,
  xlab = "Month",
  ylab = "Catchability coefficient (q)",
  text_size = 12,
  ncol = 2,
  scales = "free",
  colours = fq_palette("alisecolours"),
  scenarios = NULL,
  scenario_labels = NULL,
  scenario_order = NULL
)
}
\arguments{
\item{data}{Output from catchabilityplot_prep(). A dataframe with q (num), month (int), month_point (int), monthnames (factor), fleet (factor), scenario (factor)}

\item{xlab}{Label for x-axis (character). Default is "Month".}

\item{ylab}{Label for y-axis (character). Default is "Catchability coefficient (q)".}

\item{text_size}{Text size (num). Default is 12.}

\item{ncol}{Number of columns for facet_wrap(). Default is 2.}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'free', see ?ggplot2::facet_wrap for options.}

\item{colours}{A vector of colours used for lines (character).}

\item{scenarios}{A vector of scenario numbers to be shown on plot (numeric). This was already specified in prep file, but this is a manual override to save running the prep function again.}

\item{scenario_labels}{A vector of customised scenario names (character). Default is "Scenario 1", "Scenario 2", etc.}

\item{scenario_order}{A vector to reorder how scenarios are displayed (character). Use the label names defined in "scenario_labels".
If "scenario_labels" is left blank, the labels will be "Scenario 1", "Scenario 2" etc.
Any scenarios not included in "scenario_order" will be tacked on in the order they appear in the input data.}
}
\value{
A plot of catchability pattern
}
\description{
Plot catchability pattern
}
\examples{
data <- catchabilityplot_prep_DD(dd_mle)
catchabilityplot(data)

data <- catchabilityplot_prep_SS(ss_mle)
catchabilityplot(data)
}
