% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchplot.R
\name{catchplot}
\alias{catchplot}
\title{Catch plot}
\usage{
catchplot(
  data,
  xlab = "Year",
  ylab = NULL,
  xbreaks = NULL,
  ybreaks = NULL,
  xlabels = NULL,
  ylabels = NULL,
  xlim = NULL,
  ylim = NULL,
  xangle = NULL,
  financial_year = FALSE,
  fleet_names = NULL,
  colours = NULL,
  legend_position = "top",
  reverse = FALSE,
  strip_position = NA,
  show_annual_aggregate = FALSE,
  show_dates_on_axis = FALSE,
  scenarios = NULL,
  scenario_labels = NULL,
  scenario_order = NULL,
  scales = "free",
  ncol = 2
)
}
\arguments{
\item{data}{Output from catchplot_prep(). A dataframe with date (date), value (int), fleet (fac), scenario (int)}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{ylab}{Label for y-axis (character). Default is "Retained catch (t)" for DDUST models. Default is determined by catchplot_prep_SS partition setting for SS models.}

\item{xbreaks}{A vector of breaks between x-axis labels, used in ggplot2::scale_x_continous() (numeric).}

\item{ybreaks}{A vector of breaks between y-axis labels, used in ggplot2::scale_y_continous() (numeric).}

\item{xlabels}{A vector of labels for the x-axis breaks.}

\item{ylabels}{A vector of labels for the y-axis breaks.}

\item{xlim}{A vector of lower and upper x-axis limits (e.g. c(1950, 2020)) (numeric).}

\item{ylim}{A vector of lower and upper y-axis limits (e.g. c(0,1)) (numeric).}

\item{xangle}{Set to 90 to rotate x-axis labels 90 degrees.}

\item{financial_year}{Set to TRUE if the assessment was based on financial year (logical). Adjusts the x-axis to show full financial year notation.}

\item{fleet_names}{A vector of customised fleet names for legend (character).}

\item{colours}{A vector of colours used (character).}

\item{legend_position}{Position of the legend ("none", "left", "right", "bottom", "top", or two-element numeric vector for x and y position). Default is "top".}

\item{reverse}{Set to TRUE to reverse the default stacking order (logical). This is useful if you're rotating both the plot and legend.}

\item{strip_position}{Customise ggplot2::facet_wrap() strips. By default, the labels are displayed on the top of the plot. Using strip_position it is possible to place the labels on either of the four sides by setting strip.position = c("top", "bottom", "left", "right").}

\item{show_annual_aggregate}{Set to TRUE to aggregate annually (logical). Default is FALSE.}

\item{show_dates_on_axis}{Set to TRUE to show full dates on x-axis as opposed to years.}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{scenario_labels}{A vector of customised scenario names (character). Default is "Scenario 1", "Scenario 2", etc.}

\item{scenario_order}{A vector to reorder how scenarios are displayed (character). Use the label names defined in "scenario_labels".
If "scenario_labels" is left blank, the labels will be "Scenario 1", "Scenario 2" etc.
Any scenarios not included in "scenario_order" will be tacked on in the order they appear in the input data.}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'free', see ?ggplot2::facet_wrap for options.}

\item{ncol}{Number of columns for facet_wrap(). Default is 2.}
}
\value{
Catch plot
}
\description{
Catch plot
}
\examples{
data <- catchplot_prep_SS(ss_mle)
catchplot(data, fleet_names = "Commercial")
catchplot(data, fleet_names = "Commercial", financial_year=TRUE)

data <- catchplot_prep_DD(dd_mle)
catchplot(data)
}
