% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationplot.R
\name{correlationplot}
\alias{correlationplot}
\title{Correlation Plot}
\usage{
correlationplot(
  data,
  labels = NULL,
  show_diagonal = FALSE,
  columns = NULL,
  colours = NULL,
  sample = NULL
)
}
\arguments{
\item{data}{A data frame made from correlationplot_prep. If an MLE plot, variables are X1 (fac), X2 (fac), value (num), MCMC (log). If MCMC, variables are chain (fac), iter (int) followed by columns for each parameter to plot.}

\item{labels}{A vector of labels for the parameters shown on the plot}

\item{show_diagonal}{Set to TRUE to show values on the diagnoal for MCMC plot (logical). Default is FALSE.}

\item{columns}{A vector of columns to display on plot (character).}

\item{colours}{A vector of colours used (character).}

\item{sample}{Number of samples to plot from each MCMC chain to ease burden of rendering dense plots (numeric).}
}
\value{
Correlation plot
}
\description{
Correlation Plot
}
\examples{
parameters <- extract_SS_parameters(ss_mle)[c(2,3,4,26),]
data <- correlationplot_prep_SS(ss_mle, scenario = 1, parameters = parameters)
correlationplot(data)
}
