% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpueqplot.R
\name{cpueqplot}
\alias{cpueqplot}
\title{A plot that illustrates the impact of catchability rescaling how the model perceives catch rates}
\usage{
cpueqplot(
  data,
  xlab = "Year",
  ylab = "Catch rate (kg/fisher day)",
  text_size = 12,
  colours = c("#FFC000", "#9E480E", "#70AD47")
)
}
\arguments{
\item{data}{Output from cpueqplot_prep()}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{ylab}{Label for y-axis (character). Default is "Catch rate (kg/fisher day)".}

\item{text_size}{Text size (num). Default is 12.}

\item{colours}{A vector of colours used for scenarios (character).}
}
\value{
A plot that illustrates the impact of catchability rescaling how the model perceives catch rates
}
\description{
A plot that illustrates the impact of catchability rescaling how the model perceives catch rates
}
