% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discardplot.R
\name{discardplot}
\alias{discardplot}
\title{Discard plot}
\usage{
discardplot(
  data,
  show_fits = TRUE,
  fleets = NULL,
  fleet_names = NULL,
  scenarios = NULL,
  scenario_labels = NULL,
  scenario_order = NULL,
  xlab = "Year",
  ylab = "Discarded catch (kg)",
  xbreaks = NULL,
  ybreaks = NULL,
  xlim = NULL,
  ylim = NULL,
  xlabels = NULL,
  ylabels = NULL,
  xangle = NULL,
  financial_year = FALSE,
  text_size = 12,
  point_size = 1.5,
  show_dates_on_axis = FALSE,
  colours = NULL
)
}
\arguments{
\item{data}{A dataframe from discardplot_prep with year, fleet, obs, upper, lower, exp}

\item{show_fits}{Set to TRUE to show model fits.}

\item{fleets}{A numeric or vector of fleet numbers to plot}

\item{fleet_names}{A vector of customised fleet names for legend}

\item{scenarios}{A vector of scenario numbers to be shown on plot (numeric). This was already specified in prep file, but this is a manual override to save running the prep function again.}

\item{scenario_labels}{Customise facet labels for individual scenarios}

\item{scenario_order}{A vector to reorder how scenarios are displayed (character). Use the label names defined in "scenario_labels".
If "scenario_labels" is left blank, the labels will be "Scenario 1", "Scenario 2" etc.
Any scenarios not included in "scenario_order" will be tacked on in the order they appear in the input data.}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{ylab}{Label for y-axis (character). Default is "Catch rate (kg/operation day)".}

\item{xbreaks}{A vector of breaks between x-axis labels, used in ggplot2::scale_x_continous() (numeric).}

\item{ybreaks}{A vector of breaks between y-axis labels, used in ggplot2::scale_y_continous() (numeric).}

\item{xlim}{A vector of lower and upper x-axis limits (e.g. c(1950, 2020)) (numeric).}

\item{ylim}{A vector of lower and upper y-axis limits (e.g. c(0,1)) (numeric).}

\item{xlabels}{A vector of labels for the x-axis breaks.}

\item{ylabels}{A vector of labels for the y-axis breaks.}

\item{xangle}{Set to 90 to rotate x-axis labels 90 degrees.}

\item{financial_year}{Set to TRUE if the assessment was based on financial year (logical). Adjusts the x-axis to show full financial year notation.}

\item{text_size}{Text size (num). Default is 12.}

\item{point_size}{Size of points used in ggplot2::geom_point(). Default is 1.5.}

\item{show_dates_on_axis}{Set to TRUE to show full dates on x-axis as opposed to years.}

\item{colours}{A vector of colours used (character).}
}
\value{
Discard plot
}
\description{
Discard plot
}
\examples{
data <- discardplot_prep_SS(ss_mle)
discardplot(data)
discardplot(data,
            fleets=c(1,2),
            fleet_names = c("Commercial","Recreational"))
}
