% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicB0plot.R
\name{dynamicB0plot}
\alias{dynamicB0plot}
\title{Dynamic B0 plot}
\usage{
dynamicB0plot(
  data,
  area = NULL,
  xlab = "Year",
  ylab = "Spawning biomass",
  xbreaks = NULL,
  ybreaks = NULL,
  xlabels = NULL,
  ylabels = NULL,
  xlim = NULL,
  ylim = NULL,
  xangle = NULL,
  financial_year = FALSE,
  text_size = 12,
  colours = c("black", "grey40"),
  linetype = c("solid", "dashed"),
  scales = "free",
  ncol = 2,
  scenarios = NULL,
  scenario_labels = NULL,
  scenario_order = NULL,
  legend_position = "top",
  line_width = 0.7
)
}
\arguments{
\item{data}{Output from dynamicB0plot_prep(). A data frame with variables called year (int), era (chr), type (chr), value (num), scenario (int)}

\item{area}{Area to show (num). If left blank, plot shows entire fishery.}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{ylab}{Label for y-axis (character). Default is "Spawning biomass".}

\item{xbreaks}{A vector of breaks between x-axis labels, used in ggplot2::scale_x_continous() (numeric).}

\item{ybreaks}{A vector of breaks between y-axis labels, used in ggplot2::scale_y_continous() (numeric).}

\item{xlabels}{A vector of labels for the x-axis breaks.}

\item{ylabels}{A vector of labels for the y-axis breaks.}

\item{xlim}{A vector of lower and upper x-axis limits (e.g. c(1950, 2020)) (numeric).}

\item{ylim}{A vector of lower and upper y-axis limits (e.g. c(0,1)) (numeric).}

\item{xangle}{Set to 90 to rotate x-axis labels 90 degrees.}

\item{financial_year}{Set to TRUE if the assessment was based on financial year (logical). Adjusts the x-axis to show full financial year notation.}

\item{text_size}{Text size (num). Default is 12.}

\item{colours}{A vector of colours used (character).}

\item{linetype}{A vector of linetypes used (character)}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'free', see ?ggplot2::facet_wrap for options.}

\item{ncol}{Number of columns for facet_wrap(). Default is 2.}

\item{scenarios}{A vector of scenario numbers to be shown on plot (numeric). This was already specified in prep file, but this is a manual override to save running the prep function again.}

\item{scenario_labels}{A vector of customised scenario names (character). Default is "Scenario 1", "Scenario 2", etc.}

\item{scenario_order}{A vector to reorder how scenarios are displayed (character). Use the label names defined in "scenario_labels".
If "scenario_labels" is left blank, the labels will be "Scenario 1", "Scenario 2" etc.
Any scenarios not included in "scenario_order" will be tacked on in the order they appear in the input data.}

\item{legend_position}{Position of the legend ("none", "left", "right", "bottom", "top", or two-element numeric vector for x and y position). Default is "top".}

\item{line_width}{Width of lines. Default is 0.7.}
}
\value{
A dynamic B0 plot to show the effect of fishing
}
\description{
Dynamic B0 plot
}
\examples{
data <- dynamicB0plot_prep_SS(ss_mle)
dynamicB0plot(data)
}
