% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fplot.R
\name{Fplot}
\alias{Fplot}
\title{Fishing mortality plot}
\usage{
Fplot(
  data,
  xlab = "Year",
  ylab = expression("Fishing mortality (relative to F"[MSY] * ")"),
  xbreaks = NULL,
  ybreaks = NULL,
  xlabels = NULL,
  ylabels = NULL,
  xlim = NULL,
  ylim = NULL,
  xangle = NULL,
  point_size = 1.5,
  text_size = 12,
  scenarios = NULL,
  scenario_labels = NULL,
  scenario_order = NULL,
  scales = "free",
  ncol = 2,
  financial_year = FALSE,
  show_CI = TRUE,
  mcmc_style = "CI",
  show_median = c("median_F", "trajectory"),
  aggregate_scenarios = FALSE,
  alpha = 0.7,
  line_width = 0.7,
  hair_width = 0.5,
  legend_box = "horizontal",
  legend_position = "top",
  band_colour = "black",
  line_type = c("solid", "dashed"),
  colours = c("black", "darkred"),
  boxplot_outliers = TRUE
)
}
\arguments{
\item{data}{Output from Fplot_prep().
For MLE, a dataframe with columns year (int), value (num), upper (num), lower (num), scenario (fac), biomass_definition (chr)
For MCMC, a dataframe with columns rownum (num), scenario (int), year (num), value (num), med (chr), interval (num), prob_lower (num), prob_upper (num), biomass_definition (chr)}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{ylab}{Label for y-axis (character). Default is "Fishing mortality".}

\item{xbreaks}{A vector of breaks between x-axis labels, used in ggplot2::scale_x_continous() (numeric).}

\item{ybreaks}{A vector of breaks between y-axis labels, used in ggplot2::scale_y_continous() (numeric).}

\item{xlabels}{A vector of labels for the x-axis breaks.}

\item{ylabels}{A vector of labels for the y-axis breaks.}

\item{xlim}{A vector of lower and upper x-axis limits (e.g. c(1950, 2020)) (numeric).}

\item{ylim}{A vector of lower and upper y-axis limits (e.g. c(0,1)) (numeric).}

\item{xangle}{Set to 90 to rotate x-axis labels 90 degrees.}

\item{point_size}{Size of points used in ggplot2::geom_line(). Default is 1.5.}

\item{text_size}{Text size (num). Default is 12.}

\item{scenarios}{A vector of scenario numbers to be shown on plot (numeric). This was already specified in prep file, but this is a manual override to save running the prep function again.}

\item{scenario_labels}{A vector of customised scenario names (character). Default is "Scenario 1", "Scenario 2", etc.}

\item{scenario_order}{A vector to reorder how scenarios are displayed (character). Use the label names defined in "scenario_labels".
If "scenario_labels" is left blank, the labels will be "Scenario 1", "Scenario 2" etc.
Any scenarios not included in "scenario_order" will be tacked on in the order they appear in the input data.}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'free', see ?ggplot2::facet_wrap for options.}

\item{ncol}{Number of columns for facet_wrap(). Default is 2.}

\item{financial_year}{Set to TRUE if the assessment was based on financial year (logical). Adjusts the x-axis to show full financial year notation.}

\item{show_CI}{Set to TRUE to show CI range for MLE plots only (logical).}

\item{mcmc_style}{The type of MCMC plot to be displayed (character). Options are "banded", "hairy", "boxplot", "CI" and "joy", the default is "banded". Only one option can be selected.}

\item{show_median}{Type of median shown. Default "annual_biomass" shows the median of each year,
"trajectory" shows median trajectory based on biomass in final year,
"parameters" uses median of each parameter (not yet implemented)}

\item{aggregate_scenarios}{Set to TRUE to calculate credible intervals across all scenarios (logical). Only activated if mcmc_style==CI.}

\item{alpha}{Transparency for range (numeric) used in ggplot2::geom_density_ridges(). Default is 0.7.}

\item{line_width}{Width of median lines (numeric). Default is 1.}

\item{hair_width}{Width of fine MCMC hairs (numeric). Default is 0.5.}

\item{legend_box}{Display option for legend (character). Choose "vertical" to stack legend types vertically, or "horizontal" to keep legends in one row.}

\item{legend_position}{Position of the legend ("none", "left", "right", "bottom", "top", or two-element numeric vector for x and y position). Default is "top".}

\item{band_colour}{Colour of bands (character). Only used when mcmc_style=="banded". Input one colour, bands will be distinguished using an alpha.}

\item{line_type}{A vector of linetypes (e.g. "solid", "dashed") for median lines.}

\item{colours}{A vector of colours used (character).}

\item{boxplot_outliers}{Set to FALSE to remove outlier points from boxplot. Default is TRUE.}
}
\value{
Fishing mortality plot
}
\description{
Fishing mortality plot
}
\examples{
data <- Fplot_prep_DD(dd_mle)
Fplot(data)

data <- Fplot_prep_SS(ss_mle,ss_mcmc)
Fplot(data)
Fplot(data, mcmc_style="banded")
}
