% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthplot.R
\name{growthplot}
\alias{growthplot}
\title{Growth plot}
\usage{
growthplot(
  data,
  xlab = "Age",
  ylab = "Carapace length (cm, beginning of year)",
  show_two_sex = NULL,
  scenarios = NULL,
  scenario_labels = NULL,
  scenario_order = NULL,
  colours = NULL,
  scales = "free",
  ncol = 2,
  variation_on_variation = FALSE,
  xangle = NULL,
  legend_position = NULL,
  financial_year = FALSE,
  text_size = NULL,
  legend_text_size = NULL,
  text_colour = NULL,
  legend_text_colour = NULL,
  legend_box = NULL,
  legend_title_blank = NULL,
  panel_border = NULL,
  panel_border_colour = NULL
)
}
\arguments{
\item{data}{A data frame with variables age (int), value (num), lower (num), upper (num), sex (int), scenario (int)}

\item{xlab}{Label for x-axis (character). Default is "Age".}

\item{ylab}{Label for y-axis (character). Default is "Carapace length (cm, beginning of year)".}

\item{show_two_sex}{Set to TRUE to activate a feature that is relevant for two-sex models (logical).}

\item{scenarios}{A vector of scenario numbers to be shown on plot (numeric). This was already specified in prep file, but this is a manual override to save running the prep function again.}

\item{scenario_labels}{A vector of customised scenario names (character). Default is "Scenario 1", "Scenario 2", etc.}

\item{scenario_order}{A vector to reorder how scenarios are displayed (character). Use the label names defined in "scenario_labels".
If "scenario_labels" is left blank, the labels will be "Scenario 1", "Scenario 2" etc.
Any scenarios not included in "scenario_order" will be tacked on in the order they appear in the input data.}

\item{colours}{A vector of colours used (character).}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'free', see ?ggplot2::facet_wrap for options.}

\item{ncol}{Number of columns for facet_wrap(). Default is 2.}

\item{variation_on_variation}{Set to TRUE to illustrate MCMC variation on the CV or SD values. Default (FALSE) sets the variation in growth to the median MCMC value.}

\item{text_size}{Text size (num). Default is 12.}

\item{text_size, legend_text_size, text_colour, legend_text_colour, legend_position, legend_box, legend_title_blank, panel_border, panel_border_colour, xangle}{Optional plotting theme overrides. Defaults are controlled by \code{theme_ssand()}
and can be set globally via \code{set_ssand_style()}.}
}
\value{
Growth plot
}
\description{
Growth plot
}
\examples{
data <- growthplot_prep_SS(ss_mle)
growthplot(data)

data <- growthplot_prep_DD(dd_mle)
growthplot(data)
}
