% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Generate a customised map}
\usage{
map(
  show_coast = TRUE,
  show_gbrmp_zoning = FALSE,
  show_cities = FALSE,
  show_historic_zoning = FALSE,
  show_grids = FALSE,
  show_NSW_grids = FALSE,
  show_NSW_zones = FALSE,
  show_specific_grids = FALSE,
  show_heat_map = FALSE,
  show_monitoring_regions = FALSE,
  show_gbrmp_boundary = FALSE,
  show_reefs = FALSE,
  show_custom = FALSE,
  coast_directory = NULL,
  reefs_directory = NULL,
  gbrmp_boundary_directory = NULL,
  gbrmp_zoning_directory = NULL,
  monitoring_region_directory = NULL,
  monitoring_region_object = NULL,
  historic_directory = NULL,
  gbrmp_zoning_exclude = NULL,
  gbrmp_zoning_include = NULL,
  cities = NULL,
  specific_grids = NULL,
  heat_map = NULL,
  monitoring_regions = NULL,
  region = "QLD",
  xlim = NULL,
  ylim = NULL,
  legend_title = NULL,
  legend_position = c(0.75, 0.75),
  caption = NULL,
  colours = NULL,
  labels = NULL,
  text_size = 2.5,
  alpha = 0.5,
  grid_fill = "white",
  region_border = NA,
  coast_colour = "grey90",
  show_labels = TRUE,
  show_latbands = FALSE,
  latbands = NULL,
  latbands_level = "bottom",
  show_scale = FALSE,
  scale_location = "tl",
  show_north = FALSE,
  north_location = "tl",
  reef_fill = "white",
  reef_alpha = 0.2,
  reef_border = "grey80",
  show_reefs_in_legend = FALSE,
  plot_title = NULL,
  custom_directory = NULL,
  custom_object = NULL,
  custom_fill_variable = NULL,
  custom_border_colour = rep("grey30", 99),
  custom_fill_colour_static = NA,
  custom_border_thickness = rep(0.5, 99)
)
}
\arguments{
\item{show_coast}{Set to TRUE to show coastline (logical, default is TRUE)}

\item{show_gbrmp_zoning}{Set to TRUE to show Great Barrier Reef Marine Park zoning (logical).}

\item{show_cities}{Set to TRUE to show cities (logical).}

\item{show_historic_zoning}{Set to TRUE to show historic zoning maps for trawl fishery (logical).}

\item{show_grids}{Set to TRUE to show all grids related to your region (logical).}

\item{show_NSW_grids}{Set to TRUE to show NSW grids (logical).}

\item{show_NSW_zones}{Set to TRUE to show NSW zones (logical). These are just latitude bands but labels are added.}

\item{show_specific_grids}{Set to TRUE to show specific grids on map (logical).}

\item{show_heat_map}{Set to TRUE to show heat map (logical).}

\item{show_monitoring_regions}{Set to TRUE to show Fishery Monitoring regions (logical).}

\item{show_gbrmp_boundary}{Set to TRUE to show Great Barrier Reef Marine Park boundary (logical).}

\item{show_reefs}{Set to TRUE to show reefs (logical).}

\item{show_custom}{Set to TRUE to set up map with custom shapefile.}

\item{coast_directory}{Directory to coast shapefile (character).}

\item{reefs_directory}{Directory to reef shapefile (character).}

\item{gbrmp_boundary_directory}{Directory to GBRMP boundary shapefile (character).}

\item{gbrmp_zoning_directory}{Directory to GBRMP zoning shapefile (character).}

\item{monitoring_region_directory}{Directory to monitoring region shapefile (character).}

\item{monitoring_region_object}{Name of monitoring region sf object (variable name) if already in the R environment, as an alternative to providing the directory of the shapefile.}

\item{historic_directory}{Directory to history shapefiles (vector, character).}

\item{gbrmp_zoning_exclude}{Vector of Zoning names to exclude on map (e.g. "Light Blue Zone"). Either include or exclude zones, not both.}

\item{gbrmp_zoning_include}{Vector of Zoning names to include on map (e.g. "Light Blue Zone"). Either include or exclude zones, not both.}

\item{cities}{A vector of cities to show on the map, from a pre-made list.
If none are specified, some are selected for you, based on region.
Options are: "Sunshine Coast", "Yeppoon", "Karumba", "Pormpuraaw",
"Aurukun","Mapoon","Thursday Island","Cooktown",
"Lockhart River","Port Douglas", "Airlie Beach","Stanage",
"Agnes Water","Hervey Bay","K\'gari","Gold Coast","Cairns",
"Townsville","Mackay","Rockhampton","Brisbane","Bundaberg"}

\item{specific_grids}{A vector of specific grids to be shown. For example, grids that were used in catch rate analysis.}

\item{heat_map}{A dataframe with variables grid (chr) and value (num). Heat map colours grid according to value.}

\item{monitoring_regions}{A vector of monitoring regions to show on map.}

\item{region}{Define the broad region as a shortcut for specifying xlim and ylim. Options are "QLD" (Queensland), "EC" (east coast), "GOC" (Gulf of Carpentaria), "TS" (Torres Strait).
Default is "QLD". For other regions, specify using xlim and ylim.}

\item{xlim}{Longitude limits of graph}

\item{ylim}{Latitude limits of graph (note they should be negative)}

\item{legend_title}{Title of legend (chr)}

\item{legend_position}{Position of the legend ("none", "left", "right", "bottom", "top", or two-element numeric vector for x and y position). Default is "top".}

\item{caption}{Caption of plot (chr)}

\item{colours}{A vector of colours (chr) where the first element is the colour of the landmass.}

\item{labels}{A vector of labels for areas distinguished by 'fill'. Not required is show_monitoring_regions==TRUE.}

\item{text_size}{Text size for grids or regions (num). Default is 2.5.}

\item{alpha}{Alpha on grids (only applies if not using a heat map). Default is 0.5.}

\item{grid_fill}{Colour to fill grids. Default is "white".}

\item{region_border}{Colour for border on regions (default is NA to exclude borders).}

\item{coast_colour}{Colour of coast fill (character.)}

\item{show_labels}{Set to FALSE to turn off fill variable labels on the map. Legend remains.}

\item{show_latbands}{Set to TRUE to show latitude bands on y-axis.}

\item{latbands}{If show_latbands==TRUE, enter latbands to plot on axis. By default, ylim is used.}

\item{latbands_level}{Set to "top" to place latbands as the top layer or "bottom" to place them below the coast layer}

\item{show_scale}{Display scale on map}

\item{scale_location}{Location of scale. Default is top left ("tl")}

\item{show_north}{Display north pointing arrow on map}

\item{north_location}{Location of north pointing arrow. Default is top left ("tl")}

\item{reef_fill}{Colour to fill reefs}

\item{reef_alpha}{Alpha of reefs}

\item{reef_border}{Colour of border of reefs}

\item{show_reefs_in_legend}{Set to TRUE to show the fill colour of reefs in the legend.}

\item{plot_title}{Enter a string to add a plot title.}

\item{custom_directory}{Directory to custom shapefile (character). Can be a list of directories if plotting multiple custom layers.}

\item{custom_object}{Name of custom sf object (variable name) if already in the R environment. Can be a list of objects if plotting multiple custom layers.}

\item{custom_fill_variable}{Variable name in custom shapefile that will be used to guide fill aesthetic. If plotting multiple custom layers, must be a vector as long as the list of layers. Can enter NA for an entry.}

\item{custom_border_colour}{Colour of custom region border, if fill is not dynamic. If plotting multiple custom layers, must be a vector as long as the list of layers.}

\item{custom_fill_colour_static}{Colour to fill, if fill is not dynamic. If plotting multiple custom layers, must be a vector as long as the list of layers.}

\item{custom_border_thickness}{Thickness of border. If plotting multiple custom layers, must be a vector as long as the list of layers.}
}
\value{
A customised map.
}
\description{
Mapping components have been modularised to allow you to add components
}
\examples{
\dontrun{
# To generate a basic map with grids overlaid:
map(region = "EC", coast_directory = "shapefiles/coast.shp", show_grids=TRUE)

# To change region:
map(region = "GOC", coast_directory = "shapefiles/coast.shp", show_grids=TRUE)

# To add cities (could also specify cities using `cities` variable)
map(region = "EC", coast_directory = "shapefiles/coast.shp",
    show_grids=TRUE, show_cities = TRUE)

# To generate a map with specific grids, such as those used in catch rate analysis:
map(region = "EC", coast_directory = "shapefiles/coast.shp",
    show_specific_grids = TRUE,
    specific_grids = c("I15", "K17", "H10", "L16", "Q18", "R12", "P18"))

# To generate a heat map:
heat_map <- data.frame(grid = c("E4", "Z8", "A33", "G29", "C19",
                                "A32", "Y37", "T12", "S9", "F40",
                                "J8", "J12", "Z40", "M5", "N23"),
                       value= c(0.74, 0.14, 0.301, 1.172, 0.863,
                                0.444, 1.044, 0.558, 0.31, 1.861,
                                0.63, 0.682, 0.168, 0.82, 0.898))
map(region = "EC", coast_directory = "shapefiles/coast.shp",
    show_heat_map = TRUE, heat_map = heat_map,
    legend_position = "right")

# To show monitoring regions:
map(region = "EC", coast_directory = "shapefiles/coast.shp",
    show_monitoring_regions = TRUE,
    monitoring_region_directory = "shapefiles/monitoring.shp",
    monitoring_regions = c("Lucinda Region", "Swains Region", "Bowen Region"),
    labels = c("Lucinda Region", "Swains Region", "Bowen Region"))

# To show reefs and GBRMP boundary
map(region = "EC", show_gbrmp_boundary = TRUE, show_reefs = TRUE,
    gbrmp_boundary_directory = "shapefiles/gbrmp_boundary.shp",
    reefs_directory = "shapefiles/reefs.shp",
    colours = c("grey90", "white", "darkgreen"),
    labels = c("Great Barrier Reef Marine Park", "Reef"))

# To show GBRMP zoning
map(region = "EC", coast_directory = "shapefiles/coast.shp",
    show_gbrmp_zoning = TRUE,
    gbrmp_zoning_directory = "shapefiles/gbrmp_zoning.shp",
    labels = c("Green Zone",
               "Yellow Zone",
               "Dark Blue Zone",
               "Olive Green Zone",
               "Orange Zone",
               "Pink Zone",
               "Light Blue Zone",
               "Commonwealth Island",
               "Commonwealth Island Zone"),
    colours = c("grey70","#41ba07", "#FFC000",
                        "darkblue","#26590d", "#ff770f",
                        "#ff0fd7", "#0fe3ff", "#340942", "#453d4d"))

# To plot using a custom shapefile:
map(region="EC", coast_directory = "shapefiles/coast.shp",
    show_custom = TRUE,
    custom_directory = "custom.shp",
    custom_fill_variable = "BDM_ZONE",
    xlim = c(149,153),
    ylim = c(-24,-20),
    colours = rep(c("#F7C915","#3F73D3", "#F47E7A"),100),
    region_border = "black",
    legend_position = "none",
    coast_colour = "#dadf77",
    alpha = 1)

# To plot a map with latbands:
map(coast_directory = "shapefiles/coast.shp",
  show_latbands = TRUE, latbands_level = "bottom")

# To plot a map with QLD and NSW grids:
map(coast_directory = "shapefiles/coast.shp",
      ylim = c(-30, -20),
      show_grids = TRUE,
      show_NSW_grids = TRUE)

}
}
