% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_finalbiomassposteriorplot.R
\name{mcmc_finalbiomassposteriorplot}
\alias{mcmc_finalbiomassposteriorplot}
\title{Posterior plot for final biomass ratio with risk area (below B20) highlighted}
\usage{
mcmc_finalbiomassposteriorplot(
  data,
  colours = c("#AD3D25", "#FCC17E", "#0085B8", "#00925B"),
  xlab = NULL,
  ylab = NULL,
  CI_label_position = "bottom",
  line_type = NULL,
  line_type_label = NULL,
  show_median = TRUE,
  annotation_text_size = 5,
  legend_text_size = 11.5,
  axis_title_size = 13,
  axis_text_size = 12,
  xangle = NULL,
  legend_position = NULL,
  financial_year = FALSE,
  text_size = NULL,
  text_colour = NULL,
  legend_text_colour = NULL,
  legend_box = NULL,
  legend_title_blank = NULL,
  panel_border = NULL,
  panel_border_colour = NULL
)
}
\arguments{
\item{data}{List output from mcmc_finalbiomassposterior_prep()}

\item{colours}{A vector of colours used (character).}

\item{xlab}{Label for x-axis (character). Default is "".}

\item{ylab}{Label for y-axis (character). Default is "".}

\item{CI_label_position}{Specify position for labels for 95\% confidence interval percentages. Default "bottom", alternatives are "top" or "none".}

\item{line_type}{Linetype for vertical lines. Default is "dashed".}

\item{line_type_label}{Label for linetype legend. Default is "X\% falls within this range" where X is the credible_interval specified in the input data.}

\item{show_median}{Default is TRUE. Set to FALSE to remove median line.}

\item{annotation_text_size}{Size of annotation text (e.g. median and CI). Default is 5.}

\item{legend_text_size}{Size of legend text. Default is 11.5.}

\item{axis_title_size}{Sie of axis title. Default is 13.}

\item{axis_text_size}{Size of axis text. Default is 12.}

\item{legend_position}{Position of the legend ("none", "left", "right", "bottom", "top", or two-element numeric vector for x and y position). Default is "top".}

\item{text_size, legend_text_size, text_colour, legend_text_colour, legend_position, legend_box, legend_title_blank, panel_border, panel_border_colour, xangle}{Optional plotting theme overrides. Defaults are controlled by \code{theme_ssand()}
and can be set globally via \code{set_ssand_style()}.}
}
\value{
A posterior plot for final biomass ratio with risk area (below B20) highlighted
}
\description{
Posterior plot for final biomass ratio with risk area (below B20) highlighted
}
\examples{
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc)
data <- mcmc_finalbiomassposterior_prep_SS(ss_mle,ss_mcmc_ens)
mcmc_finalbiomassposteriorplot(data)
}
