% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_finalbiomassposteriorplot.R
\name{mcmc_finalbiomassposteriorplot}
\alias{mcmc_finalbiomassposteriorplot}
\title{Posterior plot for final biomass ratio with risk area (below B20) highlighted}
\usage{
mcmc_finalbiomassposteriorplot(
  data,
  colours = c("#AD3D25", "#FCC17E", "#0085B8", "#00925B"),
  xlab = NULL,
  ylab = NULL,
  legend_position = "top",
  CI_label_position = "bottom",
  line_type = "dashed",
  line_type_label = NULL
)
}
\arguments{
\item{data}{List output from mcmc_finalbiomassposterior_prep()}

\item{colours}{A vector of colours used (character).}

\item{xlab}{Label for x-axis (character). Default is "".}

\item{ylab}{Label for y-axis (character). Default is "".}

\item{legend_position}{Position of the legend ("none", "left", "right", "bottom", "top", or two-element numeric vector for x and y position). Default is "top".}

\item{CI_label_position}{Specify position for labels for 95\% confidence interval percentages. Default "bottom", alternative is "top".}

\item{line_type}{Linetype for vertical lines. Default is "dashed".}

\item{line_type_label}{Label for linetype legend. Default is "X\% falls within this range" where X is the credible_interval specified in the input data.}
}
\value{
A posterior plot for final biomass ratio with risk area (below B20) highlighted
}
\description{
Posterior plot for final biomass ratio with risk area (below B20) highlighted
}
\examples{
ss_mcmc_ens <- mcmc_ensemble_SS(ss_mcmc)
data <- mcmc_finalbiomassposterior_prep_SS(ss_mle,ss_mcmc_ens)
mcmc_finalbiomassposteriorplot(data)
}
