% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_posteriordensityplot.R
\name{mcmc_posteriordensityplot}
\alias{mcmc_posteriordensityplot}
\title{Posterior density plot for MCMC run}
\usage{
mcmc_posteriordensityplot(
  data,
  add_lines = c("optimised", "median parameter", "median trajectory"),
  xlab = "",
  ylab = "Density",
  ncol = 3,
  parameter_labels = NULL,
  xmax = NULL
)
}
\arguments{
\item{data}{Output from mcmc_posterior_densityplot_prep(). List of data frames with columns parameter (character), value (numeric), year (numeric), chain (numeric).}

\item{add_lines}{Lines to add to density plots. e.g. c('optimised', 'median parameter', 'median trajectory')}

\item{xlab}{Label for x-axis (character). Default is "".}

\item{ylab}{Label for y-axis (character). Default is "Density".}

\item{ncol}{Number of columns for facet_wrap()}

\item{parameter_labels}{Vector of customised parameter labels for facet wrap, e.g. expression(xi)}

\item{xmax}{Named vector of maximum x-axis limit, e.g. c('R\[0\]' = 50).}
}
\value{
Posterior density plot
}
\description{
Posterior density plot for MCMC run
}
\examples{
parameters <- extract_SS_parameters(ss_mcmc)[c(2:10,449),]
data <- mcmc_posteriordensityplot_prep_SS(ss_mle, ss_mcmc, scenario = 1, parameters)
mcmc_posteriordensityplot(data)
}
