% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_traceplot.R
\name{mcmc_traceplot}
\alias{mcmc_traceplot}
\title{MCMC traceplot}
\usage{
mcmc_traceplot(
  data,
  colours = c("#b4d9eb", "#cc8c97", "#ddd3a0", "#7ed19a", "#988dd3"),
  show_opt = FALSE,
  ncol = 2,
  sample = NULL,
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{data}{Output from mcmc_posteriordensityplot_prep(). A list of dataframes}

\item{colours}{A vector of colours used (character).}

\item{show_opt}{Show maximum likelihood estimate from optim()}

\item{ncol}{Number of columns for facet_wrap(). Default is .}

\item{sample}{Number of samples to plot from each MCMC chain to ease burden of rendering dense plots (numeric).}

\item{xlab}{Label for x-axis (character). Default is "".}

\item{ylab}{Label for y-axis (character). Default is "".}
}
\value{
Trace plot
}
\description{
MCMC traceplot
}
\examples{
parameters <- extract_SS_parameters(ss_mcmc)[c(2:9),]
data <- mcmc_posteriordensityplot_prep_SS(ss_mle,
                                          ss_mcmc,
                                          scenario = 1,
                                          parameters,
                                          show_objective_function=TRUE)
mcmc_traceplot(data)

\dontrun{
library(DDUST)
dd_sim <- simulate_DDUST(dd_mle,dd_mcmc)
data <- mcmc_posteriordensityplot_prep_DD(dd_mle, dd_mcmc,
                                          dd_sim, scenario = 1)
mcmc_traceplot(data)
}
}
