% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseplot.R
\name{phaseplot}
\alias{phaseplot}
\title{Phase plot}
\usage{
phaseplot(
  data,
  xbreaks = seq(0, 2, 0.2),
  scales = "free_y",
  year_label_size = 3,
  ncol = 2,
  year_labels = NULL,
  scenarios = NULL,
  scenario_labels = NULL,
  scenario_order = NULL,
  xlab = "Biomass (relative)",
  ylab = expression(Fishing ~ pressure ~ ratio ~ (relative ~ to ~ F[MSY])),
  xangle = NULL,
  legend_position = NULL,
  text_size = NULL,
  legend_text_size = NULL,
  text_colour = NULL,
  legend_text_colour = NULL,
  legend_box = NULL,
  legend_title_blank = NULL,
  panel_border = NULL,
  panel_border_colour = NULL
)
}
\arguments{
\item{data}{Output from phaseplot_prep()}

\item{xbreaks}{A vector of breaks between x-axis labels, used in ggplot2::scale_x_continous() (numeric).}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'free', see ?ggplot2::facet_wrap for options.}

\item{year_label_size}{Text size for year labels (num). Default is 3.}

\item{ncol}{Number of columns for facet_wrap(). Default is 2.}

\item{year_labels}{A vector of years to be displayed on plot (numeric).}

\item{scenarios}{A vector of scenario numbers to be shown on plot (numeric). This was already specified in prep file, but this is a manual override to save running the prep function again.}

\item{scenario_labels}{A vector of customised scenario names (character). Default is "Scenario 1", "Scenario 2", etc.}

\item{scenario_order}{A vector to reorder how scenarios are displayed (character). Use the label names defined in "scenario_labels".
If "scenario_labels" is left blank, the labels will be "Scenario 1", "Scenario 2" etc.
Any scenarios not included in "scenario_order" will be tacked on in the order they appear in the input data.}

\item{xlab}{Label for x-axis (character). Default is "Biomass (relative)".}

\item{ylab}{Label for y-axis (character). Default is "Fishing pressure ratio (relative to FMSY))".}

\item{text_size, legend_text_size, text_colour, legend_text_colour, legend_position, legend_box, legend_title_blank, panel_border, panel_border_colour, xangle}{Optional plotting theme overrides. Defaults are controlled by \code{theme_ssand()}
and can be set globally via \code{set_ssand_style()}.}
}
\value{
A phase plot
}
\description{
Phase plot
}
\examples{
data <- phaseplot_prep_SS(ss_mle)
phaseplot(data)

data <- phaseplot_prep_DD(dd_mle)
phaseplot(data)
}
