% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportionspawningplot.R
\name{proportionspawningplot}
\alias{proportionspawningplot}
\title{Proportion spawning plot}
\usage{
proportionspawningplot(
  data,
  scenarios = NULL,
  scenario_labels = NULL,
  scenario_order = NULL,
  ylim = NULL,
  ylabels = NULL,
  ybreaks = NULL,
  scales = "free",
  ncol = 2,
  xlab = "Month",
  ylab = "Proportion spawning",
  xangle = NULL,
  legend_position = NULL,
  text_size = NULL,
  legend_text_size = NULL,
  text_colour = NULL,
  legend_text_colour = NULL,
  legend_box = NULL,
  legend_title_blank = NULL,
  panel_border = NULL,
  panel_border_colour = NULL
)
}
\arguments{
\item{data}{Output from proportionspawningplot_prep() with columns month (fac), value (num), scenario (fac)}

\item{scenarios}{A vector of scenario numbers to be shown on plot (numeric). This was already specified in prep file, but this is a manual override to save running the prep function again.}

\item{scenario_labels}{A vector of customised scenario names (character). Default is "Scenario 1", "Scenario 2", etc.}

\item{scenario_order}{A vector to reorder how scenarios are displayed (character). Use the label names defined in "scenario_labels".
If "scenario_labels" is left blank, the labels will be "Scenario 1", "Scenario 2" etc.
Any scenarios not included in "scenario_order" will be tacked on in the order they appear in the input data.}

\item{ylim}{A vector of lower and upper y-axis limits (e.g. c(0,1)) (numeric).}

\item{ylabels}{A vector of labels for the y-axis breaks.}

\item{ybreaks}{A vector of breaks between y-axis labels, used in ggplot2::scale_y_continous() (numeric).}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'free', see ?ggplot2::facet_wrap for options.}

\item{ncol}{Number of columns for facet_wrap(). Default is .}

\item{xlab}{Label for x-axis (character). Default is "Month".}

\item{ylab}{Label for y-axis (character). Default is "Proportion spawning".}

\item{text_size, legend_text_size, text_colour, legend_text_colour, legend_position, legend_box, legend_title_blank, panel_border, panel_border_colour, xangle}{Optional plotting theme overrides. Defaults are controlled by \code{theme_ssand()}
and can be set globally via \code{set_ssand_style()}.}
}
\value{
Proportion spawning plot
}
\description{
Proportion spawning plot
}
\examples{
data <- proportionspawningplot_prep_DD(dd_mle)
proportionspawningplot(data)
}
