% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitmentpatternplot.R
\name{recruitmentpatternplot}
\alias{recruitmentpatternplot}
\title{Plot of recruitment pattern}
\usage{
recruitmentpatternplot(
  data,
  xlab = "Month",
  ylab = "Recruitment \%",
  scenarios = NULL,
  scenario_labels = NULL,
  scenario_order = NULL,
  ncol = 2,
  scales = "free",
  colours = c("black", "grey80", "grey20", "grey60", "grey40"),
  xangle = NULL,
  legend_position = NULL,
  financial_year = FALSE,
  text_size = NULL,
  legend_text_size = NULL,
  text_colour = NULL,
  legend_text_colour = NULL,
  legend_box = NULL,
  legend_title_blank = NULL,
  panel_border = NULL,
  panel_border_colour = NULL
)
}
\arguments{
\item{data}{Output from recruitmentpatternplot_prep(). A dataframe with recruitment (num), months (num), monthnames (fac), scenario (fac)}

\item{xlab}{Label for x-axis (character). Default is "Month".}

\item{ylab}{Label for y-axis (character). Default is "Recruitment \%".}

\item{scenarios}{A vector of scenario numbers to be shown on plot (numeric). This was already specified in prep file, but this is a manual override to save running the prep function again.}

\item{scenario_labels}{A vector of customised scenario names (character). Default is "Scenario 1", "Scenario 2", etc.}

\item{scenario_order}{A vector to reorder how scenarios are displayed (character). Use the label names defined in "scenario_labels".
If "scenario_labels" is left blank, the labels will be "Scenario 1", "Scenario 2" etc.
Any scenarios not included in "scenario_order" will be tacked on in the order they appear in the input data.}

\item{ncol}{Number of columns for facet_wrap(). Default is 2.}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'free', see ?ggplot2::facet_wrap for options.}

\item{colours}{A vector of colours used (character).}

\item{text_size, legend_text_size, text_colour, legend_text_colour, legend_position, legend_box, legend_title_blank, panel_border, panel_border_colour, xangle}{Optional plotting theme overrides. Defaults are controlled by \code{theme_ssand()}
and can be set globally via \code{set_ssand_style()}.}
}
\value{
Plot of recruitment pattern
}
\description{
Stock synthesis functionality not yet introduced
}
\examples{
data <- recruitmentpatternplot_prep_DD(dd_mle)
recruitmentpatternplot(data)
}
