% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghettiplot.R
\name{spaghettiplot}
\alias{spaghettiplot}
\title{Spaghetti plot}
\usage{
spaghettiplot(
  data,
  template = "coloured",
  show_base_case = TRUE,
  legend_labels = NULL,
  legend_position = "top",
  xlab = "Year",
  ylab = "Biomass (relative unfished)",
  xlim = c(NA, NA),
  ylim = c(0, NA),
  xbreaks = NULL,
  ybreaks = NULL,
  xlabels = NULL,
  ylabels = NULL,
  xangle = NULL,
  show_scenario_labels = TRUE,
  linetype_categories = NA,
  linetype_labels = NA,
  linetypes = NA,
  colour_categories = NA,
  colour_labels = NA,
  colours = NA,
  scenarios = NULL,
  scenario_labels = NULL,
  scenario_order = NULL,
  line_width = 1,
  financial_year = FALSE,
  force = 1,
  text_size = 12,
  scenario_text_size = 4
)
}
\arguments{
\item{data}{Output from spaghettiplot_prep(). A data frame with variables called year, scenario, base and value}

\item{template}{Specify a pre-built template to use: options are "coloured" or "greyscale"}

\item{show_base_case}{Set to TRUE to highlight the base case (logical)}

\item{legend_labels}{A vector to specify legend labels}

\item{legend_position}{Legend position (default is "top")}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{ylab}{Label for y-axis (character). Default is "Biomass (relative unfished)".}

\item{xlim}{A vector of lower and upper x-axis limits (e.g. c(1950, 2020)) (numeric).}

\item{ylim}{A vector of lower and upper y-axis limits (e.g. c(0,1)) (numeric).}

\item{xbreaks}{A vector of breaks between x-axis labels, used in ggplot2::scale_x_continous() (numeric).}

\item{ybreaks}{A vector of breaks between y-axis labels, used in ggplot2::scale_y_continous() (numeric).}

\item{xlabels}{A vector of labels for the x-axis breaks.}

\item{ylabels}{A vector of labels for the y-axis breaks.}

\item{xangle}{Set to 90 to rotate x-axis labels 90 degrees.}

\item{show_scenario_labels}{Set to TRUE to include scenario labels on the plots (logical).}

\item{linetype_categories}{For a custom design scheme. A vector, as long as your list of scenarios, where each element of the vector specifies which linetype category the scenario belongs to.}

\item{linetype_labels}{A vector of labels for the groups specified in linetype_categories}

\item{linetypes}{A vector of the linetypes desired for the groups specified in linetype_labels}

\item{colour_categories}{For a custom design scheme. A vector, as long as your list of scenarios, where each element of the vector specifies which colour category the scenario belongs to.}

\item{colour_labels}{A vector of labels for the groups specified in colour_categories}

\item{colours}{A vector of the colours desired for the groups specified in colour_labels}

\item{scenarios}{A vector of scenarios to show on plot. Already specified in prep file, but this is a manual override to save running the prep function again}

\item{scenario_labels}{A vector of customised scenario names (character). Default is "Scenario 1", "Scenario 2", etc.}

\item{scenario_order}{A vector to reorder how scenarios are displayed (character). Use the label names defined in "scenario_labels".
If "scenario_labels" is left blank, the labels will be "Scenario 1", "Scenario 2" etc.
Any scenarios not included in "scenario_order" will be tacked on in the order they appear in the input data.}

\item{line_width}{Width of lines}

\item{financial_year}{Set to TRUE if the assessment was based on financial year (logical). Adjusts the x-axis to show full financial year notation.}

\item{force}{Force of repulsion between overlapping text labels. Default is 1.}

\item{text_size}{text_size, default 12}

\item{scenario_text_size}{scenario_text_size, default 4}
}
\value{
A spaghetti plot
}
\description{
Display biomass trajectories for multiple scenarios. See Details for customisation options.
}
\details{
\itemize{
\item Can use pre-built template ("coloured" or "greyscale") or design a custom scheme on the fly.
\item A custom scheme can distinguish groups of scenarios using line-type and colour.
\item For example, you might want to show catch rate settings using colours, and natural mortality settings using linetype.
\item The variable linetype_categories expects a vector, as long as your list of scenarios, where each element of the vector specifies which linetype category the scenario belongs to.
\item For example, if you had six scenarios and wanted the first half belonging to group 1 and the second belonging to group 2, you would enter linetype_categories = c(1,1,1,2,2,2)
\item You can then specify labels for these groups using linetype_labels.
\item For example, linetype_labels = c("M = 1", "M = 2")
\item You can also specify the linetypes used for these categories using linetypes
\item For example, linetypes = c("solid", "dashed")
\item You can set up the colours a similar way, simply replacing "linetype" with "colour" in the examples above.
}
}
\examples{
data <- spaghettiplot_prep_DD(dd_mle)
spaghettiplot(data)
spaghettiplot(data, financial_year=TRUE)

data <- spaghettiplot_prep_SS(ss_mle, ss_mcmc)
spaghettiplot(data)
spaghettiplot(data, template = "greyscale", line_width = 0.5)
spaghettiplot(data,
              linetype_categories = c(1,2),
              linetype_labels = c("Catch rate A", "Catch rate B"),
              linetypes = c("solid", "dashed"),
              colour_categories = c(1,2),
              colour_labels = c("High M", "Medium M"),
              colours = fq_palette("alisecolours")[1:2])
}
