% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawningoutputplot.R
\name{spawningoutputplot}
\alias{spawningoutputplot}
\title{Spawning output plot}
\usage{
spawningoutputplot(
  data,
  xaxis = "age",
  xlab = "Age (years)",
  ylab = "Spawning output",
  colours = "black",
  scenarios = NULL,
  scenario_labels = NULL,
  scenario_order = NULL,
  scales = "free",
  ncol = 2,
  xangle = NULL,
  legend_position = NULL,
  financial_year = FALSE,
  text_size = NULL,
  legend_text_size = NULL,
  text_colour = NULL,
  legend_text_colour = NULL,
  legend_box = NULL,
  legend_title_blank = NULL,
  panel_border = NULL,
  panel_border_colour = NULL
)
}
\arguments{
\item{data}{Output from spawningoutputplot_prep(). A dataframe with xvar (int), maturityxfecundity (num), x (chr), scenario (int)}

\item{xaxis}{The variable you would like to appear on the x-axis. Options are "age" or "length"}

\item{xlab}{Label for x-axis (character). Default is "Year".}

\item{ylab}{Label for y-axis (character). Default is "Retained catch (t)".}

\item{colours}{A vector of colours used (character).}

\item{scenarios}{A vector of scenarios to plot (numeric). Shows all scenarios if left blank. Can be overridden in the plotting function.}

\item{scenario_labels}{A vector of customised scenario names (character). Default is "Scenario 1", "Scenario 2", etc.}

\item{scenario_order}{A vector to reorder how scenarios are displayed (character). Use the label names defined in "scenario_labels".
If "scenario_labels" is left blank, the labels will be "Scenario 1", "Scenario 2" etc.
Any scenarios not included in "scenario_order" will be tacked on in the order they appear in the input data.}

\item{scales}{Scales for ggplot2::facet_wrap(). Default is 'free', see ?ggplot2::facet_wrap for options.}

\item{ncol}{Number of columns for facet_wrap(). Default is 2.}

\item{text_size, legend_text_size, text_colour, legend_text_colour, legend_position, legend_box, legend_title_blank, panel_border, panel_border_colour, xangle}{Optional plotting theme overrides. Defaults are controlled by \code{theme_ssand()}
and can be set globally via \code{set_ssand_style()}.}
}
\value{
Spawning output plot
}
\description{
Spawning output plot
}
\examples{
data <- spawningoutputplot_prep_SS(ss_mle)
spawningoutputplot(data,xaxis="length")
}
