% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSAND.R
\docType{package}
\name{SSAND-package}
\alias{SSAND}
\alias{SSAND-package}
\title{SSAND: Standardised Stock Assessment Numerical Displays}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolkit for visualising data used in, and produced by, stock assessment models. Raw model outputs from specific stock assessment models (so far Stock Synthesis and DDUST) can be transformed into report-ready graphics, tables and values, and plugged directly into a report template. Raw data can also be visualised using individual functions or through comprehensive investigation tools.
}
\author{
\strong{Maintainer}: Alise Fox \email{fisheriesassessment@daf.qld.gov.au} (\href{https://orcid.org/0009-0002-5634-0963}{ORCID})

Authors:
\itemize{
  \item Montana Wickens (\href{https://orcid.org/0000-0002-7316-014X}{ORCID})
  \item Robyn Lovett
  \item Justin Cameron
  \item Mai Tanimoto
}

Other contributors:
\itemize{
  \item Fisheries Queensland, FQ [copyright holder, funder]
}

}
\keyword{internal}
